Module glium::backend

The backend module allows one to link between glium and the OpenGL context..

There are three concepts in play:

The Backend trait describes the glue between glium and the OpenGL context provider like glutin, SDL, GLFW, etc.

The Context struct is the main brick of glium. It manages everything that glium needs to execute OpenGL commands. Creating a Context requires a Backend.

The Facade trait. Calling functions like VertexBuffer::new requires passing an object that implements this trait. It is implemented on Rc<Context>.

Modules

glutin
Backend implementation for the glutin library

Structs

Context
Stores the state and information required for glium to execute commands. Most public glium functions require passing a Rc<Context>.

Enums

ReleaseBehavior
Defines what happens when you change the current context.

Traits

Backend
Trait for types that can be used as a backend for a glium context.

Facade
Trait for types that provide a safe access for glium functions.