Module glium::backend::glutin

Backend implementation for the glutin library

Features

Only available if the 'glutin' feature is enabled.

Reexports

pub extern crate glutin;

Modules

headless
Backend implementation for a glutin headless renderer.

Structs

Display
A GL context combined with a facade for drawing upon.

GlutinBackend
An implementation of the Backend trait for glutin.

Enums

DisplayCreationError
Error that can happen while creating a glium display.