Struct glium::backend::glutin::Display

pub struct Display { /* fields omitted */ }

A GL context combined with a facade for drawing upon.

The Display uses glutin for the Window and its associated GL Context.

These are stored alongside a glium-specific context.

Methods

impl Display

pub fn new(
window_builder: WindowBuilder,
context_builder: ContextBuilder,
events_loop: &EventsLoop
) -> Result<Self, DisplayCreationError>

Create a new glium Display from the given context and window builders.

Performs a compatibility check to make sure that all core elements of glium are supported by the implementation.

pub fn from_gl_window(gl_window: GlWindow) -> Result<Self, IncompatibleOpenGl>

Create a new glium Display.

Performs a compatibility check to make sure that all core elements of glium are supported by the implementation.

pub unsafe fn unchecked(gl_window: GlWindow) -> Result<Self, IncompatibleOpenGl>

Create a new glium Display.

This function does the same as build_glium, except that the resulting context will assume that the current OpenGL context will never change.

pub fn with_debug(
gl_window: GlWindow,
debug: DebugCallbackBehavior
) -> Result<Self, IncompatibleOpenGl>

The same as the new constructor, but allows for specifying debug callback behaviour.

pub unsafe fn unchecked_with_debug(
gl_window: GlWindow,
debug: DebugCallbackBehavior
) -> Result<Self, IncompatibleOpenGl>

The same as the unchecked constructor, but allows for specifying debug callback behaviour.

pub fn rebuild(
&self,
window_builder: WindowBuilder,
context_builder: ContextBuilder,
events_loop: &EventsLoop
) -> Result<(), DisplayCreationError>

Rebuilds the Display's GlWindow with the given window and context builders.

This method ensures that the new GlWindow's Context will share the display lists of the original GlWindow's Context.

pub fn gl_window(&self) -> Ref<GlWindow>

Borrow the inner glutin GlWindow.

pub fn draw(&self) -> Frame

Start drawing on the backbuffer.

This function returns a Frame, which can be used to draw on it. When the Frame is destroyed, the buffers are swapped.

Note that destroying a Frame is immediate, even if vsync is enabled.

If the framebuffer dimensions have changed since the last call to draw, the inner glutin context will be resized accordingly before returning the Frame.