Module glium::backend::glutin_backend

Backend implementation for the glutin library

Features

Only available if the 'glutin' feature is enabled.

Structs

GlutinFacade
Facade implementation for glutin. Wraps both glium and glutin.

GlutinHeadlessBackend
An implementation of the Backend trait for a glutin headless context.

GlutinWindowBackend
An implementation of the Backend trait for a glutin window.

PollEventsIter
Iterator for all the events received by the window.

WaitEventsIter
Blocking iterator over all the events received by the window.

WinRef
Borrow of the glutin window.