Struct glium::backend::glutin_backend::GlutinFacade

pub struct GlutinFacade {
    // some fields omitted
}

Facade implementation for glutin. Wraps both glium and glutin.

Methods

impl GlutinFacade

fn poll_events(&self) -> PollEventsIter

Reads all events received by the window.

This iterator polls for events and can be exhausted.

fn wait_events(&self) -> WaitEventsIter

Reads all events received by the window.

fn is_closed(&self) -> bool

Returns true if the window has been closed.

fn get_window(&self) -> Option<WinRef>

Returns the underlying window, or None if glium uses a headless context.

fn get_framebuffer_dimensions(&self) -> (u32, u32)

Returns the dimensions of the main framebuffer.

fn get_opengl_version(&self) -> Version

Returns the OpenGL version of the current context.

fn get_supported_glsl_version(&self) -> Version

Returns the supported GLSL version.

fn is_glsl_version_supported(&self, version: &Version) -> bool

Returns true if the given GLSL version is supported.

fn draw(&self) -> Frame

Start drawing on the backbuffer.

This function returns a Frame, which can be used to draw on it. When the Frame is destroyed, the buffers are swapped.

Note that destroying a Frame is immediate, even if vsync is enabled.

fn get_max_anisotropy_support(&self) -> Option<u16>

Returns the maximum value that can be used for anisotropic filtering, or None if the hardware doesn't support it.

fn get_max_viewport_dimensions(&self) -> (u32, u32)

Returns the maximum dimensions of the viewport.

Glium will panic if you request a larger viewport than this when drawing.

fn release_shader_compiler(&self)

Releases the shader compiler, indicating that no new programs will be created for a while.

Features

This method is always available, but is a no-op if it's not available in the implementation.

fn get_free_video_memory(&self) -> Option<usize>

Returns an estimate of the amount of video memory available in bytes.

Returns None if no estimate is available.

fn read_front_buffer<P, T>(&self) -> T where P: PixelValue + Clone + Send, T: Texture2dDataSink<Data=P>

Reads the content of the front buffer.

You will only see the data that has finished being drawn.

This function can return any type that implements Texture2dData.

Example

let pixels: Vec<Vec<(u8, u8, u8)>> = display.read_front_buffer();

unsafe fn exec_in_context<'a, T, F>(&self, action: F) -> T where T: Send + 'static, F: FnOnce() -> T + 'a

Execute an arbitrary closure with the OpenGL context active. Useful if another component needs to directly manipulate OpenGL state.

If action manipulates any OpenGL state, it must be restored before action completes.

fn assert_no_error(&self)

Asserts that there are no OpenGL errors pending.

This function should be used in tests.

fn synchronize(&self)

Waits until all the previous commands have finished being executed.

When you execute OpenGL functions, they are not executed immediately. Instead they are put in a queue. This function waits until all commands have finished being executed, and the queue is empty.

You don't need to call this function manually, except when running benchmarks.