Struct glium::buffer::Buffer

pub struct Buffer<T: ?Sized> where
    T: Content,  { /* fields omitted */ }

Represents a view of a buffer.

Methods

impl<T: ?Sized> Buffer<T> where
T: Content,

pub fn new<F: ?Sized>(
facade: &F,
data: &T,
ty: BufferType,
mode: BufferMode
) -> Result<Buffer<T>, BufferCreationError> where
F: Facade,

Builds a new buffer containing the given data. The size of the buffer is equal to the size of the data.

pub fn empty_unsized<F: ?Sized>(
facade: &F,
ty: BufferType,
size: usize,
mode: BufferMode
) -> Result<Buffer<T>, BufferCreationError> where
F: Facade,

Builds a new buffer of the given size.

pub fn get_context(&self) -> &Rc<Context>

Returns the context corresponding to this buffer.

pub fn get_size(&self) -> usize

Returns the size in bytes of this buffer.

pub fn is_persistent(&self) -> bool

Returns true if this buffer uses persistent mapping.

pub fn write(&self, data: &T)

Uploads some data in this buffer.

Implementation

For persistent-mapped buffers, waits untils the data is no longer used by the GPU then memcpies the data to the mapping.

For immutable buffers, creates a temporary buffer that contains the data then calls glCopyBufferSubData to copy from the temporary buffer to the real one.

For other types, calls glBufferSubData.

Panic

Panics if the length of data is different from the length of this buffer.

pub fn invalidate(&self)

Invalidates the content of the buffer. The data becomes undefined.

You should call this if you only use parts of a buffer. For example if you want to use the first half of the buffer, you invalidate the whole buffer then write the first half.

This operation is a no-op if the backend doesn't support it and for persistent-mapped buffers.

Implementation

Calls glInvalidateBufferData if supported. Otherwise, calls glBufferData with a null pointer for data. If glBufferStorage has been used to create the buffer and glInvalidateBufferData is not supported, does nothing.

pub fn read(&self) -> Result<T::Owned, ReadError>

Reads the content of the buffer.

pub fn map(&mut self) -> Mapping<T>

Maps the buffer in memory for both reading and writing.

Implementation

For persistent-mapped buffers, waits until the data is no longer accessed by the GPU then returns a pointer to the existing mapping.

For immutable buffers, creates a temporary buffer containing the data of the buffer and maps it. When the mapping object is destroyed, copies the content of the temporary buffer to the real buffer.

For other types, calls glMapBuffer or glMapSubBuffer.

pub fn map_read(&mut self) -> ReadMapping<T>

Maps the buffer in memory for reading.

Implementation

For persistent-mapped buffers, waits until the data is no longer accessed by the GPU then returns a pointer to the existing mapping.

For immutable buffers, creates a temporary buffer containing the data of the buffer and maps it.

For other types, calls glMapBuffer or glMapSubBuffer.

pub fn map_write(&mut self) -> WriteMapping<T>

Maps the buffer in memory for writing only.

Implementation

For persistent-mapped buffers, waits until the data is no longer accessed by the GPU then returns a pointer to the existing mapping.

For immutable buffers, creates a temporary buffer and maps it. When the mapping object is destroyed, copies the content of the temporary buffer to the real buffer.

For other types, calls glMapBuffer or glMapSubBuffer.

pub fn copy_to<'a, S>(&self, target: S) -> Result<(), CopyError> where
S: Into<BufferSlice<'a, T>>,
T: 'a,

Copies the content of the buffer to another buffer.

Panic

Panics if T is unsized and the other buffer is too small.

pub unsafe fn slice_custom<F, R: ?Sized>(&self, f: F) -> BufferSlice<R> where
F: for<'r> FnOnce(&'r T) -> &'r R,
R: Content,

Builds a slice that contains an element from inside the buffer.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

Example

#[derive(Copy, Clone)]
struct BufferContent {
    value1: u16,
    value2: u16,
}
let slice = unsafe { buffer.slice_custom(|content| &content.value2) };

Safety

The object whose reference is passed to the closure is uninitialized. Therefore you must not access the content of the object.

You must return a reference to an element from the parameter. The closure must not panic.

pub unsafe fn slice_custom_mut<F, R: ?Sized>(
&mut self,
f: F
) -> BufferMutSlice<R> where
F: for<'r> FnOnce(&'r T) -> &'r R,
R: Content,

Same as slice_custom but returns a mutable slice.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

pub fn as_slice(&self) -> BufferSlice<T>

Builds a slice containing the whole subbuffer.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

pub fn as_mut_slice(&mut self) -> BufferMutSlice<T>

Builds a slice containing the whole subbuffer.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

pub fn as_slice_any(&self) -> BufferAnySlice

Builds a slice-any containing the whole subbuffer.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

impl<T> Buffer<T> where
T: Content + Copy,

pub fn empty<F: ?Sized>(
facade: &F,
ty: BufferType,
mode: BufferMode
) -> Result<Buffer<T>, BufferCreationError> where
F: Facade,

Builds a new buffer of the given size.

impl<T> Buffer<[T]> where
[T]: Content,
T: Copy,

pub fn empty_array<F: ?Sized>(
facade: &F,
ty: BufferType,
len: usize,
mode: BufferMode
) -> Result<Buffer<[T]>, BufferCreationError> where
F: Facade,

Builds a new buffer of the given size.

pub fn len(&self) -> usize

Returns the number of elements in this buffer.

pub fn slice<R: RangeArgument<usize>>(
&self,
range: R
) -> Option<BufferSlice<[T]>>

Builds a slice of this subbuffer. Returns None if out of range.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

pub fn slice_mut<R: RangeArgument<usize>>(
&mut self,
range: R
) -> Option<BufferMutSlice<[T]>>

Builds a slice of this subbuffer. Returns None if out of range.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

impl<T> Buffer<[T]> where
T: PixelValue,

pub fn read_as_texture_1d<S>(&self) -> Result<S, ReadError> where
S: Texture1dDataSink<T>,

Reads the content of the buffer.