Struct glium::buffer::BufferAny

pub struct BufferAny {
    // some fields omitted
}

Represents a sub-part of a buffer.

Doesn't contain any information about the content, contrary to Buffer.

Methods

impl BufferAny

fn as_slice_any(&self) -> BufferAnySlice

Builds a slice-any containing the whole subbuffer.

fn get_elements_size(&self) -> usize

Returns the size in bytes of each element in the buffer.

fn get_elements_count(&self) -> usize

Returns the number of elements in the buffer.

fn get_context(&self) -> &Rc<Context>

Returns the context corresponding to this buffer.

fn get_size(&self) -> usize

Returns the number of bytes in this subbuffer.

fn invalidate(&self)

Invalidates the content of the buffer. The data becomes undefined.

This operation is a no-op if the backend doesn't support it and for persistent-mapped buffers.

unsafe fn read<T>(&self) -> Result<T::Owned, ReadError> where T: Content

UNSTABLE. This function can be removed at any moment without any further notice.

Considers that the buffer is filled with elements of type T and reads them.

Panic

Panicks if the size of the buffer is not a multiple of the size of the data. For example, trying to read some (u8, u8, u8, u8)s from a buffer of 7 bytes will panic.