Struct glium::buffer::BufferViewMutSlice

pub struct BufferViewMutSlice<'a, T: ?Sized> where
    T: Content,  { /* fields omitted */ }

Represents a sub-part of a buffer.

Methods

impl<'a, T: ?Sized> BufferMutSlice<'a, T> where
T: Content + 'a,

pub fn get_size(&self) -> usize

Returns the size in bytes of this slice.

pub fn map(self) -> Mapping<'a, T>

Maps the buffer in memory for both reading and writing.

Implementation

For persistent-mapped buffers, waits until the data is no longer accessed by the GPU then returns a pointer to the existing mapping.

For immutable buffers, creates a temporary buffer containing the data of the buffer and maps it. When the mapping object is destroyed, copies the content of the temporary buffer to the real buffer.

For other types, calls glMapBuffer or glMapSubBuffer.

pub fn map_read(self) -> ReadMapping<'a, T>

Maps the buffer in memory for reading.

Implementation

For persistent-mapped buffers, waits until the data is no longer accessed by the GPU then returns a pointer to the existing mapping.

For immutable buffers, creates a temporary buffer containing the data of the buffer and maps it.

For other types, calls glMapBuffer or glMapSubBuffer.

pub fn map_write(self) -> WriteMapping<'a, T>

Maps the buffer in memory for writing only.

Implementation

For persistent-mapped buffers, waits until the data is no longer accessed by the GPU then returns a pointer to the existing mapping.

For immutable buffers, creates a temporary buffer and maps it. When the mapping object is destroyed, copies the content of the temporary buffer to the real buffer.

For other types, calls glMapBuffer or glMapSubBuffer.

pub fn write(&self, data: &T)

Uploads some data in this buffer.

Implementation

For persistent-mapped buffers, waits untils the data is no longer used by the GPU then memcpies the data to the mapping.

For immutable buffers, creates a temporary buffer that contains the data then calls glCopyBufferSubData to copy from the temporary buffer to the real one.

For other types, calls glBufferSubData.

Panic

Panics if the length of data is different from the length of this buffer.

pub fn invalidate(&self)

Invalidates the content of the slice. The data becomes undefined.

This operation is a no-op if the backend doesn't support it and for persistent-mapped buffers.

Implementation

Calls glInvalidateBufferSubData if supported.

pub fn read(&self) -> Result<T::Owned, ReadError>

Reads the content of the buffer.

pub fn copy_to<S>(&self, target: S) -> Result<(), CopyError> where
S: Into<BufferSlice<'a, T>>,

Copies the content of this slice to another slice.

Panic

Panics if T is unsized and the other buffer is too small.

pub unsafe fn slice_custom<F, R: ?Sized>(self, f: F) -> BufferMutSlice<'a, R> where
F: for<'r> FnOnce(&'r T) -> &'r R,
R: Content,

Builds a slice that contains an element from inside the buffer.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

Example

#[derive(Copy, Clone)]
struct BufferContent {
    value1: u16,
    value2: u16,
}
let slice = unsafe { buffer.slice_custom(|content| &content.value2) };

Safety

The object whose reference is passed to the closure is uninitialized. Therefore you must not access the content of the object.

You must return a reference to an element from the parameter. The closure must not panic.

pub fn as_slice_any(self) -> BufferAnySlice<'a>

Builds a slice-any containing the whole subbuffer.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

impl<'a, T> BufferMutSlice<'a, [T]> where
[T]: Content,
T: Copy + 'a,

pub fn len(&self) -> usize

Returns the number of elements in this slice.

pub fn slice<R: RangeArgument<usize>>(
self,
range: R
) -> Option<BufferMutSlice<'a, [T]>>

Builds a subslice of this slice. Returns None if out of range.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

impl<'a, T> BufferMutSlice<'a, [T]> where
T: PixelValue + 'a,

pub fn read_as_texture_1d<S>(&self) -> Result<S, ReadError> where
S: Texture1dDataSink<T>,

Reads the content of the buffer.