Struct glium::buffer::BufferViewSlice

pub struct BufferViewSlice<'a, T: ?Sized> where T: Content + 'a {
    // some fields omitted
}

Represents a sub-part of a buffer.

Methods

impl<'a, T: ?Sized> BufferSlice<'a, T> where T: Content + 'a

fn get_size(&self) -> usize

Returns the size in bytes of this slice.

fn get_context(&self) -> &Rc<Context>

Returns the context corresponding to this buffer.

fn write(&self, data: &T)

Uploads some data in this buffer.

Implementation

For persistent-mapped buffers, waits untils the data is no longer used by the GPU then memcpies the data to the mapping.

For immutable buffers, creates a temporary buffer that contains the data then calls glCopyBufferSubData to copy from the temporary buffer to the real one.

For other types, calls glBufferSubData.

Panic

Panics if the length of data is different from the length of this buffer.

fn invalidate(&self)

Invalidates the content of the slice. The data becomes undefined.

This operation is a no-op if the backend doesn't support it and for persistent-mapped buffers.

Implementation

Calls glInvalidateBufferSubData if supported.

fn read(&self) -> Result<T::Owned, ReadError>

Reads the content of the buffer.

fn copy_to<S>(&self, target: S) -> Result<(), CopyError> where S: Into<BufferSlice<'a, T>>

Copies the content of this slice to another slice.

Panic

Panics if T is unsized and the other buffer is too small.

unsafe fn slice_custom<F, R: ?Sized>(&self, f: F) -> BufferSlice<'a, R> where F: for<'r> FnOnce(&'r T) -> &'r R, R: Content

Builds a slice that contains an element from inside the buffer.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

Example

#[derive(Copy, Clone)]
struct BufferContent {
    value1: u16,
    value2: u16,
}
let slice = unsafe { buffer.slice_custom(|content| &content.value2) };

Safety

The object whose reference is passed to the closure is uninitialized. Therefore you must not access the content of the object.

You must return a reference to an element from the parameter. The closure must not panic.

fn as_slice_any(&self) -> BufferAnySlice<'a>

Builds a slice-any containing the whole subbuffer.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

impl<'a, T> BufferSlice<'a, [T]> where [T]: Content + 'a

fn len(&self) -> usize

Returns the number of elements in this slice.

fn slice<R: RangeArgument<usize>>(&self, range: R) -> Option<BufferSlice<'a, [T]>>

Builds a subslice of this slice. Returns None if out of range.

This method builds an object that represents a slice of the buffer. No actual operation OpenGL is performed.

impl<'a, T> BufferSlice<'a, [T]> where T: PixelValue + 'a

fn read_as_texture_1d<S>(&self) -> Result<S, ReadError> where S: Texture1dDataSink<T>

Reads the content of the buffer.