Struct glium::buffer::Builder

pub struct Builder<R> {
    // some fields omitted
}

Struct that allows you to build multiple subbuffers stored inside the same buffer.

Example

let (vb1, vb2) = glium::buffer::Builder::new(&display)
                        .add_empty(12)    // vb1
                        .add_empty(16)    // vb2
                        .build();

// rustc needs some hints about the types of the buffers,
// so you may have to write something like this:
let vb1: glium::vertex::VertexBuffer<Vertex> = vb1;
let vb2: glium::vertex::VertexBuffer<Vertex> = vb2;

Methods

impl Builder<()>

fn new<F>(facade: &F) -> Builder<()> where F: Facade

Builds a new empty builder.

impl<R> Builder<R>

fn add_empty<'a, O>(self, len: usize) -> Builder<R::Output> where R: BuilderTupleAdd<'a, O>

Adds an empty buffer to the list of buffers to create.

fn add_data<'a, O>(self, data: &'a [O]) -> Builder<R::Output> where R: BuilderTupleAdd<'a, O>

Adds a buffer with some data to the list of buffers to create.

fn build<O>(self) -> O where R: BuilderParamsList<O>

Consumes the builder and returns the generated buffers.