Struct glium::buffer::SubBuffer

pub struct SubBuffer<T> where T: Copy + Send + 'static {
    // some fields omitted
}

Represents a sub-part of a buffer.

Methods

impl<T> SubBuffer<T> where T: Copy + Send + 'static

fn new<F>(facade: &F, data: &[T], ty: BufferType, dynamic: bool) -> Result<SubBuffer<T>, BufferCreationError> where F: Facade

Builds a new buffer containing the given data. The size of the buffer is equal to the size of the data.

If dynamic is false, glium will attempt to use buffer storage to create an immutable buffer (without the DYNAMIC_STORAGE_FLAG).

If dynamic is true, glium will attempt to use persistent mapping and manage synchronizations manually.

fn empty<F>(facade: &F, ty: BufferType, len: usize, dynamic: bool) -> Result<SubBuffer<T>, BufferCreationError> where F: Facade

Builds a new buffer of the given size.

If dynamic is false, glium will attempt to use buffer storage to create an immutable buffer (without the DYNAMIC_STORAGE_FLAG).

If dynamic is true, glium will attempt to use persistent mapping and manage synchronizations manually.

fn get_context(&self) -> &Rc<Context>

Returns the context corresponding to this buffer.

fn len(&self) -> usize

Returns the number of elements in this subbuffer.

fn is_persistent(&self) -> bool

Returns true if this buffer uses persistent mapping.

fn write<P>(&self, data: P) where P: AsRef<[T]>

Uploads some data in this buffer.

Panic

Panics if the length of data is different from the length of this buffer.

fn invalidate(&self)

Invalidates the content of the buffer. The data becomes undefined.

You should call this if you only use parts of a buffer. For example if you want to use the first half of the buffer, you invalidate the whole buffer then write the first half.

This operation is a no-op if the backend doesn't support it.

fn read(&self) -> Vec<T>

Reads the content of the buffer.

Features

Only available if the gl_read_buffer feature is enabled.

fn map(&mut self) -> Mapping<T>

Maps the buffer in memory.

fn read_if_supported(&self) -> Option<Vec<T>>

Reads the content of the subbuffer. Returns None if this operation is not supported.

fn slice(&self, range: Range<usize>) -> Option<SubBufferSlice<T>>

Builds a slice of this subbuffer. Returns None if out of range.

fn slice_mut(&mut self, range: Range<usize>) -> Option<SubBufferMutSlice<T>>

Builds a slice of this subbuffer. Returns None if out of range.

fn as_slice(&self) -> SubBufferSlice<T>

Builds a slice containing the whole subbuffer.

fn as_mut_slice(&mut self) -> SubBufferMutSlice<T>

Builds a slice containing the whole subbuffer.

fn as_slice_any(&self) -> SubBufferAnySlice

Builds a slice-any containing the whole subbuffer.