Struct glium::buffer::SubBufferAny

pub struct SubBufferAny {
    // some fields omitted
}

Represents a sub-part of a buffer.

Doesn't contain any information about the content, contrary to SubBuffer.

Methods

impl SubBufferAny

fn as_slice_any(&self) -> SubBufferAnySlice

Builds a slice-any containing the whole subbuffer.

fn get_context(&self) -> &Rc<Context>

Returns the context corresponding to this buffer.

fn get_elements_size(&self) -> usize

Returns the size of each element in this buffer.

This information is taken from the original SubBuffer that was used to construct this object.

fn get_elements_count(&self) -> usize

Returns the number of elements in this buffer.

This information is taken from the original SubBuffer that was used to construct this object.

fn get_size(&self) -> usize

Returns the number of bytes in this subbuffer.

fn invalidate(&self)

Invalidates the content of the buffer. The data becomes undefined.

This operation is a no-op if the backend doesn't support it.

unsafe fn read_if_supported<T>(&self) -> Option<Vec<T>> where T: Copy + Send + 'static

UNSTABLE. This function can be removed at any moment without any further notice.

Considers that the buffer is filled with elements of type T and reads them.

Panic

Panicks if the size of the buffer is not a multiple of the size of the data. For example, trying to read some (u8, u8, u8, u8)s from a buffer of 7 bytes will panic.