Struct glium::debug::TimestampQuery

pub struct TimestampQuery {
    // some fields omitted
}

Allows you to obtain the timestamp inside the OpenGL commands queue.

When you call functions in glium, they are not instantly executed. Instead they are added in a commands queue that the backend executes asynchronously.

When you call TimestampQuery::new, a command is added to this list asking the backend to send us the current timestamp. Thanks to this, you can know how much time it takes to execute commands.

Example

let before = glium::debug::TimestampQuery::new(&display);
// do some stuff here
let after = glium::debug::TimestampQuery::new(&display);

match (after, before) {
    (Some(after), Some(before)) => {
        let elapsed = after.get() - before.get();
        println!("Time it took to do stuff: {}", elapsed);
    },
    _ => ()
}

Methods

impl TimestampQuery

fn new<F>(facade: &F) -> Option<TimestampQuery> where F: Facade

Creates a new TimestampQuery. Returns None if the backend doesn't support it.

fn is_ready(&self) -> bool

Queries the counter to see if the timestamp is already available.

It takes some time to retreive the value, during which you can execute other functions.

fn get(self) -> u64

Returns the value of the timestamp. Blocks until it is available.

This function doesn't block if is_ready returns true.