Type Definition glium::debug::DebugCallback

type DebugCallback = Box<dyn FnMut(Source, MessageType, Severity, u32, bool, &str)>;

Represents a callback that can be used for the debug output feature of OpenGL.

The first three parameters are self-explanatory. The fourth parameter is an identifier for this message whose meaning is implementation-defined. The fifth parameter indicates whether glium is already handling any possible error condition, so you don't need to print an error. The last parameter is a message generated by the OpenGL implementation.