Struct glium::draw_parameters::PrimitivesGeneratedQuery

pub struct PrimitivesGeneratedQuery { /* fields omitted */ }

Query that allows you to know the number of primitives generated by the geometry shader. Will stay at 0 if you use it without any active geometry shader.

Methods

impl PrimitivesGeneratedQuery

pub fn new<F: ?Sized>(
facade: &F
) -> Result<PrimitivesGeneratedQuery, QueryCreationError> where
F: Facade,

Builds a new query.

impl PrimitivesGeneratedQuery

pub fn is_ready(&self) -> bool

Queries the counter to see if the result is already available.

pub fn get(self) -> u32

Returns the value of the query. Blocks until it is available.

This function doesn't block if is_ready would return true.

Note that you are strongly discouraged from calling this in the middle of the rendering process, as it may block for a long time.

Queries should either have their result written into a buffer, be used for conditional rendering, or stored and checked during the next frame.

pub fn to_buffer_u32(
&self,
target: BufferSlice<u32>
) -> Result<(), ToBufferError>

Writes the result of the query to a buffer when it is available.

This function doesn't block. Instead it submits a commands to the GPU's commands queue and orders the GPU to write the result of the query to a buffer.

This operation is not necessarily supported everywhere.