Enum glium::PolygonModeUnstable

pub enum PolygonMode {
    Point,
    Line,
    Fill,
}

Defines how the device should render polygons.

The usual value is Fill, which fills the content of polygon with the color. However other values are sometimes useful, especially for debugging purposes.

Example

The same triangle drawn respectively with Fill, Line and Point (barely visible).

Variants

Point

Only draw a single point at each vertex.

All attributes that apply to points are used when using this mode.

Line

Only draw a line in the boundaries of each polygon.

All attributes that apply to lines (line_width) are used when using this mode.

Fill

Fill the content of the polygon. This is the default mode.