Enum glium::StencilOperation

pub enum StencilOperation {
    Keep,
    Zero,
    Replace,
    Increment,
    IncrementWrap,
    Decrement,
    DecrementWrap,
    Invert,
}

Specificies which operation the GPU will do depending on the result of the stencil test.

Variants

Keep

Keeps the value currently in the stencil buffer.

Zero

Writes zero in the stencil buffer.

Replace

Writes the reference value (stencil_reference_value_clockwise or stencil_reference_value_counter_clockwise) in the stencil buffer.

Increment

Increments the value currently in the stencil buffer. If the value is the maximum, don't do anything.

IncrementWrap

Increments the value currently in the stencil buffer. If the value is the maximum, wrap to 0.

Decrement

Decrements the value currently in the stencil buffer. If the value is 0, don't do anything.

DecrementWrap

Decrements the value currently in the stencil buffer. If the value is 0, wrap to -1.

Invert

Inverts each bit of the value.