Function glium::get_supported_glsl_version

pub fn get_supported_glsl_version(gl_version: &Version) -> Version

Given an API version, this function returns the GLSL version that the implementation is required to support.

Panic

Panics if the version is invalid or is not supposed to support a GLSL version.