Enum glium::framebuffer::ValidationError

pub enum ValidationError {
    EmptyFramebufferObjectsNotSupported,
    EmptyFramebufferUnsupportedDimensions,
    DimensionsMismatchNotSupported,
    SamplesCountMismatch,
    TooManyColorAttachments {
        maximum: usize,
        obtained: usize,
    },
}

An error that can happen while validating attachments.

Variants
EmptyFramebufferObjectsNotSupported

You requested an empty framebuffer object, but they are not supported.

EmptyFramebufferUnsupportedDimensions

The requested characteristics of an empty framebuffer object are out of range.

DimensionsMismatchNotSupported

The backend doesn't support attachments with various dimensions.

Note that almost all OpenGL implementations support attachments with various dimensions. Only very old versions don't.

SamplesCountMismatch

All attachments must have the same number of samples.

TooManyColorAttachments

Backends only support a certain number of color attachments.

Fields of TooManyColorAttachments

maximum: usize

Maximum number of attachments.

obtained: usize

Number of attachments that were given.