Struct glium::framebuffer::DepthRenderBuffer

pub struct DepthRenderBuffer {
    // some fields omitted
}

A render buffer is similar to a texture, but is optimized for usage as a draw target.

Contrary to a texture, you can't sample or modify the content of the DepthRenderBuffer directly.

Methods

impl DepthRenderBuffer

fn new<F>(facade: &F, format: DepthFormat, width: u32, height: u32) -> Result<DepthRenderBuffer, CreationError> where F: Facade

Builds a new render buffer.