Struct glium::framebuffer::EmptyFrameBuffer

pub struct EmptyFrameBuffer {
    // some fields omitted
}

A framebuffer with no attachment at all.

Note that this is only supported on recent hardware.

Methods

impl<'a> EmptyFrameBuffer

fn is_supported<C>(context: &C) -> bool where C: CapabilitiesSource

Returns true if empty framebuffers are supported by the backend.

fn is_layered_supported<C>(context: &C) -> bool where C: CapabilitiesSource

Returns true if layered empty framebuffers are supported by the backend.

fn get_max_supported_width<C>(context: &C) -> Option<u32> where C: CapabilitiesSource

Returns the maximum width of empty framebuffers that the backend supports, or None if empty framebuffers are not supported.

fn get_max_supported_height<C>(context: &C) -> Option<u32> where C: CapabilitiesSource

Returns the maximum height of empty framebuffers that the backend supports, or None if empty framebuffers are not supported.

fn get_max_supported_samples<C>(context: &C) -> Option<u32> where C: CapabilitiesSource

Returns the maximum number of samples of empty framebuffers that the backend supports, or None if empty framebuffers are not supported.

fn get_max_supported_layers<C>(context: &C) -> Option<u32> where C: CapabilitiesSource

Returns the maximum number of layers of empty framebuffers that the backend supports, or None if layered empty framebuffers are not supported.

fn new<F>(facade: &F, width: u32, height: u32, layers: Option<u32>, samples: Option<u32>, fixed_samples: bool) -> Result<EmptyFrameBuffer, ValidationError> where F: Facade

Creates a EmptyFrameBuffer.

Panic

Panicks if layers or samples is equal to Some(0).