Struct glium::framebuffer::MultiOutputFrameBuffer

pub struct MultiOutputFrameBuffer<'a> {
    // some fields omitted
}

This struct is useless for the moment.

Methods

impl<'a> MultiOutputFrameBuffer<'a>

fn new<F, I, A>(facade: &F, color_attachments: I) -> Result<MultiOutputFrameBuffer<'a>, ValidationError> where F: Facade, I: IntoIterator<Item=(&'a str, A)>, A: ToColorAttachment<'a>

Creates a new MultiOutputFrameBuffer.

Panic

Panics if all attachments don't have the same dimensions.

fn with_depth_buffer<F, D, I, A>(facade: &F, color_attachments: I, depth: D) -> Result<MultiOutputFrameBuffer<'a>, ValidationError> where F: Facade, D: ToDepthAttachment<'a>, I: IntoIterator<Item=(&'a str, A)>, A: ToColorAttachment<'a>

Creates a MultiOutputFrameBuffer with a depth buffer.

Panic

Panics if all attachments don't have the same dimensions.

fn with_depth_and_stencil_buffer<A, F, I, D, S>(facade: &F, color: I, depth: D, stencil: S) -> Result<MultiOutputFrameBuffer<'a>, ValidationError> where D: ToDepthAttachment<'a>, I: IntoIterator<Item=(&'a str, A)>, S: ToStencilAttachment<'a>, A: ToColorAttachment<'a>, F: Facade

Creates a MultiOutputFrameBuffer with a depth buffer, and a stencil buffer.

Panic

Panics if all attachments don't have the same dimensions.

fn with_stencil_buffer<A, F, I, S>(facade: &F, color: I, stencil: S) -> Result<MultiOutputFrameBuffer<'a>, ValidationError> where S: ToStencilAttachment<'a>, F: Facade, I: IntoIterator<Item=(&'a str, A)>, A: ToColorAttachment<'a>

Creates a MultiOutputFrameBuffer with a stencil buffer, but no depth buffer.

Panic

Panics if all attachments don't have the same dimensions.

fn with_depth_stencil_buffer<A, F, I, D>(facade: &F, color: I, depthstencil: D) -> Result<MultiOutputFrameBuffer<'a>, ValidationError> where D: ToDepthStencilAttachment<'a>, F: Facade, I: IntoIterator<Item=(&'a str, A)>, A: ToColorAttachment<'a>

Creates a MultiOutputFrameBuffer with a depth-stencil buffer.

Panic

Panics if all attachments don't have the same dimensions.