Struct glium::framebuffer::SimpleFrameBuffer

pub struct SimpleFrameBuffer<'a> { /* fields omitted */ }

A framebuffer which has only one color attachment.

Implementations

impl<'a> SimpleFrameBuffer<'a>

pub fn new<F: ?Sized, C>(
facade: &F,
color: C
) -> Result<SimpleFrameBuffer<'a>, ValidationError> where
C: ToColorAttachment<'a>,
F: Facade,

Creates a SimpleFrameBuffer with a single color attachment and no depth nor stencil buffer.

pub fn with_depth_buffer<F: ?Sized, C, D>(
facade: &F,
color: C,
depth: D
) -> Result<SimpleFrameBuffer<'a>, ValidationError> where
C: ToColorAttachment<'a>,
D: ToDepthAttachment<'a>,
F: Facade,

Creates a SimpleFrameBuffer with a single color attachment and a depth buffer, but no stencil buffer.

pub fn depth_only<F: ?Sized, D>(
facade: &F,
depth: D
) -> Result<SimpleFrameBuffer<'a>, ValidationError> where
D: ToDepthAttachment<'a>,
F: Facade,

Creates a SimpleFrameBuffer with a single color attachment and no depth nor stencil buffer.

pub fn with_depth_and_stencil_buffer<F: ?Sized, C, D, S>(
facade: &F,
color: C,
depth: D,
stencil: S
) -> Result<SimpleFrameBuffer<'a>, ValidationError> where
C: ToColorAttachment<'a>,
D: ToDepthAttachment<'a>,
S: ToStencilAttachment<'a>,
F: Facade,

Creates a SimpleFrameBuffer with a single color attachment, a depth buffer, and a stencil buffer.

pub fn depth_and_stencil_only<F: ?Sized, D, S>(
facade: &F,
depth: D,
stencil: S
) -> Result<SimpleFrameBuffer<'a>, ValidationError> where
D: ToDepthAttachment<'a>,
S: ToStencilAttachment<'a>,
F: Facade,

Creates a SimpleFrameBuffer with a single color attachment and no depth nor stencil buffer.

pub fn with_stencil_buffer<F: ?Sized, C, S>(
facade: &F,
color: C,
stencil: S
) -> Result<SimpleFrameBuffer<'a>, ValidationError> where
C: ToColorAttachment<'a>,
S: ToStencilAttachment<'a>,
F: Facade,

Creates a SimpleFrameBuffer with a single color attachment and a stencil buffer, but no depth buffer.

pub fn stencil_only<F: ?Sized, S>(
facade: &F,
stencil: S
) -> Result<SimpleFrameBuffer<'a>, ValidationError> where
S: ToStencilAttachment<'a>,
F: Facade,

Creates a SimpleFrameBuffer with a single color attachment and a stencil buffer, but no depth buffer.

pub fn with_depth_stencil_buffer<F: ?Sized, C, D>(
facade: &F,
color: C,
depthstencil: D
) -> Result<SimpleFrameBuffer<'a>, ValidationError> where
C: ToColorAttachment<'a>,
D: ToDepthStencilAttachment<'a>,
F: Facade,

Creates a SimpleFrameBuffer with a single color attachment and a depth-stencil buffer.

pub fn depth_stencil_only<F: ?Sized, D>(
facade: &F,
depthstencil: D
) -> Result<SimpleFrameBuffer<'a>, ValidationError> where
D: ToDepthStencilAttachment<'a>,
F: Facade,

Creates a SimpleFrameBuffer with a single color attachment and a depth-stencil buffer.