Enum glium::glutin::DeviceEvent

pub enum DeviceEvent {
    Added,
    Removed,
    MouseMotion {
        delta: (f64, f64),
    },
    MouseWheel {
        delta: MouseScrollDelta,
    },
    Motion {
        axis: u32,
        value: f64,
    },
    Button {
        button: u32,
        state: ElementState,
    },
    Key(KeyboardInput),
    Text {
        codepoint: char,
    },
}

Represents raw hardware events that are not associated with any particular window.

Useful for interactions that diverge significantly from a conventional 2D GUI, such as 3D camera or first-person game controls. Many physical actions, such as mouse movement, can produce both device and window events. Because window events typically arise from virtual devices (corresponding to GUI cursors and keyboard focus) the device IDs may not match.

Note that these events are delivered regardless of input focus.

Variants
AddedRemovedMouseMotion

Change in physical position of a pointing device.

This represents raw, unfiltered physical motion. Not to be confused with WindowEvent::CursorMoved.

Fields of MouseMotion

delta: (f64, f64)

(x, y) change in position in unspecified units.

Different devices may use different units.

MouseWheel

Physical scroll event

Fields of MouseWheel

delta: MouseScrollDelta

Motion

Motion on some analog axis. This event will be reported for all arbitrary input devices that winit supports on this platform, including mouse devices. If the device is a mouse device then this will be reported alongside the MouseMotion event.

Fields of Motion

axis: u32

value: f64

Button
Fields of Button

button: u32

state: ElementState

Key(KeyboardInput)Text
Fields of Text

codepoint: char