Enum glium::glutin::WindowEvent

pub enum WindowEvent {
    Resized(u32, u32),
    Moved(i32, i32),
    Closed,
    DroppedFile(PathBuf),
    HoveredFile(PathBuf),
    HoveredFileCancelled,
    ReceivedCharacter(char),
    Focused(bool),
    KeyboardInput {
        device_id: DeviceId,
        input: KeyboardInput,
    },
    CursorMoved {
        device_id: DeviceId,
        position: (f64, f64),
        modifiers: ModifiersState,
    },
    CursorEntered {
        device_id: DeviceId,
    },
    CursorLeft {
        device_id: DeviceId,
    },
    MouseWheel {
        device_id: DeviceId,
        delta: MouseScrollDelta,
        phase: TouchPhase,
        modifiers: ModifiersState,
    },
    MouseInput {
        device_id: DeviceId,
        state: ElementState,
        button: MouseButton,
        modifiers: ModifiersState,
    },
    TouchpadPressure {
        device_id: DeviceId,
        pressure: f32,
        stage: i64,
    },
    AxisMotion {
        device_id: DeviceId,
        axis: u32,
        value: f64,
    },
    Refresh,
    Touch(Touch),
    HiDPIFactorChanged(f32),
}

Describes an event from a Window.

Variants
Resized(u32, u32)

The size of the window has changed.

Moved(i32, i32)

The position of the window has changed.

Closed

The window has been closed.

DroppedFile(PathBuf)

A file has been dropped into the window.

HoveredFile(PathBuf)

A file is being hovered over the window.

HoveredFileCancelled

A file was hovered, but has exited the window.

ReceivedCharacter(char)

The window received a unicode character.

Focused(bool)

The window gained or lost focus.

The parameter is true if the window has gained focus, and false if it has lost focus.

KeyboardInput

An event from the keyboard has been received.

Fields of KeyboardInput

device_id: DeviceId

input: KeyboardInput

CursorMoved

The cursor has moved on the window.

Fields of CursorMoved

device_id: DeviceId

position: (f64, f64)

(x,y) coords in pixels relative to the top-left corner of the window. Because the range of this data is limited by the display area and it may have been transformed by the OS to implement effects such as cursor acceleration, it should not be used to implement non-cursor-like interactions such as 3D camera control.

modifiers: ModifiersState

CursorEntered

The cursor has entered the window.

Fields of CursorEntered

device_id: DeviceId

CursorLeft

The cursor has left the window.

Fields of CursorLeft

device_id: DeviceId

MouseWheel

A mouse wheel movement or touchpad scroll occurred.

Fields of MouseWheel

device_id: DeviceId

delta: MouseScrollDelta

phase: TouchPhase

modifiers: ModifiersState

MouseInput

An mouse button press has been received.

Fields of MouseInput

device_id: DeviceId

state: ElementState

button: MouseButton

modifiers: ModifiersState

TouchpadPressure

Touchpad pressure event.

At the moment, only supported on Apple forcetouch-capable macbooks. The parameters are: pressure level (value between 0 and 1 representing how hard the touchpad is being pressed) and stage (integer representing the click level).

Fields of TouchpadPressure

device_id: DeviceId

pressure: f32

stage: i64

AxisMotion

Motion on some analog axis. May report data redundant to other, more specific events.

Fields of AxisMotion

device_id: DeviceId

axis: u32

value: f64

Refresh

The window needs to be redrawn.

Touch(Touch)

Touch event has been received

HiDPIFactorChanged(f32)

DPI scaling factor of the window has changed.

The following actions cause DPI changes:

A user changes the resolution.

A user changes the desktop scaling value (e.g. in Control Panel on Windows).

A user moves the application window to a display with a different DPI.