Struct glium::glutin::PixelFormatRequirements

pub struct PixelFormatRequirements {
    pub hardware_accelerated: Option<bool>,
    pub color_bits: Option<u8>,
    pub float_color_buffer: bool,
    pub alpha_bits: Option<u8>,
    pub depth_bits: Option<u8>,
    pub stencil_bits: Option<u8>,
    pub double_buffer: Option<bool>,
    pub multisampling: Option<u16>,
    pub stereoscopy: bool,
    pub srgb: bool,
    pub release_behavior: ReleaseBehavior,
}

Describes how the backend should choose a pixel format.

Fields
hardware_accelerated: Option<bool>

If true, only hardware-accelerated formats will be considered. If false, only software renderers. None means "don't care". Default is Some(true).

color_bits: Option<u8>

Minimum number of bits for the color buffer, excluding alpha. None means "don't care". The default is Some(24).

float_color_buffer: bool

If true, the color buffer must be in a floating point format. Default is false.

Using floating points allows you to write values outside of the [0.0, 1.0] range.

alpha_bits: Option<u8>

Minimum number of bits for the alpha in the color buffer. None means "don't care". The default is Some(8).

depth_bits: Option<u8>

Minimum number of bits for the depth buffer. None means "don't care". The default value is Some(24).

stencil_bits: Option<u8>

Minimum number of bits for the depth buffer. None means "don't care". The default value is Some(8).

double_buffer: Option<bool>

If true, only double-buffered formats will be considered. If false, only single-buffer formats. None means "don't care". The default is Some(true).

multisampling: Option<u16>

Contains the minimum number of samples per pixel in the color, depth and stencil buffers. None means "don't care". Default is None. A value of Some(0) indicates that multisampling must not be enabled.

stereoscopy: bool

If true, only stereoscopic formats will be considered. If false, only non-stereoscopic formats. The default is false.

srgb: bool

If true, only sRGB-capable formats will be considered. If false, don't care. The default is false.

release_behavior: ReleaseBehavior

The behavior when changing the current context. Default is Flush.