Struct glium::glutin::WindowAttributes

pub struct WindowAttributes {
    pub dimensions: Option<(u32, u32)>,
    pub min_dimensions: Option<(u32, u32)>,
    pub max_dimensions: Option<(u32, u32)>,
    pub fullscreen: Option<MonitorId>,
    pub title: String,
    pub maximized: bool,
    pub visible: bool,
    pub transparent: bool,
    pub decorations: bool,
    pub multitouch: bool,
}

Attributes to use when creating a window.

Fields
dimensions: Option<(u32, u32)>

The dimensions of the window. If this is None, some platform-specific dimensions will be used.

The default is None.

min_dimensions: Option<(u32, u32)>

The minimum dimensions a window can be, If this is None, the window will have no minimum dimensions (aside from reserved).

The default is None.

max_dimensions: Option<(u32, u32)>

The maximum dimensions a window can be, If this is None, the maximum will have no maximum or will be set to the primary monitor's dimensions by the platform.

The default is None.

fullscreen: Option<MonitorId>

Whether the window should be set as fullscreen upon creation.

The default is None.

title: String

The title of the window in the title bar.

The default is "winit window".

maximized: bool

Whether the window should be maximized upon creation.

The default is false.

visible: bool

Whether the window should be immediately visible upon creation.

The default is true.

transparent: bool

Whether the the window should be transparent. If this is true, writing colors with alpha values different than 1.0 will produce a transparent window.

The default is false.

decorations: bool

Whether the window should have borders and bars.

The default is true.

multitouch: bool

[iOS only] Enable multitouch, see multipleTouchEnabled