Struct glium::index::DrawCommandsIndicesBuffer

pub struct DrawCommandsIndicesBuffer {
    // some fields omitted
}

A buffer containing a list of draw commands.

Methods

impl DrawCommandsIndicesBuffer

fn empty<F>(facade: &F, elements: usize) -> Result<DrawCommandsIndicesBuffer, BufferCreationError> where F: Facade

Builds an empty buffer.

The parameter indicates the number of elements.

fn empty_dynamic<F>(facade: &F, elements: usize) -> Result<DrawCommandsIndicesBuffer, BufferCreationError> where F: Facade

Builds an empty buffer.

The parameter indicates the number of elements.

fn empty_persistent<F>(facade: &F, elements: usize) -> Result<DrawCommandsIndicesBuffer, BufferCreationError> where F: Facade

Builds an empty buffer.

The parameter indicates the number of elements.

fn empty_immutable<F>(facade: &F, elements: usize) -> Result<DrawCommandsIndicesBuffer, BufferCreationError> where F: Facade

Builds an empty buffer.

The parameter indicates the number of elements.

fn with_index_buffer<'a, T>(&'a self, index_buffer: &'a IndexBuffer<T>) -> IndicesSource<'a> where T: Index

Builds an indices source from this buffer and a primitives type. This indices source can be passed to the draw() function.