Struct glium::index::DrawCommandsNoIndicesBuffer

pub struct DrawCommandsNoIndicesBuffer {
    // some fields omitted
}

A buffer containing a list of draw commands.

Methods

impl DrawCommandsNoIndicesBuffer

fn empty<F>(facade: &F, elements: usize) -> Result<DrawCommandsNoIndicesBuffer, BufferCreationError> where F: Facade

Builds an empty buffer.

The parameter indicates the number of elements.

fn empty_dynamic<F>(facade: &F, elements: usize) -> Result<DrawCommandsNoIndicesBuffer, BufferCreationError> where F: Facade

Builds an empty buffer.

The parameter indicates the number of elements.

fn empty_persistent<F>(facade: &F, elements: usize) -> Result<DrawCommandsNoIndicesBuffer, BufferCreationError> where F: Facade

Builds an empty buffer.

The parameter indicates the number of elements.

fn empty_immutable<F>(facade: &F, elements: usize) -> Result<DrawCommandsNoIndicesBuffer, BufferCreationError> where F: Facade

Builds an empty buffer.

The parameter indicates the number of elements.

fn with_primitive_type(&self, primitives: PrimitiveType) -> IndicesSource

Builds an indices source from this buffer and a primitives type. This indices source can be passed to the draw() function.