Struct glium::index::IndexBuffer

pub struct IndexBuffer<T> where
    T: Index,  { /* fields omitted */ }

A list of indices loaded in the graphics card's memory.

Methods

impl<T> IndexBuffer<T> where
T: Index,

pub fn new<F: ?Sized>(
facade: &F,
prim: PrimitiveType,
data: &[T]
) -> Result<IndexBuffer<T>, CreationError> where
F: Facade,

Builds a new index buffer from a list of indices and a primitive type.

pub fn dynamic<F: ?Sized>(
facade: &F,
prim: PrimitiveType,
data: &[T]
) -> Result<IndexBuffer<T>, CreationError> where
F: Facade,

Builds a new index buffer from a list of indices and a primitive type.

pub fn persistent<F: ?Sized>(
facade: &F,
prim: PrimitiveType,
data: &[T]
) -> Result<IndexBuffer<T>, CreationError> where
F: Facade,

Builds a new index buffer from a list of indices and a primitive type.

pub fn immutable<F: ?Sized>(
facade: &F,
prim: PrimitiveType,
data: &[T]
) -> Result<IndexBuffer<T>, CreationError> where
F: Facade,

Builds a new index buffer from a list of indices and a primitive type.

pub fn empty<F: ?Sized>(
facade: &F,
prim: PrimitiveType,
len: usize
) -> Result<IndexBuffer<T>, CreationError> where
F: Facade,

Builds a new empty index buffer.

pub fn empty_dynamic<F: ?Sized>(
facade: &F,
prim: PrimitiveType,
len: usize
) -> Result<IndexBuffer<T>, CreationError> where
F: Facade,

Builds a new empty index buffer.

pub fn empty_persistent<F: ?Sized>(
facade: &F,
prim: PrimitiveType,
len: usize
) -> Result<IndexBuffer<T>, CreationError> where
F: Facade,

Builds a new empty index buffer.

pub fn empty_immutable<F: ?Sized>(
facade: &F,
prim: PrimitiveType,
len: usize
) -> Result<IndexBuffer<T>, CreationError> where
F: Facade,

Builds a new empty index buffer.

pub fn get_primitives_type(&self) -> PrimitiveType

Returns the type of primitives associated with this index buffer.

pub fn get_indices_type(&self) -> IndexType

Returns the data type of the indices inside this index buffer.

pub fn slice<R: RangeArgument<usize>>(
&self,
range: R
) -> Option<IndexBufferSlice<T>>

Returns None if out of range.