Struct glium::index::IndexBufferAny

pub struct IndexBufferAny {
    // some fields omitted
}

An IndexBuffer without any type information.

Makes it easier to store in a Vec or return from a function, for example.

Methods

impl IndexBufferAny

fn get_primitives_type(&self) -> PrimitiveType

Returns the type of primitives associated with this index buffer.

fn get_indices_type(&self) -> IndexType

Returns the data type of the indices inside this index buffer.