Struct glium::index::IndexBufferSlice

pub struct IndexBufferSlice<'a, T: 'a> where T: Index {
    // some fields omitted
}

Slice of an IndexBuffer.

Methods

impl<'a, T: 'a> IndexBufferSlice<'a, T> where T: Index

fn get_primitives_type(&self) -> PrimitiveType

Returns the type of primitives associated with this index buffer.

fn get_indices_type(&self) -> IndexType

Returns the data type of the indices inside this index buffer.

fn slice(&self, range: Range<usize>) -> Option<IndexBufferSlice<'a, T>>

Returns None if out of range.