Struct glium::index::NoIndices

pub struct NoIndices(pub PrimitiveType);

Marker that can be used as an indices source when you don't need indices.

If you use this, then the primitives will be constructed using the order in which the vertices are in the vertices sources.