Module glium::index_bufferUnstable

In order to draw, you need to provide a source of indices which is used to link the vertices together into primitives.

There are height types of primitives, each one with a corresponding struct: - PointsList - LinesList - LinesListAdjacency - LineStrip - LineStripAdjacency - TrianglesList - TrianglesListAdjacency - TriangleStrip - TriangleStripAdjacency - TriangleFan

Each struct contains a vector with the indices and can be used as an IndicesSource.

However the most optimal way to draw something is to load the data in the video memory by creating an IndexBuffer.

Structs

IndexBuffer
A list of indices loaded in the graphics card's memory.

LineStrip
A list of lines connected together stored in RAM.

LineStripAdjacency
A list of lines connected together with adjacency infos stored in RAM.

LinesList
A list of lines stored in RAM.

LinesListAdjacency
A list of lines with adjacency infos stored in RAM.

PointsList
A list of points stored in RAM.

TriangleFan
A list of triangles stored in RAM.

TriangleStrip
A list of triangles connected together stored in RAM.

TriangleStripAdjacency
A list of triangles connected together with adjacency infos stored in RAM.

TrianglesList
A list of triangles stored in RAM.

TrianglesListAdjacency
A list of triangles with adjacency infos stored in RAM.

Enums

IndexType
Types of indices in an indices source.

IndicesSource
Describes a source of indices used for drawing.

PrimitiveType
List of available primitives.

Traits

Index
An index from the index buffer.

IntoIndexBuffer
Object is convertible to an index buffer.

ToIndicesSource
Can be used as a source of indices when drawing.