Enum glium::index_buffer::IndicesSourceUnstable

pub enum IndicesSource<'a, T: 'a> {
    IndexBuffer {
        buffer: &'a IndexBuffer,
        offset: usize,
        length: usize,
    },
    Buffer {
        pointer: &'a [T],
        primitives: PrimitiveType,
        offset: usize,
        length: usize,
    },
}

Describes a source of indices used for drawing.

Variants

IndexBuffer

A buffer uploaded in video memory.

Fields

buffer

The buffer.

offset

Offset of the first element of the buffer to use.

length

Number of elements in the buffer to use.

Buffer

A buffer in RAM.

Fields

pointer

Slice of data to use.

primitives

Type of primitives contained in the buffer.

offset

Offset of the first element of the buffer to use.

length

Number of elements in the buffer to use.

Methods

impl<'a, T> IndicesSource<'a, T> where T: Index

fn get_primitives_type(&self) -> PrimitiveType

Returns the types of primitives.

fn get_indices_type(&self) -> IndexType

Returns the types of indices.

fn get_offset(&self) -> usize

Returns the first element to use from the buffer.

fn get_length(&self) -> usize

Returns the lgnth of the buffer to use.