Struct glium::pixel_buffer::PixelBuffer

pub struct PixelBuffer<T> {
    // some fields omitted
}

Buffer that stores the content of a texture.

The generic type represents the type of pixels that the buffer contains.

Methods

impl<T> PixelBuffer<T>

fn new_empty<F>(facade: &F, capacity: usize) -> PixelBuffer<T> where F: Facade

Builds a new buffer with an uninitialized content.

fn get_size(&self) -> usize

Returns the size of the buffer, in bytes.

impl<T> PixelBuffer<T> where T: Texture2dDataSink

fn read(&self) -> T

Copies the content of the pixel buffer to RAM.

This operation is slow and should be done outside of the rendering loop.

Panic

Panics if the pixel buffer is empty.

Features

This function is only available if the gl_read_buffer feature is enabled. Otherwise, you should use read_if_supported.

fn read_if_supported(&self) -> Option<T>

Copies the content of the pixel buffer to RAM.

This operation is slow and should be done outside of the rendering loop.

Panic

Panics if the pixel buffer is empty.