Enum glium::program::ProgramCreationError

pub enum ProgramCreationError {
    CompilationError(String),
    LinkingError(String),
    ShaderTypeNotSupported,
    CompilationNotSupported,
    TransformFeedbackNotSupported,
    PointSizeNotSupported,
    BinaryHeaderError,
}

Error that can be triggered when creating a Program.

Variants
CompilationError(String)

Error while compiling one of the shaders.

LinkingError(String)

Error while linking the program.

ShaderTypeNotSupported

One of the requested shader types is not supported by the backend.

Usually the case for geometry shaders.

CompilationNotSupported

The OpenGL implementation doesn't provide a compiler.

TransformFeedbackNotSupported

You have requested transform feedback varyings, but transform feedback is not supported by the backend.

PointSizeNotSupported

You have requested point size setting from the shader, but it's not supported by the backend.

BinaryHeaderError

The glium-specific binary header was not found or is corrupt.