Struct glium::program::SubroutineData

pub struct SubroutineData {
    pub location_counts: HashMap<ShaderStage, usize, BuildHasherDefault<FnvHasher>>,
    pub subroutine_uniforms: HashMap<(String, ShaderStage), SubroutineUniform, BuildHasherDefault<FnvHasher>>,
}

Contains all subroutine data of a program.

Fields
location_counts: HashMap<ShaderStage, usize, BuildHasherDefault<FnvHasher>>

Number of subroutine uniform locations per shader stage. This is not equal to the number of subroutine uniforms per stage, because users can use #layout(location=...).

subroutine_uniforms: HashMap<(String, ShaderStage), SubroutineUniform, BuildHasherDefault<FnvHasher>>

The list of all subroutine uniforms of the program stored in a structured way to enable fast lookups. A subroutine uniform is uniquely defined by a name and a shader stage.