Struct glium::program::SubroutineUniform

pub struct SubroutineUniform {
    pub index: u32,
    pub location: i32,
    pub size: Option<usize>,
    pub compatible_subroutines: Vec<Subroutine>,
}

Information about a Subroutine Uniform (except name)

Fields
index: u32

The index of the subroutine uniform. Needed to query information from the OpenGL backend.

location: i32

The location of the uniform. This is used to bind subroutines to this subroutine uniform.

size: Option<usize>

If the uniform is an array, the size of the array.

compatible_subroutines: Vec<Subroutine>

A list of subroutines that can potentially be used with this uniform.