Struct glium::render_buffer::StencilRenderBufferUnstable

pub struct StencilRenderBuffer {
    // some fields omitted
}

A render buffer is similar to a texture, but is optimized for usage as a draw target.

Contrary to a texture, you can't sample or modify the content of the StencilRenderBuffer directly.

Methods

impl StencilRenderBuffer

fn new(display: &Display, format: StencilFormat, width: u32, height: u32) -> StencilRenderBuffer

Builds a new render buffer.

fn get_dimensions(&self) -> (u32, u32)

Returns the dimensions of the render buffer.