Struct glium::BlitTarget

pub struct BlitTarget {
    pub left: u32,
    pub bottom: u32,
    pub width: i32,
    pub height: i32,
}

Area of a surface in pixels. Similar to a Rect except that dimensions can be negative.

In the OpenGL ecosystem, the (0,0) coordinate is at the bottom-left hand corner of the images.

Fields

left

Number of pixels between the left border of the surface and the left border of the rectangle.

bottom

Number of pixels between the bottom border of the surface and the bottom border of the rectangle.

width

Width of the area in pixels. Can be negative.

height

Height of the area in pixels. Can be negative.