Struct glium::Depth

pub struct Depth {
    pub test: DepthTest,
    pub write: bool,
    pub range: (f32, f32),
    pub clamp: DepthClamp,
}

Represents the depth parameters of a draw command.

Fields

test

The function that the GPU will use to determine whether to write over an existing pixel on the target. Don't forget to set depth_write appropriately if you use a depth test.

See the DepthTest documentation for more details.

The default is Overwrite.

write

Sets whether the GPU will write the depth values on the depth buffer if they pass the depth test.

The default is false. You most likely want true if you're doing depth testing.

If you pass true but don't have a depth buffer available, drawing will produce a NoDepthBuffer error.

range

The range of possible Z values in surface coordinates.

Just like OpenGL turns X and Y coordinates between -1.0 and 1.0 into surface coordinates, it will also map your Z coordinates to a certain range which you can specify here.

The two values must be between 0.0 and 1.0, anything outside this range will result in a panic. By default the depth range is (0.0, 1.0).

The first value of the tuple must be the "near" value, where -1.0 will be mapped. The second value must be the "far" value, where 1.0 will be mapped. It is possible for the "near" value to be greater than the "far" value.

clamp

Sets whether the depth values of samples should be clamped to 0.0 and 1.0.

The default value is NoClamp.