Struct glium::IndexBuffer

pub struct IndexBuffer {
    // some fields omitted
}

A list of indices loaded in the graphics card's memory.

Methods

impl IndexBuffer

fn new<T: IntoIndexBuffer, F>(facade: &F, data: T) -> IndexBuffer where F: Facade

Builds a new index buffer.

fn from_raw<T, F>(facade: &F, data: Vec<T>, prim: PrimitiveType) -> IndexBuffer where T: Index, F: Facade

Builds a new index buffer from raw data and a primitive type.

fn get_primitives_type(&self) -> PrimitiveType

Returns the type of primitives associated with this index buffer.

fn get_indices_type(&self) -> IndexType

Returns the data type of the indices inside this index buffer.

fn slice(&self, Range { start: start, end: end }: Range<usize>) -> Option<IndexBufferSlice>

Returns None if out of range.