Struct glium::VertexBuffer

pub struct VertexBuffer<T> {
    // some fields omitted
}

A list of vertices loaded in the graphics card's memory.

Methods

impl<T: Vertex + 'static + Send> VertexBuffer<T>

fn new(display: &Display, data: Vec<T>) -> VertexBuffer<T>

Builds a new vertex buffer.

Example

#[derive(Copy)]
struct Vertex {
    position: [f32; 3],
    texcoords: [f32; 2],
}

implement_vertex!(Vertex, position, texcoords);

let vertex_buffer = glium::VertexBuffer::new(&display, vec![
    Vertex { position: [0.0,  0.0, 0.0], texcoords: [0.0, 1.0] },
    Vertex { position: [5.0, -3.0, 2.0], texcoords: [1.0, 0.0] },
]);

fn new_dynamic(display: &Display, data: Vec<T>) -> VertexBuffer<T>

Builds a new vertex buffer.

This function will create a buffer that has better performance when it is modified frequently.

fn new_persistent(display: &Display, data: Vec<T>) -> VertexBuffer<T>

Builds a new vertex buffer with persistent mapping.

Features

Only available if the gl_persistent_mapping feature is enabled.

fn new_persistent_if_supported(display: &Display, data: Vec<T>) -> Option<VertexBuffer<T>>

Builds a new vertex buffer with persistent mapping, or None if this is not supported.

impl<T: Send + Copy + 'static> VertexBuffer<T>

unsafe fn new_raw(display: &Display, data: Vec<T>, bindings: VertexFormat, elements_size: usize) -> VertexBuffer<T>

Builds a new vertex buffer from an indeterminate data type and bindings.

Example

let bindings = vec![(
        format!("position"), 0,
        glium::vertex::AttributeType::F32F32,
    ), (
        format!("color"), 2 * ::std::mem::size_of::<f32>(),
        glium::vertex::AttributeType::F32,
    ),
];

let data = vec![
    1.0, -0.3, 409.0,
    -0.4, 2.8, 715.0f32
];

let vertex_buffer = unsafe {
    glium::VertexBuffer::new_raw(&display, data, bindings, 3 * ::std::mem::size_of::<f32>())
};

fn slice(&self, offset: usize, len: usize) -> Option<VertexBufferSlice<T>>

Accesses a slice of the buffer.

Returns None if the slice is out of range.

fn map<'a>(&'a mut self) -> Mapping<'a, T>

Maps the buffer to allow write access to it.

This function will block until the buffer stops being used by the backend. This operation is much faster if the buffer is persistent.

fn read(&self) -> Vec<T>

Reads the content of the buffer.

This function is usually better if are just doing one punctual read, while map is better if you want to have multiple small reads.

Features

Only available if the gl_read_buffer feature is enabled.

fn read_if_supported(&self) -> Option<Vec<T>>

Reads the content of the buffer.

This function is usually better if are just doing one punctual read, while map is better if you want to have multiple small reads.

fn write(&self, data: Vec<T>)

Replaces the content of the buffer.

Panic

Panics if the length of data is different from the length of this buffer.

impl<T> VertexBuffer<T>

fn is_persistent(&self) -> bool

Returns true if the buffer is mapped in a permanent way in memory.

fn get_elements_size(&self) -> usize

Returns the number of bytes between two consecutive elements in the buffer.

fn get_bindings(&self) -> &VertexFormat

Returns the associated VertexFormat.

fn into_vertex_buffer_any(self) -> VertexBufferAny

Discard the type information and turn the vertex buffer into a VertexBufferAny.

fn len(&self) -> usize

Returns the number of elements in the buffer.