Struct glium::WaitEventsIterUnstable

pub struct WaitEventsIter<'a> {
    // some fields omitted
}

Blocking iterator over all the events received by the window.

This iterator polls for events, until the window associated with its context is closed.