Struct glium::texture::bindless::ResidentTexture

pub struct ResidentTexture { /* fields omitted */ }

A texture that is resident in video memory. This allows you to use bindless textures in your shaders.

Methods

impl ResidentTexture

pub fn new(
texture: TextureAny
) -> Result<ResidentTexture, BindlessTexturesNotSupportedError>

Takes ownership of the given texture and makes it resident.

pub fn into_inner(self) -> TextureAny

Unwraps the texture and restores it.