Struct glium::texture::buffer_texture::BufferTexture

pub struct BufferTexture<T> where
    [T]: BufferContent,  { /* fields omitted */ }

A one-dimensional texture that gets its data from a buffer.

Methods

impl<T> BufferTexture<T> where
[T]: BufferContent,
T: TextureBufferContent + Copy,

pub fn new<F: ?Sized>(
facade: &F,
data: &[T],
ty: BufferTextureType
) -> Result<BufferTexture<T>, CreationError> where
F: Facade,

Builds a new texture buffer from data.

pub fn dynamic<F: ?Sized>(
facade: &F,
data: &[T],
ty: BufferTextureType
) -> Result<BufferTexture<T>, CreationError> where
F: Facade,

Builds a new texture buffer from data.

pub fn persistent<F: ?Sized>(
facade: &F,
data: &[T],
ty: BufferTextureType
) -> Result<BufferTexture<T>, CreationError> where
F: Facade,

Builds a new texture buffer from data.

pub fn immutable<F: ?Sized>(
facade: &F,
data: &[T],
ty: BufferTextureType
) -> Result<BufferTexture<T>, CreationError> where
F: Facade,

Builds a new texture buffer from data.

pub fn empty<F: ?Sized>(
facade: &F,
len: usize,
ty: BufferTextureType
) -> Result<BufferTexture<T>, CreationError> where
F: Facade,

Builds a new empty buffer buffer.

pub fn empty_dynamic<F: ?Sized>(
facade: &F,
len: usize,
ty: BufferTextureType
) -> Result<BufferTexture<T>, CreationError> where
F: Facade,

Builds a new empty buffer buffer.

pub fn empty_persistent<F: ?Sized>(
facade: &F,
len: usize,
ty: BufferTextureType
) -> Result<BufferTexture<T>, CreationError> where
F: Facade,

Builds a new empty buffer buffer.

pub fn empty_immutable<F: ?Sized>(
facade: &F,
len: usize,
ty: BufferTextureType
) -> Result<BufferTexture<T>, CreationError> where
F: Facade,

Builds a new empty buffer buffer.

pub fn from_buffer<F: ?Sized>(
context: &F,
buffer: Buffer<[T]>,
ty: BufferTextureType
) -> Result<BufferTexture<T>, (TextureCreationError, Buffer<[T]>)> where
F: Facade,

Builds a new buffer texture by taking ownership of a buffer.

impl<T> BufferTexture<T> where
[T]: BufferContent,

pub fn as_buffer_texture_ref(&self) -> BufferTextureRef

Builds a BufferTextureRef.