Struct glium::texture::compressed_cubemap::CompressedCubemapMipmap

pub struct CompressedCubemapMipmap<'t>(_, _);

Represents a single mipmap level of a CompressedCubemap.

Can be obtained by calling CompressedCubemap::mipmap(), CompressedCubemap::main_level(), CompressedCubemapLayer::mipmap() or CompressedCubemapLayer::main_level().

Methods

impl<'t> CompressedCubemapMipmap<'t>

pub fn width(&self) -> u32

Returns the width of that image.

pub fn height(&self) -> u32

Returns the height of that image.

pub fn dimensions(&self) -> u32

Returns the dimension of that image.

pub fn read_compressed_data(&self) -> Option<(CompressedFormat, Vec<u8>)>

Reads the content of the texture level to RAM without decompressing it before.

You should avoid doing this at all cost during performance-critical operations (for example, while you're drawing).

Returns the compressed format of the texture and the compressed data, gives None when the internal compression format is generic or unknown.

pub fn get_texture(&self) -> &'t CompressedCubemap

Returns the corresponding texture.

pub fn get_level(&self) -> u32

Returns the texture level.

pub fn image(&self, layer: CubeLayer) -> CompressedCubemapImage<'t>

Provides an object representing a single layer of this cubemap.