Struct glium::texture::compressed_cubemap_array::CompressedCubemapArrayLayerMipmap

pub struct CompressedCubemapArrayLayerMipmap<'t>(_, _);

Represents a single layer of a mipmap level of a CompressedCubemapArray.

Methods

impl<'t> CompressedCubemapArrayLayerMipmap<'t>

fn image(&self, layer: CubeLayer) -> CompressedCubemapArrayImage<'t>

Provides an object representing a single layer of this cubemap.