Struct glium::texture::compressed_srgb_texture2d::CompressedSrgbTexture2d

pub struct CompressedSrgbTexture2d(_);

A compressed two-dimensional texture containing sRGB floating-point data.

Methods

impl CompressedSrgbTexture2d

fn sampled(&self) -> Sampler<CompressedSrgbTexture2d>

Builds a Sampler marker object that allows you to indicate how the texture should be sampled from inside a shader.

Example

let uniforms = uniform! {
    color_texture: texture.sampled().magnify_filter(glium::uniforms::MagnifySamplerFilter::Nearest)
};

impl CompressedSrgbTexture2d

fn get_internal_format(&self) -> Result<InternalFormat, GetFormatError>

Determines the internal format of this texture.

The backend may not support querying the actual format, in which case an error is returned.

fn new<'a, F, T>(facade: &F, data: T) -> Result<CompressedSrgbTexture2d, TextureCreationError> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

No mipmap level (except for the main level) will be allocator nor generated.

fn with_mipmaps<'a, F, T>(facade: &F, data: T, mipmaps: CompressedMipmapsOption) -> Result<CompressedSrgbTexture2d, TextureCreationError> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

fn with_compressed_data<F>(facade: &F, data: &[u8], width: u32, height: u32, format: CompressedSrgbFormat, mipmaps: CompressedMipmapsOption) -> Result<CompressedSrgbTexture2d, TextureCreationError> where F: Facade

Builds a new texture with a specific format. The input data must also be of the specified compressed format.

fn with_format<'a, F, T>(facade: &F, data: T, format: CompressedSrgbFormat, mipmaps: CompressedMipmapsOption) -> Result<CompressedSrgbTexture2d, TextureCreationError> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture with a specific format.

fn empty_with_format<F>(facade: &F, format: CompressedSrgbFormat, mipmaps: CompressedMipmapsOption, width: u32, height: u32) -> Result<CompressedSrgbTexture2d, TextureCreationError> where F: Facade

Creates an empty texture with a specific format.

The texture (and its mipmaps) will contain undefined data.

unsafe fn from_id<F: Facade>(facade: &F, format: CompressedSrgbFormat, id: GLuint, owned: bool, mipmap: MipmapsOption, ty: Dimensions) -> CompressedSrgbTexture2d

Builds a new texture reference from an existing, externally created OpenGL texture. If owned is true, this reference will take ownership of the texture and be responsible for cleaning it up. Otherwise, the texture must be cleaned up externally, but only after this reference's lifetime has ended.

fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

The minimum value is 1, since there is always a main texture.

fn read<T>(&self) -> T where T: Texture2dDataSink<(u8, u8, u8, u8)>

Reads the content of the texture to RAM.

You should avoid doing this at all cost during performance-critical operations (for example, while you're drawing). Use read_to_pixel_buffer instead.

fn read_to_pixel_buffer(&self) -> PixelBuffer<(u8, u8, u8, u8)>

Reads the content of the texture into a buffer in video memory.

This operation copies the texture's data into a buffer in video memory (a pixel buffer). Contrary to the read function, this operation is done asynchronously and doesn't need a synchronization.

fn read_compressed_data(&self) -> Option<(CompressedSrgbFormat, Vec<u8>)>

Reads the content of the texture to RAM without decompressing it before.

You should avoid doing this at all cost during performance-critical operations (for example, while you're drawing).

Returns the compressed format of the texture and the compressed data, gives None when the internal compression format is generic or unknown.

fn write<'a, T>(&self, rect: Rect, data: T) where T: Texture2dDataSource<'a>

Uploads some data in the texture.

Note that this may cause a synchronization if you use the texture right before or right after this call. Prefer creating a whole new texture if you change a huge part of it.

Panic

Panics if the the dimensions of data don't match the Rect.

Calling this for compressed textures will result in a panic of type INVALID_OPERATION if Rect::bottom or Rect::width is not equal to 0 (border). In addition, the contents of any texel outside the region modified by such a call are undefined. These restrictions may be relaxed for specific compressed internal formats whose images are easily edited.

fn write_compressed_data(&self, rect: Rect, data: &[u8], width: u32, height: u32, format: CompressedSrgbFormat) -> Result<(), ()>

Uploads some data in the texture by using a compressed format as input.

Note that this may cause a synchronization if you use the texture right before or right after this call. Prefer creating a whole new texture if you change a huge part of it.

Panic

Panics if the the dimensions of data don't match the Rect.

Calling this will result in a panic of type INVALID_OPERATION error if Rect::width or Rect::height is not equal to 0 (border), or if the written dimensions do not match the original texture dimensions. The contents of any texel outside the region modified by the call are undefined. These restrictions may be relaxed for specific compressed internal formats whose images are easily edited.

fn resident(self) -> Result<ResidentTexture, BindlessTexturesNotSupportedError>

Turns the texture into a ResidentTexture.

This allows you to use the texture in a much more efficient way by storing a "reference to it" in a buffer (actually not a reference but a raw pointer).

See the documentation of ResidentTexture for more infos.

fn mipmap(&self, level: u32) -> Option<CompressedSrgbTexture2dMipmap>

Access a single mipmap level of this texture.

fn main_level(&self) -> CompressedSrgbTexture2dMipmap

Access the main mipmap level of this texture.