Struct glium::texture::compressed_srgb_texture2d_array::CompressedSrgbTexture2dArray

pub struct CompressedSrgbTexture2dArray(_);

An array of compressed two-dimensional textures containing sRGB floating-point data.

Implementations

impl CompressedSrgbTexture2dArray

pub fn sampled(&self) -> Sampler<'_, CompressedSrgbTexture2dArray>

Builds a Sampler marker object that allows you to indicate how the texture should be sampled from inside a shader.

Example

let uniforms = uniform! {
    color_texture: texture.sampled().magnify_filter(glium::uniforms::MagnifySamplerFilter::Nearest)
};

impl CompressedSrgbTexture2dArray

pub fn get_internal_format(&self) -> Result<InternalFormat, GetFormatError>

Determines the internal format of this texture.

The backend may not support querying the actual format, in which case an error is returned.

pub fn new<'a, F: ?Sized, T>(
facade: &F,
data: Vec<T>
) -> Result<CompressedSrgbTexture2dArray, TextureCreationError> where
T: Texture2dDataSource<'a>,
F: Facade,

Builds a new texture by uploading data.

No mipmap level (except for the main level) will be allocator nor generated.

pub fn with_mipmaps<'a, F: ?Sized, T>(
facade: &F,
data: Vec<T>,
mipmaps: CompressedMipmapsOption
) -> Result<CompressedSrgbTexture2dArray, TextureCreationError> where
T: Texture2dDataSource<'a>,
F: Facade,

Builds a new texture by uploading data.

pub fn with_compressed_data<F: ?Sized>(
facade: &F,
data: Vec<&[u8]>,
width: u32,
height: u32,
array_size: u32,
format: CompressedSrgbFormat,
mipmaps: CompressedMipmapsOption
) -> Result<CompressedSrgbTexture2dArray, TextureCreationError> where
F: Facade,

Builds a new texture with a specific format. The input data must also be of the specified compressed format.

pub fn with_format<'a, F: ?Sized, T>(
facade: &F,
data: Vec<T>,
format: CompressedSrgbFormat,
mipmaps: CompressedMipmapsOption
) -> Result<CompressedSrgbTexture2dArray, TextureCreationError> where
T: Texture2dDataSource<'a>,
F: Facade,

Builds a new texture with a specific format.

pub fn empty_with_format<F: ?Sized>(
facade: &F,
format: CompressedSrgbFormat,
mipmaps: CompressedMipmapsOption,
width: u32,
height: u32,
array_size: u32
) -> Result<CompressedSrgbTexture2dArray, TextureCreationError> where
F: Facade,

Creates an empty texture with a specific format.

The texture (and its mipmaps) will contain undefined data.

pub unsafe fn from_id<F: Facade + ?Sized>(
facade: &F,
format: CompressedSrgbFormat,
id: c_uint,
owned: bool,
mipmap: MipmapsOption,
ty: Dimensions
) -> CompressedSrgbTexture2dArray

Builds a new texture reference from an existing, externally created OpenGL texture. If owned is true, this reference will take ownership of the texture and be responsible for cleaning it up. Otherwise, the texture must be cleaned up externally, but only after this reference's lifetime has ended.

pub fn width(&self) -> u32

Returns the width of that image.

pub fn height(&self) -> u32

Returns the height of that image.

pub fn array_size(&self) -> u32

Returns the number of array layers.

pub fn dimensions(&self) -> (u32, u32)

Returns the width and height of that image.

pub fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

The minimum value is 1, since there is always a main texture.

pub fn resident(
self
) -> Result<ResidentTexture, BindlessTexturesNotSupportedError>

Turns the texture into a ResidentTexture.

This allows you to use the texture in a much more efficient way by storing a "reference to it" in a buffer (actually not a reference but a raw pointer).

See the documentation of ResidentTexture for more infos.

pub fn first_layer(&self) -> CompressedSrgbTexture2dArrayLayer<'_>

Access the first layer of this texture.

pub fn layer(&self, layer: u32) -> Option<CompressedSrgbTexture2dArrayLayer<'_>>

Access a single layer of this texture.

pub fn mipmap(
&self,
level: u32
) -> Option<CompressedSrgbTexture2dArrayMipmap<'_>>

Access a single mipmap level of this texture.

pub fn main_level(&self) -> CompressedSrgbTexture2dArrayMipmap<'_>

Access the main mipmap level of this texture.