Module glium::texture::depth_stencil_cubemap

Contains the implementation of DepthStencilCubemap.

Structs

DepthStencilCubemap
A cube texture containing both depth and stencil data.

DepthStencilCubemapImage
Represents a single image of a mipmap level of a layer of DepthStencilCubemap.

DepthStencilCubemapLayer
Represents a single layer of a DepthStencilCubemap.

DepthStencilCubemapLayerMipmap
Represents a single layer of a mipmap level of a DepthStencilCubemap.

DepthStencilCubemapMipmap
Represents a single mipmap level of a DepthStencilCubemap.