Enum glium::texture::ClientFormat

pub enum ClientFormat {
    U8,
    U8U8,
    U8U8U8,
    U8U8U8U8,
    I8,
    I8I8,
    I8I8I8,
    I8I8I8I8,
    U16,
    U16U16,
    U16U16U16,
    U16U16U16U16,
    I16,
    I16I16,
    I16I16I16,
    I16I16I16I16,
    U32,
    U32U32,
    U32U32U32,
    U32U32U32U32,
    I32,
    I32I32,
    I32I32I32,
    I32I32I32I32,
    U3U3U2,
    U5U6U5,
    U4U4U4U4,
    U5U5U5U1,
    U10U10U10U2,
    F16,
    F16F16,
    F16F16F16,
    F16F16F16F16,
    F32,
    F32F32,
    F32F32F32,
    F32F32F32F32,
}

List of client-side pixel formats.

These are all the possible formats of input data when uploading to a texture.

Variants
U8U8U8U8U8U8U8U8U8U8I8I8I8I8I8I8I8I8I8I8U16U16U16U16U16U16U16U16U16U16I16I16I16I16I16I16I16I16I16I16U32U32U32U32U32U32U32U32U32U32I32I32I32I32I32I32I32I32I32I32U3U3U2U5U6U5U4U4U4U4U5U5U5U1U10U10U10U2F16F16F16F16F16F16F16F16F16F16F32F32F32F32F32F32F32F32F32F32
Methods

impl ClientFormat

pub fn get_size(&self) -> usize

Returns the size in bytes of a pixel of this type.

pub fn get_num_components(&self) -> u8

Returns the number of components of this client format.