Enum glium::texture::InternalFormat

pub enum InternalFormat {
    OneComponent {
        ty1: InternalFormatType,
        bits1: usize,
    },
    TwoComponents {
        ty1: InternalFormatType,
        bits1: usize,
        ty2: InternalFormatType,
        bits2: usize,
    },
    ThreeComponents {
        ty1: InternalFormatType,
        bits1: usize,
        ty2: InternalFormatType,
        bits2: usize,
        ty3: InternalFormatType,
        bits3: usize,
    },
    FourComponents {
        ty1: InternalFormatType,
        bits1: usize,
        ty2: InternalFormatType,
        bits2: usize,
        ty3: InternalFormatType,
        bits3: usize,
        ty4: InternalFormatType,
        bits4: usize,
    },
}

Internal format of a texture.

The actual format of a texture is not necessarly one of the predefined ones, so we have to use a very generic description.

Variants

OneComponent

The format has one component.

Fields

ty1

Type of the first component of the format.

bits1

Number of bits of the first component.

TwoComponents

The format has two components.

Fields

ty1

Type of the first component of the format.

bits1

Number of bits of the first component.

ty2

Type of the second component.

bits2

Number of bits of the second component.

ThreeComponents

The format has three components.

Fields

ty1

Type of the first component of the format.

bits1

Number of bits of the first component.

ty2

Type of the second component.

bits2

Number of bits of the second component.

ty3

Type of the third component.

bits3

Number of bits of the third component.

FourComponents

The format has four components.

Fields

ty1

Type of the first component of the format.

bits1

Number of bits of the first component.

ty2

Type of the second component.

bits2

Number of bits of the second component.

ty3

Type of the third component.

bits3

Number of bits of the third component.

ty4

Type of the fourth component.

bits4

Number of bits of the fourth component.

Methods

impl InternalFormat

fn get_total_bits(&self) -> usize

Returns the total number of bits of this format.