Enum glium::texture::SrgbFormat

pub enum SrgbFormat {
    U8U8U8,
    U8U8U8U8,
}

List of uncompressed pixel formats that contain floating-point data in the sRGB color space.

Variants
U8U8U8U8U8U8U8
Methods

impl SrgbFormat

pub fn get_formats_list() -> Vec<SrgbFormat>

Returns a list of all the possible values of this enumeration.

pub fn to_texture_format(self) -> TextureFormat

Turns this format into a more generic TextureFormat.

pub fn is_supported<C: ?Sized>(&self, context: &C) -> bool where
C: CapabilitiesSource,

Returns true if this format is supported by the backend.

pub fn is_color_renderable<C: ?Sized>(&self, context: &C) -> bool where
C: CapabilitiesSource,

Returns true if a texture or renderbuffer with this format can be used as a framebuffer attachment.