Module glium::texture::integral_cubemap_array

Contains the implementation of IntegralCubemapArray.

Structs

IntegralCubemapArray
An array of cube textures containing signed integral data.

IntegralCubemapArrayImage
Represents a single image of a mipmap level of a layer of IntegralCubemapArray.

IntegralCubemapArrayLayer
Represents a single layer of a IntegralCubemapArray.

IntegralCubemapArrayLayerMipmap
Represents a single layer of a mipmap level of a IntegralCubemapArray.

IntegralCubemapArrayMipmap
Represents a single mipmap level of a IntegralCubemapArray.