Struct glium::texture::integral_cubemap_array::IntegralCubemapArray

pub struct IntegralCubemapArray(_);

An array of cube textures containing signed integral data.

Methods

impl IntegralCubemapArray

pub fn sampled(&self) -> Sampler<IntegralCubemapArray>

Builds a Sampler marker object that allows you to indicate how the texture should be sampled from inside a shader.

Example

let uniforms = uniform! {
    color_texture: texture.sampled().magnify_filter(glium::uniforms::MagnifySamplerFilter::Nearest)
};

impl IntegralCubemapArray

pub fn get_internal_format(&self) -> Result<InternalFormat, GetFormatError>

Determines the internal format of this texture.

The backend may not support querying the actual format, in which case an error is returned.

pub fn empty<F: ?Sized>(
facade: &F,
dimension: u32,
array_size: u32
) -> Result<IntegralCubemapArray, TextureCreationError> where
F: Facade,

Creates an empty texture.

No mipmap level (except for the main level) will be allocated or generated.

The texture will contain undefined data.

pub fn empty_with_format<F: ?Sized>(
facade: &F,
format: UncompressedIntFormat,
mipmaps: MipmapsOption,
dimension: u32,
array_size: u32
) -> Result<IntegralCubemapArray, TextureCreationError> where
F: Facade,

Creates an empty texture with a specific format.

The texture (and its mipmaps) will contain undefined data.

pub fn empty_with_mipmaps<F: ?Sized>(
facade: &F,
mipmaps: MipmapsOption,
dimension: u32,
array_size: u32
) -> Result<IntegralCubemapArray, TextureCreationError> where
F: Facade,

Creates an empty texture. Specifies whether is has mipmaps.

The texture (and its mipmaps) will contain undefined data.

pub unsafe fn from_id<F: Facade + ?Sized>(
facade: &F,
format: UncompressedIntFormat,
id: c_uint,
owned: bool,
mipmap: MipmapsOption,
ty: Dimensions
) -> IntegralCubemapArray

Builds a new texture reference from an existing, externally created OpenGL texture. If owned is true, this reference will take ownership of the texture and be responsible for cleaning it up. Otherwise, the texture must be cleaned up externally, but only after this reference's lifetime has ended.

pub fn width(&self) -> u32

Returns the width of that image.

pub fn height(&self) -> u32

Returns the height of that image.

pub fn array_size(&self) -> u32

Returns the number of array layers.

pub fn dimensions(&self) -> u32

Returns the dimension of that image.

pub fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

The minimum value is 1, since there is always a main texture.

pub fn resident(
self
) -> Result<ResidentTexture, BindlessTexturesNotSupportedError>

Turns the texture into a ResidentTexture.

This allows you to use the texture in a much more efficient way by storing a "reference to it" in a buffer (actually not a reference but a raw pointer).

See the documentation of ResidentTexture for more infos.

pub fn first_layer(&self) -> IntegralCubemapArrayLayer

Access the first layer of this texture.

pub fn layer(&self, layer: u32) -> Option<IntegralCubemapArrayLayer>

Access a single layer of this texture.

pub fn mipmap(&self, level: u32) -> Option<IntegralCubemapArrayMipmap>

Access a single mipmap level of this texture.

pub fn main_level(&self) -> IntegralCubemapArrayMipmap

Access the main mipmap level of this texture.