Struct glium::texture::integral_texture2d::IntegralTexture2d

pub struct IntegralTexture2d(_);

A two-dimensional texture containing signed integral data.

Methods

impl IntegralTexture2d

fn sampled(&self) -> Sampler<IntegralTexture2d>

Builds a Sampler marker object that allows you to indicate how the texture should be sampled from inside a shader.

Example

let uniforms = uniform! {
    color_texture: texture.sampled().magnify_filter(glium::uniforms::MagnifySamplerFilter::Nearest)
};

impl IntegralTexture2d

fn get_internal_format(&self) -> Result<InternalFormat, GetFormatError>

Determines the internal format of this texture.

The backend may not support querying the actual format, in which case an error is returned.

fn new<'a, F, T>(facade: &F, data: T) -> Result<IntegralTexture2d, TextureCreationError> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

This function will automatically generate all mipmaps of the texture.

fn with_mipmaps<'a, F, T>(facade: &F, data: T, mipmaps: MipmapsOption) -> Result<IntegralTexture2d, TextureCreationError> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

fn with_format<'a, F, T>(facade: &F, data: T, format: UncompressedIntFormat, mipmaps: MipmapsOption) -> Result<IntegralTexture2d, TextureCreationError> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture with a specific format.

fn empty<F>(facade: &F, width: u32, height: u32) -> Result<IntegralTexture2d, TextureCreationError> where F: Facade

Creates an empty texture.

No mipmap level (except for the main level) will be allocated or generated.

The texture will contain undefined data.

fn empty_with_format<F>(facade: &F, format: UncompressedIntFormat, mipmaps: MipmapsOption, width: u32, height: u32) -> Result<IntegralTexture2d, TextureCreationError> where F: Facade

Creates an empty texture with a specific format.

The texture (and its mipmaps) will contain undefined data.

fn empty_with_mipmaps<F>(facade: &F, mipmaps: MipmapsOption, width: u32, height: u32) -> Result<IntegralTexture2d, TextureCreationError> where F: Facade

Creates an empty texture. Specifies whether is has mipmaps.

The texture (and its mipmaps) will contain undefined data.

unsafe fn from_id<F: Facade>(facade: &F, format: UncompressedIntFormat, id: GLuint, owned: bool, mipmap: MipmapsOption, ty: Dimensions) -> IntegralTexture2d

Builds a new texture reference from an existing, externally created OpenGL texture. If owned is true, this reference will take ownership of the texture and be responsible for cleaning it up. Otherwise, the texture must be cleaned up externally, but only after this reference's lifetime has ended.

fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

The minimum value is 1, since there is always a main texture.

fn resident(self) -> Result<ResidentTexture, BindlessTexturesNotSupportedError>

Turns the texture into a ResidentTexture.

This allows you to use the texture in a much more efficient way by storing a "reference to it" in a buffer (actually not a reference but a raw pointer).

See the documentation of ResidentTexture for more infos.

fn mipmap(&self, level: u32) -> Option<IntegralTexture2dMipmap>

Access a single mipmap level of this texture.

fn main_level(&self) -> IntegralTexture2dMipmap

Access the main mipmap level of this texture.