Module glium::texture::srgb_cubemap_array

Contains the implementation of SrgbCubemapArray.

Structs

SrgbCubemapArray
An array of cube textures containing sRGB floating-point data.

SrgbCubemapArrayImage
Represents a single image of a mipmap level of a layer of SrgbCubemapArray.

SrgbCubemapArrayLayer
Represents a single layer of a SrgbCubemapArray.

SrgbCubemapArrayLayerMipmap
Represents a single layer of a mipmap level of a SrgbCubemapArray.

SrgbCubemapArrayMipmap
Represents a single mipmap level of a SrgbCubemapArray.