Struct glium::texture::stencil_cubemap_array::StencilCubemapArrayLayerMipmap

pub struct StencilCubemapArrayLayerMipmap<'t>(_, _);

Represents a single layer of a mipmap level of a StencilCubemapArray.

Methods

impl<'t> StencilCubemapArrayLayerMipmap<'t>

pub fn width(&self) -> u32

Returns the width of that image.

pub fn height(&self) -> u32

Returns the height of that image.

pub fn dimensions(&self) -> u32

Returns the dimension of that image.

pub fn image(&self, layer: CubeLayer) -> StencilCubemapArrayImage<'t>

Provides an object representing a single layer of this cubemap.