Struct glium::texture::stencil_texture1d::StencilTexture1d

pub struct StencilTexture1d(_);

A one-dimensional texture containing stencil data.

Methods

impl StencilTexture1d

pub fn get_internal_format(&self) -> Result<InternalFormat, GetFormatError>

Determines the internal format of this texture.

The backend may not support querying the actual format, in which case an error is returned.

pub fn new<'a, F: ?Sized, T>(
facade: &F,
data: T
) -> Result<StencilTexture1d, TextureCreationError> where
T: Texture1dDataSource<'a>,
F: Facade,

Builds a new texture by uploading data.

This function will automatically generate all mipmaps of the texture.

pub fn with_mipmaps<'a, F: ?Sized, T>(
facade: &F,
data: T,
mipmaps: MipmapsOption
) -> Result<StencilTexture1d, TextureCreationError> where
T: Texture1dDataSource<'a>,
F: Facade,

Builds a new texture by uploading data.

pub fn with_format<'a, F: ?Sized, T>(
facade: &F,
data: T,
format: StencilFormat,
mipmaps: MipmapsOption
) -> Result<StencilTexture1d, TextureCreationError> where
T: Texture1dDataSource<'a>,
F: Facade,

Builds a new texture with a specific format.

pub fn empty<F: ?Sized>(
facade: &F,
width: u32
) -> Result<StencilTexture1d, TextureCreationError> where
F: Facade,

Creates an empty texture.

No mipmap level (except for the main level) will be allocated or generated.

The texture will contain undefined data.

pub fn empty_with_format<F: ?Sized>(
facade: &F,
format: StencilFormat,
mipmaps: MipmapsOption,
width: u32
) -> Result<StencilTexture1d, TextureCreationError> where
F: Facade,

Creates an empty texture with a specific format.

The texture (and its mipmaps) will contain undefined data.

pub fn empty_with_mipmaps<F: ?Sized>(
facade: &F,
mipmaps: MipmapsOption,
width: u32
) -> Result<StencilTexture1d, TextureCreationError> where
F: Facade,

Creates an empty texture. Specifies whether is has mipmaps.

The texture (and its mipmaps) will contain undefined data.

pub unsafe fn from_id<F: Facade + ?Sized>(
facade: &F,
format: StencilFormat,
id: c_uint,
owned: bool,
mipmap: MipmapsOption,
ty: Dimensions
) -> StencilTexture1d

Builds a new texture reference from an existing, externally created OpenGL texture. If owned is true, this reference will take ownership of the texture and be responsible for cleaning it up. Otherwise, the texture must be cleaned up externally, but only after this reference's lifetime has ended.

pub fn width(&self) -> u32

Returns the width of that image.

pub fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

The minimum value is 1, since there is always a main texture.

pub fn resident(
self
) -> Result<ResidentTexture, BindlessTexturesNotSupportedError>

Turns the texture into a ResidentTexture.

This allows you to use the texture in a much more efficient way by storing a "reference to it" in a buffer (actually not a reference but a raw pointer).

See the documentation of ResidentTexture for more infos.

pub fn mipmap(&self, level: u32) -> Option<StencilTexture1dMipmap>

Access a single mipmap level of this texture.

pub fn main_level(&self) -> StencilTexture1dMipmap

Access the main mipmap level of this texture.