Struct glium::texture::CompressedSrgbTexture2d

pub struct CompressedSrgbTexture2d(_);

A two-dimensional texture containing sRGB floating-point data.

Methods

impl CompressedSrgbTexture2d

fn sampled(&self) -> Sampler<CompressedSrgbTexture2d>

Builds a Sampler marker object that allows you to indicate how the texture should be sampled from inside a shader.

Example

let uniforms = uniform! {
    color_texture: texture.sampled().magnify_filter(glium::uniforms::MagnifySamplerFilter::Nearest)
};

impl CompressedSrgbTexture2d

fn get_internal_format_if_supported(&self) -> Option<InternalFormat>

Determines the internal format of this texture.

Returns None if the backend doesn't allow querying the actual format.

fn new<'a, F, T>(facade: &F, data: T) -> CompressedSrgbTexture2d where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

This function will automatically generate all mipmaps of the texture.

fn with_mipmaps<'a, F, T>(facade: &F, data: T, mipmaps: bool) -> CompressedSrgbTexture2d where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn with_format<'a, F, T>(facade: &F, data: T, format: CompressedSrgbFormat, mipmaps: bool) -> Result<CompressedSrgbTexture2d, TextureCreationError> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

The minimum value is 1, since there is always a main texture.

fn mipmap(&self, level: u32) -> Option<CompressedSrgbTexture2dMipmap>

Access a single mipmap level of this texture.

fn main_level(&self) -> CompressedSrgbTexture2dMipmap

Access the main mipmap level of this texture.