Struct glium::texture::CompressedTexture2dArray

pub struct CompressedTexture2dArray(_);

An array of compressed two-dimensional textures containing floating-point data.

Methods

impl CompressedTexture2dArray

fn new<'a, F, T>(facade: &F, data: Vec<T>) -> CompressedTexture2dArray where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

This function will automatically generate all mipmaps of the texture.

fn new_if_supported<'a, F, T>(facade: &F, data: Vec<T>) -> Option<CompressedTexture2dArray> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

This function will automatically generate all mipmaps of the texture.

fn with_mipmaps<'a, F, T>(facade: &F, data: Vec<T>, mipmaps: bool) -> CompressedTexture2dArray where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn with_mipmaps_if_supported<'a, F, T>(facade: &F, data: Vec<T>, mipmaps: bool) -> Option<CompressedTexture2dArray> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn with_format<'a, F, T>(facade: &F, data: Vec<T>, format: CompressedFormat, mipmaps: bool) -> Result<CompressedTexture2dArray, TextureCreationError> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn with_format_if_supported<'a, F, T>(facade: &F, data: Vec<T>, format: CompressedFormat, mipmaps: bool) -> Result<CompressedTexture2dArray, TextureMaybeSupportedCreationError> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

The minimum value is 1, since there is always a main texture.

fn layer(&self, layer: u32) -> Option<CompressedTexture2dArrayLayer>

Access a single layer of this texture.