Struct glium::texture::CompressedTexture3dMipmap

pub struct CompressedTexture3dMipmap<'t>(_, _);

Represents a single mipmap level of a CompressedTexture3d.

Can be obtained by calling CompressedTexture3d::mipmap() or CompressedTexture3d::main_level().

Methods

impl<'t> CompressedTexture3dMipmap<'t>

fn read_compressed_data(&self) -> Option<(CompressedFormat, Vec<u8>)>

Reads the content of the texture level to RAM without decompressing it before.

You should avoid doing this at all cost during performance-critical operations (for example, while you're drawing).

Returns the compressed format of the texture and the compressed data, gives None when the internal compression format is generic or unknown.

fn get_texture(&self) -> &'t CompressedTexture3d

Returns the corresponding texture.

fn get_level(&self) -> u32

Returns the texture level.