Struct glium::texture::DepthStencilTexture1dArray

pub struct DepthStencilTexture1dArray(_);

An array of one-dimensional textures containing both depth and stencil data.

Methods

impl DepthStencilTexture1dArray

fn get_internal_format(&self) -> Result<InternalFormat, GetFormatError>

Determines the internal format of this texture.

The backend may not support querying the actual format, in which case an error is returned.

fn new<'a, F, T>(facade: &F, data: Vec<T>) -> Result<DepthStencilTexture1dArray, TextureCreationError> where T: Texture1dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

This function will automatically generate all mipmaps of the texture.

fn with_mipmaps<'a, F, T>(facade: &F, data: Vec<T>, mipmaps: MipmapsOption) -> Result<DepthStencilTexture1dArray, TextureCreationError> where T: Texture1dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

fn with_format<'a, F, T>(facade: &F, data: Vec<T>, format: DepthStencilFormat, mipmaps: MipmapsOption) -> Result<DepthStencilTexture1dArray, TextureCreationError> where T: Texture1dDataSource<'a>, F: Facade

Builds a new texture with a specific format.

fn empty<F>(facade: &F, width: u32, array_size: u32) -> Result<DepthStencilTexture1dArray, TextureCreationError> where F: Facade

Creates an empty texture.

No mipmap level (except for the main level) will be allocated or generated.

The texture will contain undefined data.

fn empty_with_format<F>(facade: &F, format: DepthStencilFormat, mipmaps: MipmapsOption, width: u32, array_size: u32) -> Result<DepthStencilTexture1dArray, TextureCreationError> where F: Facade

Creates an empty texture with a specific format.

The texture (and its mipmaps) will contain undefined data.

fn empty_with_mipmaps<F>(facade: &F, mipmaps: MipmapsOption, width: u32, array_size: u32) -> Result<DepthStencilTexture1dArray, TextureCreationError> where F: Facade

Creates an empty texture. Specifies whether is has mipmaps.

The texture (and its mipmaps) will contain undefined data.

fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

The minimum value is 1, since there is always a main texture.

fn resident(self) -> Result<ResidentTexture, BindlessTexturesNotSupportedError>

Turns the texture into a ResidentTexture.

This allows you to use the texture in a much more efficient way by storing a "reference to it" in a buffer (actually not a reference but a raw pointer).

See the documentation of ResidentTexture for more infos.

fn layer(&self, layer: u32) -> Option<DepthStencilTexture1dArrayLayer>

Access a single layer of this texture.