Struct glium::texture::DepthTexture1d

pub struct DepthTexture1d(_);

A one-dimensional texture containing depth data.

Methods

impl DepthTexture1d

fn new<'a, T>(display: &Display, data: T) -> DepthTexture1d where T: Texture1dDataSource<'a>

Builds a new texture by uploading data.

This function will automatically generate all mipmaps of the texture.

fn new_if_supported<'a, T>(display: &Display, data: T) -> Option<DepthTexture1d> where T: Texture1dDataSource<'a>

Builds a new texture by uploading data.

This function will automatically generate all mipmaps of the texture.

fn with_mipmaps<'a, T>(display: &Display, data: T, mipmaps: bool) -> DepthTexture1d where T: Texture1dDataSource<'a>

Builds a new texture by uploading data.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn with_mipmaps_if_supported<'a, T>(display: &Display, data: T, mipmaps: bool) -> Option<DepthTexture1d> where T: Texture1dDataSource<'a>

Builds a new texture by uploading data.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn with_format<'a, T>(display: &Display, data: T, format: DepthFormat, mipmaps: bool) -> Result<DepthTexture1d, TextureCreationError> where T: Texture1dDataSource<'a>

Builds a new texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn with_format_if_supported<'a, T>(display: &Display, data: T, format: DepthFormat, mipmaps: bool) -> Result<DepthTexture1d, TextureMaybeSupportedCreationError> where T: Texture1dDataSource<'a>

Builds a new texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn new_empty(display: &Display, format: DepthFormat, width: u32) -> DepthTexture1d

Creates an empty texture.

The texture will contain undefined data.

fn empty(display: &Display, width: u32) -> DepthTexture1d

Creates an empty texture.

No mipmaps will be created.

The texture will contain undefined data.

fn empty_if_supported(display: &Display, width: u32) -> Option<DepthTexture1d>

Creates an empty texture.

No mipmaps will be created.

The texture will contain undefined data.

fn empty_with_format(display: &Display, format: DepthFormat, mipmaps: bool, width: u32) -> Result<DepthTexture1d, TextureCreationError>

Creates an empty texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

The texture (and its mipmaps, if you pass true) will contain undefined data.

fn empty_with_format_if_supported(display: &Display, format: DepthFormat, mipmaps: bool, width: u32) -> Result<DepthTexture1d, TextureMaybeSupportedCreationError>

Creates an empty texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

The texture (and its mipmaps, if you pass true) will contain undefined data.

fn empty_with_mipmaps(display: &Display, mipmaps: bool, width: u32) -> DepthTexture1d

Creates an empty texture. Specifies whether is has mipmaps.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

The texture (and its mipmaps, if you pass true) will contain undefined data.

fn empty_with_mipmaps_if_supported(display: &Display, mipmaps: bool, width: u32) -> Option<DepthTexture1d>

Creates an empty texture. Specifies whether is has mipmaps.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

The texture (and its mipmaps, if you pass true) will contain undefined data.

fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

The minimum value is 1, since there is always a main texture.

fn mipmap(&self, level: u32) -> Option<DepthTexture1dMipmap>

Access a single mipmap level of this texture.

fn main_level(&self) -> DepthTexture1dMipmap

Access the main mipmap level of this texture.