Struct glium::texture::RawImage2d

pub struct RawImage2d<'a, T: Clone + 'a> {
    pub data: Cow<'a, [T]>,
    pub width: u32,
    pub height: u32,
    pub format: ClientFormat,
}

Represents raw data for a two-dimensional image.

Fields
data: Cow<'a, [T]>

A contiguous array of pixel data.

The data must start by the bottom-left hand corner pixel and progress left-to-right and bottom-to-top.

data.len() must be equal to width * height * format.get_size() / mem::size_of::<T>().

width: u32

Number of pixels per column.

height: u32

Number of pixels per row.

format: ClientFormat

Formats of the pixels.

Methods

impl<'a, T: Clone + 'a> RawImage2d<'a, T>

pub fn from_raw_rgb(data: Vec<T>, dimensions: (u32, u32)) -> RawImage2d<'a, T> where
T: ToClientFormat,

Builds a raw image from a vector of interleaved RGB values.

The first pixel is at (0, 0), the last pixel is at (1, 1).

pub fn from_raw_rgba(data: Vec<T>, dimensions: (u32, u32)) -> RawImage2d<'a, T> where
T: ToClientFormat,

Builds a raw image from a vector of interleaved RGBA values.

The first pixel is at (0, 0), the last pixel is at (1, 1).

pub fn from_raw_rgb_reversed(
data: &[T],
dimensions: (u32, u32)
) -> RawImage2d<'a, T> where
T: ToClientFormat,

Builds a raw image from a vector of interleaved RGB values, flipping it vertically.

The first pixel is at (0, 1), the last pixel is at (1, 0).

pub fn from_raw_rgba_reversed(
data: &[T],
dimensions: (u32, u32)
) -> RawImage2d<'a, T> where
T: ToClientFormat,

Builds a raw image from a vector of interleaved RGBA values, flipping it vertically.

The first pixel is at (0, 1), the last pixel is at (1, 0).

pub fn from_vec_raw1d(arr: &Vec<RawImage1d<'a, T>>) -> RawImage2d<'a, T>

Transforms a Vec into a RawImage2d