Struct glium::texture::StencilTexture2d

pub struct StencilTexture2d(_);

A two-dimensional texture containing stencil data.

Methods

impl StencilTexture2d

fn new<'a, F, T>(facade: &F, data: T) -> StencilTexture2d where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

This function will automatically generate all mipmaps of the texture.

fn new_if_supported<'a, F, T>(facade: &F, data: T) -> Option<StencilTexture2d> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

This function will automatically generate all mipmaps of the texture.

fn with_mipmaps<'a, F, T>(facade: &F, data: T, mipmaps: bool) -> StencilTexture2d where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn with_mipmaps_if_supported<'a, F, T>(facade: &F, data: T, mipmaps: bool) -> Option<StencilTexture2d> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn with_format<'a, F, T>(facade: &F, data: T, format: StencilFormat, mipmaps: bool) -> Result<StencilTexture2d, TextureCreationError> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn with_format_if_supported<'a, F, T>(facade: &F, data: T, format: StencilFormat, mipmaps: bool) -> Result<StencilTexture2d, TextureMaybeSupportedCreationError> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn new_empty<F>(facade: &F, format: StencilFormat, width: u32, height: u32) -> StencilTexture2d where F: Facade

Creates an empty texture.

The texture will contain undefined data.

fn empty<F>(facade: &F, width: u32, height: u32) -> StencilTexture2d where F: Facade

Creates an empty texture.

No mipmaps will be created.

The texture will contain undefined data.

fn empty_if_supported<F>(facade: &F, width: u32, height: u32) -> Option<StencilTexture2d> where F: Facade

Creates an empty texture.

No mipmaps will be created.

The texture will contain undefined data.

fn empty_with_format<F>(facade: &F, format: StencilFormat, mipmaps: bool, width: u32, height: u32) -> Result<StencilTexture2d, TextureCreationError> where F: Facade

Creates an empty texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

The texture (and its mipmaps, if you pass true) will contain undefined data.

fn empty_with_format_if_supported<F>(facade: &F, format: StencilFormat, mipmaps: bool, width: u32, height: u32) -> Result<StencilTexture2d, TextureMaybeSupportedCreationError> where F: Facade

Creates an empty texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

The texture (and its mipmaps, if you pass true) will contain undefined data.

fn empty_with_mipmaps<F>(facade: &F, mipmaps: bool, width: u32, height: u32) -> StencilTexture2d where F: Facade

Creates an empty texture. Specifies whether is has mipmaps.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

The texture (and its mipmaps, if you pass true) will contain undefined data.

fn empty_with_mipmaps_if_supported<F>(facade: &F, mipmaps: bool, width: u32, height: u32) -> Option<StencilTexture2d> where F: Facade

Creates an empty texture. Specifies whether is has mipmaps.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

The texture (and its mipmaps, if you pass true) will contain undefined data.

fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

The minimum value is 1, since there is always a main texture.

fn mipmap(&self, level: u32) -> Option<StencilTexture2dMipmap>

Access a single mipmap level of this texture.

fn main_level(&self) -> StencilTexture2dMipmap

Access the main mipmap level of this texture.