Struct glium::texture::Texture2d

pub struct Texture2d(_);

A two-dimensional texture containing floating-point data.

Methods

impl Texture2d

fn new<'a, F, T>(facade: &F, data: T) -> Texture2d where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

This function will automatically generate all mipmaps of the texture.

fn with_mipmaps<'a, F, T>(facade: &F, data: T, mipmaps: bool) -> Texture2d where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture by uploading data.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn with_format<'a, F, T>(facade: &F, data: T, format: UncompressedFloatFormat, mipmaps: bool) -> Result<Texture2d, TextureCreationError> where T: Texture2dDataSource<'a>, F: Facade

Builds a new texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

fn new_empty<F>(facade: &F, format: UncompressedFloatFormat, width: u32, height: u32) -> Texture2d where F: Facade

Creates an empty texture.

The texture will contain undefined data.

fn empty<F>(facade: &F, width: u32, height: u32) -> Texture2d where F: Facade

Creates an empty texture.

No mipmaps will be created.

The texture will contain undefined data.

fn empty_if_supported<F>(facade: &F, width: u32, height: u32) -> Option<Texture2d> where F: Facade

Creates an empty texture.

No mipmaps will be created.

The texture will contain undefined data.

fn empty_with_format<F>(facade: &F, format: UncompressedFloatFormat, mipmaps: bool, width: u32, height: u32) -> Result<Texture2d, TextureCreationError> where F: Facade

Creates an empty texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

The texture (and its mipmaps, if you pass true) will contain undefined data.

fn empty_with_format_if_supported<F>(facade: &F, format: UncompressedFloatFormat, mipmaps: bool, width: u32, height: u32) -> Result<Texture2d, TextureMaybeSupportedCreationError> where F: Facade

Creates an empty texture with a specific format.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

The texture (and its mipmaps, if you pass true) will contain undefined data.

fn empty_with_mipmaps<F>(facade: &F, mipmaps: bool, width: u32, height: u32) -> Texture2d where F: Facade

Creates an empty texture. Specifies whether is has mipmaps.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

The texture (and its mipmaps, if you pass true) will contain undefined data.

fn empty_with_mipmaps_if_supported<F>(facade: &F, mipmaps: bool, width: u32, height: u32) -> Option<Texture2d> where F: Facade

Creates an empty texture. Specifies whether is has mipmaps.

Note that passing true for mipmaps does not mean that you will get mipmaps. Instead it indicates that mipmaps are allowed to be created if possible.

The texture (and its mipmaps, if you pass true) will contain undefined data.

fn as_surface<'a>(&'a self) -> TextureSurface<'a>

Starts drawing on the texture.

All the function calls to the TextureSurface will draw on the texture instead of the screen.

Low-level information

The first time that this function is called, a FrameBuffer Object will be created and cached. The following calls to as_surface will load the existing FBO and re-use it. When the texture is destroyed, the FBO is destroyed too.

fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

The minimum value is 1, since there is always a main texture.

fn read<P, T>(&self) -> T where T: Texture2dDataSink<Data=P>, P: PixelValue + Clone

Reads the content of the texture to RAM.

You should avoid doing this at all cost during performance-critical operations (for example, while you're drawing). Use read_to_pixel_buffer instead.

fn read_to_pixel_buffer<P, T>(&self) -> PixelBuffer<T> where T: Texture2dDataSink<Data=P>, P: PixelValue + Clone

Reads the content of the texture into a buffer in video memory.

This operation copies the texture's data into a buffer in video memory (a pixel buffer). Contrary to the read function, this operation is done asynchronously and doesn't need a synchronization.

fn write<'a, T>(&self, rect: Rect, data: T) where T: Texture2dDataSource<'a>

Uploads some data in the texture.

Note that this may cause a synchronization if you use the texture right before or right after this call. Prefer creating a whole new texture if you change a huge part of it.

Panic

Panics if the the dimensions of data don't match the Rect.

fn mipmap(&self, level: u32) -> Option<Texture2dMipmap>

Access a single mipmap level of this texture.

fn main_level(&self) -> Texture2dMipmap

Access the main mipmap level of this texture.