Struct glium::texture::Texture2dMipmap

pub struct Texture2dMipmap<'t>(_, _);

Represents a single mipmap level of a Texture2d.

Can be obtained by calling Texture2d::mipmap() or Texture2d::main_level().

Methods

impl<'t> Texture2dMipmap<'t>

fn write<'a, T>(&self, rect: Rect, data: T) where T: Texture2dDataSource<'a>

Uploads some data in the texture level.

Note that this may cause a synchronization if you use the texture right before or right after this call.

Panic

Panics if the the dimensions of data don't match the Rect.

fn get_texture(&self) -> &'t Texture2d

Returns the corresponding texture.

fn get_level(&self) -> u32

Returns the texture level.