Struct glium::texture::TextureAny

pub struct TextureAny { /* fields omitted */ }

A texture whose type isn't fixed at compile-time.

Methods

impl TextureAny

pub fn get_width(&self) -> u32

Returns the width of the texture.

pub fn get_height(&self) -> Option<u32>

Returns the height of the texture.

pub fn get_depth(&self) -> Option<u32>

Returns the depth of the texture.

pub fn kind(&self) -> TextureKind

Returns the kind of texture.

pub fn dimensions(&self) -> Dimensions

Returns the dimensions of the texture.

pub fn get_array_size(&self) -> Option<u32>

Returns the array size of the texture.

pub fn get_samples(&self) -> Option<u32>

Returns the number of samples of the texture if it is a multisampling texture.

pub fn first_layer(&self) -> TextureAnyLayer

Returns a structure that represents the first layer of the texture. All textures have a first layer.

pub fn layer(&self, layer: u32) -> Option<TextureAnyLayer>

Returns a structure that represents a specific layer of the texture.

Non-array textures have only one layer. The number of layers can be queried with get_array_size.

Returns None if out of range.

pub fn get_texture_type(&self) -> Dimensions

Returns the type of the texture (1D, 2D, 3D, etc.).

pub fn get_internal_format(&self) -> Result<InternalFormat, GetFormatError>

Determines the internal format of this texture.

pub fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

pub fn main_level(&self) -> TextureAnyMipmap

Returns a structure that represents the main mipmap level of the texture.

pub fn mipmap(&self, level: u32) -> Option<TextureAnyMipmap>

Returns a structure that represents a specific mipmap of the texture.

Returns None if out of range.

pub unsafe fn generate_mipmaps(&self)

Binds this texture and generates mipmaps.