Struct glium::texture::TextureAnyLayer

pub struct TextureAnyLayer<'a> { /* fields omitted */ }

Represents a specific layer of an array texture and 3D textures.

Methods

impl<'a> TextureAnyLayer<'a>

pub fn get_texture(&self) -> &'a TextureAny

Returns the texture.

pub fn get_samples(&self) -> Option<u32>

Returns the number of samples of the texture.

pub fn get_layer(&self) -> u32

Returns the layer of the texture.

pub fn main_level(&self) -> TextureAnyLayerMipmap<'a>

Returns a structure that represents the main mipmap level of this layer of the texture.

pub fn mipmap(&self, level: u32) -> Option<TextureAnyLayerMipmap<'a>>

Returns a structure that represents a specific mipmap of this layer of the texture.

Returns None if out of range.