Struct glium::texture::TextureAnyLayerMipmap

pub struct TextureAnyLayerMipmap<'a> {
    // some fields omitted
}

Represents a specific layer of a specific mipmap. This is the same as TextureAnyImage, except for 3D textures, cubemaps and cubemap arrays.

Methods

impl<'a> TextureAnyLayerMipmap<'a>

fn get_texture(&self) -> &'a TextureAny

Returns the texture.

fn get_level(&self) -> u32

Returns the level of the texture.

fn get_layer(&self) -> u32

Returns the layer of the texture.

fn get_width(&self) -> u32

Returns the width of this texture slice.

fn get_height(&self) -> Option<u32>

Returns the height of this texture slice.

fn get_samples(&self) -> Option<u32>

Returns the number of samples of the texture.

fn into_image(&self, cube_layer: Option<CubeLayer>) -> Option<TextureAnyImage<'a>>

Turns this into an image.

Returns None if cube_layer is None and the texture is a cubemap. Returns None if cube_layer is Some and the texture is not a cubemap.