Module glium::texture::unsigned_cubemap_array

Contains the implementation of UnsignedCubemapArray.

Structs

UnsignedCubemapArray
An array of cube textures containing unsigned integral data.

UnsignedCubemapArrayImage
Represents a single image of a mipmap level of a layer of UnsignedCubemapArray.

UnsignedCubemapArrayLayer
Represents a single layer of a UnsignedCubemapArray.

UnsignedCubemapArrayLayerMipmap
Represents a single layer of a mipmap level of a UnsignedCubemapArray.

UnsignedCubemapArrayMipmap
Represents a single mipmap level of a UnsignedCubemapArray.