Trait glium::DisplayBuild

pub trait DisplayBuild {
    type Facade: Facade;
    type Err;
    fn build_glium_debug(self, DebugCallbackBehavior) -> Result<Self::Facade, Self::Err>;
    unsafe fn build_glium_unchecked_debug(self, DebugCallbackBehavior) -> Result<Self::Facade, Self::Err>;
    fn rebuild_glium(self, &Self::Facade) -> Result<(), Self::Err>;

    fn build_glium(self) -> Result<Self::Facade, Self::Err> where Self: Sized { ... }
    unsafe fn build_glium_unchecked(self) -> Result<Self::Facade, Self::Err> where Self: Sized { ... }
}

Objects that can build a facade object.

Associated Types

type Facade: Facade

The object that this DisplayBuild builds.

type Err

The type of error that initialization can return.

Required Methods

fn build_glium_debug(self, DebugCallbackBehavior) -> Result<Self::Facade, Self::Err>

Build a context and a facade to draw on it.

Performs a compatibility check to make sure that all core elements of glium are supported by the implementation.

unsafe fn build_glium_unchecked_debug(self, DebugCallbackBehavior) -> Result<Self::Facade, Self::Err>

Build a context and a facade to draw on it

This function does the same as build_glium, except that the resulting context will assume that the current OpenGL context will never change.

fn rebuild_glium(self, &Self::Facade) -> Result<(), Self::Err>

Changes the settings of an existing facade.

Provided Methods

fn build_glium(self) -> Result<Self::Facade, Self::Err> where Self: Sized

Build a context and a facade to draw on it.

Performs a compatibility check to make sure that all core elements of glium are supported by the implementation.

unsafe fn build_glium_unchecked(self) -> Result<Self::Facade, Self::Err> where Self: Sized

Build a context and a facade to draw on it

This function does the same as build_glium, except that the resulting context will assume that the current OpenGL context will never change.