Trait glium::SurfaceUnstable

pub trait Surface {
    fn clear_color(&mut self, red: f32, green: f32, blue: f32, alpha: f32);
    fn clear_depth(&mut self, value: f32);
    fn clear_stencil(&mut self, value: int);
    fn get_dimensions(&self) -> (uint, uint);
    fn get_depth_buffer_bits(&self) -> Option<u16>;
    fn get_stencil_buffer_bits(&self) -> Option<u16>;
    fn draw<'a, 'b, V, I, ID, U>(&mut self, V, &I, program: &Program, uniforms: U, draw_parameters: &DrawParameters) where V: IntoVerticesSource<'b>, I: ToIndicesSource<ID>, U: Uniforms;
    fn get_blit_helper(&self) -> BlitHelper;

    fn has_depth_buffer(&self) -> bool { ... }
    fn has_stencil_buffer(&self) -> bool { ... }
    fn blit_color<S>(&self, source_rect: &Rect, target: &S, target_rect: &Rect, filter: MagnifySamplerFilter) where S: Surface { ... }
    fn blit_whole_color_to<S>(&self, target: &S, target_rect: &Rect, filter: MagnifySamplerFilter) where S: Surface { ... }
    fn fill<S>(&self, target: &S, filter: MagnifySamplerFilter) where S: Surface { ... }
}

Object which can be drawn upon.

Required Methods

fn clear_color(&mut self, red: f32, green: f32, blue: f32, alpha: f32)

Clears the color components of the target.

fn clear_depth(&mut self, value: f32)

Clears the depth component of the target.

fn clear_stencil(&mut self, value: int)

Clears the stencil component of the target.

fn get_dimensions(&self) -> (uint, uint)

Returns the dimensions in pixels of the target.

fn get_depth_buffer_bits(&self) -> Option<u16>

Returns the number of bits of each pixel of the depth buffer.

Returns None if there is no depth buffer.

fn get_stencil_buffer_bits(&self) -> Option<u16>

Returns the number of bits of each pixel of the stencil buffer.

Returns None if there is no stencil buffer.

fn draw<'a, 'b, V, I, ID, U>(&mut self, V, &I, program: &Program, uniforms: U, draw_parameters: &DrawParameters) where V: IntoVerticesSource<'b>, I: ToIndicesSource<ID>, U: Uniforms

Draws.

Panic

Panics if the requested depth function requires a depth buffer and none is attached.

Panics if the type of some of the vertex source's attributes do not match the program's.

Panics if a program's attribute is not in the vertex source (does not panic if a vertex's attribute is not used by the program).

Panics if the viewport is larger than the dimensions supported by the hardware.

Panics if the depth range is outside of (0, 1).

fn get_blit_helper(&self) -> BlitHelper

Returns an opaque type that is used by the implementation of blit functions.

Provided Methods

fn has_depth_buffer(&self) -> bool

Returns true if the surface has a depth buffer available.

fn has_stencil_buffer(&self) -> bool

Returns true if the surface has a stencil buffer available.

fn blit_color<S>(&self, source_rect: &Rect, target: &S, target_rect: &Rect, filter: MagnifySamplerFilter) where S: Surface

Copies a rectangle of pixels from this surface to another surface.

The source_rect defines the area of the source (self) that will be copied, and the target_rect defines the area where the copied image will be pasted. If the source and target areas don't have the same dimensions, the image will be resized to match. The filter parameter is relevant only in this situation.

It is possible for the source and the target to be the same surface. However if the rectangles overlap, then the behavior is undefined.

Note that there is no alpha blending, depth/stencil checking, etc. or anything ; this function just copies pixels.

fn blit_whole_color_to<S>(&self, target: &S, target_rect: &Rect, filter: MagnifySamplerFilter) where S: Surface

Copies the entire surface to a target surface. See blit_color.

fn fill<S>(&self, target: &S, filter: MagnifySamplerFilter) where S: Surface

Copies the entire surface to the entire target. See blit_color.