glium::VertexFormatUnstable

type VertexFormat = Vec<(String, usize, AttributeType)>;

Describes the layout of each vertex in a vertex buffer.

The first element is the name of the binding, the second element is the offset from the start of each vertex to this element, and the third element is the type.