Struct glium::uniforms::UniformBuffer

pub struct UniformBuffer<T> {
    // some fields omitted
}

Buffer that contains a uniform block.

Methods

impl<T> UniformBuffer<T> where T: Copy + Send + 'static

fn new(display: &Display, data: T) -> UniformBuffer<T>

Uploads data in the uniforms buffer.

Features

Only available if the gl_uniform_blocks feature is enabled.

fn new_if_supported(display: &Display, data: T) -> Option<UniformBuffer<T>>

Uploads data in the uniforms buffer.

fn new_persistent(display: &Display, data: T) -> UniformBuffer<T>

Builds a new uniform buffer with persistent mapping.

Features

Only available if the gl_uniform_blocks and gl_persistent_mapping features are both enabled.

fn new_persistent_if_supported(display: &Display, data: T) -> Option<UniformBuffer<T>>

Builds a new uniform buffer with persistent mapping, or None if this is not supported.

fn upload(&mut self, data: T)

Modifies the content of the buffer.

fn map<'a>(&'a mut self) -> Mapping<'a, T>

Maps the buffer to allow write access to it.

This function will block until the buffer stops being used by the backend. This operation is much faster if the buffer is persistent.

fn read(&self) -> T

Reads the content of the buffer.

This function is usually better if are just doing one punctual read, while map is better if you want to have multiple small reads.

Features

Only available if the gl_read_buffer feature is enabled.

fn read_if_supported(&self) -> Option<T>

Reads the content of the buffer.