Enum glium::vertex::BufferCreationError

pub enum BufferCreationError {
    FormatNotSupported,
    BufferCreationError(BufferCreationError),
}

Error that can happen when creating a vertex buffer.

Variants
FormatNotSupported

The vertex format is not supported by the backend.

Anything 64bits-related may not be supported.

BufferCreationError(BufferCreationError)

Error while creating the vertex buffer.