Struct glium::vertex::VertexBuffer

pub struct VertexBuffer<T> where
    T: Copy,  { /* fields omitted */ }

A list of vertices loaded in the graphics card's memory.

Methods

impl<T> VertexBuffer<T> where
T: Vertex,

pub fn new<F: ?Sized>(
facade: &F,
data: &[T]
) -> Result<VertexBuffer<T>, CreationError> where
F: Facade,

Builds a new vertex buffer.

Note that operations such as write will be very slow. If you want to modify the buffer from time to time, you should use the dynamic function instead.

Example

#[derive(Copy, Clone)]
struct Vertex {
    position: [f32; 3],
    texcoords: [f32; 2],
}

implement_vertex!(Vertex, position, texcoords);

let vertex_buffer = glium::VertexBuffer::new(&display, &[
    Vertex { position: [0.0,  0.0, 0.0], texcoords: [0.0, 1.0] },
    Vertex { position: [5.0, -3.0, 2.0], texcoords: [1.0, 0.0] },
]);

pub fn dynamic<F: ?Sized>(
facade: &F,
data: &[T]
) -> Result<VertexBuffer<T>, CreationError> where
F: Facade,

Builds a new vertex buffer.

This function will create a buffer that is intended to be modified frequently.

pub fn persistent<F: ?Sized>(
facade: &F,
data: &[T]
) -> Result<VertexBuffer<T>, CreationError> where
F: Facade,

Builds a new vertex buffer.

pub fn immutable<F: ?Sized>(
facade: &F,
data: &[T]
) -> Result<VertexBuffer<T>, CreationError> where
F: Facade,

Builds a new vertex buffer.

pub fn empty<F: ?Sized>(
facade: &F,
elements: usize
) -> Result<VertexBuffer<T>, CreationError> where
F: Facade,

Builds an empty vertex buffer.

The parameter indicates the number of elements.

pub fn empty_dynamic<F: ?Sized>(
facade: &F,
elements: usize
) -> Result<VertexBuffer<T>, CreationError> where
F: Facade,

Builds an empty vertex buffer.

The parameter indicates the number of elements.

pub fn empty_persistent<F: ?Sized>(
facade: &F,
elements: usize
) -> Result<VertexBuffer<T>, CreationError> where
F: Facade,

Builds an empty vertex buffer.

The parameter indicates the number of elements.

pub fn empty_immutable<F: ?Sized>(
facade: &F,
elements: usize
) -> Result<VertexBuffer<T>, CreationError> where
F: Facade,

Builds an empty vertex buffer.

The parameter indicates the number of elements.

impl<T> VertexBuffer<T> where
T: Copy,

pub unsafe fn new_raw<F: ?Sized>(
facade: &F,
data: &[T],
bindings: VertexFormat,
elements_size: usize
) -> Result<VertexBuffer<T>, CreationError> where
F: Facade,

Builds a new vertex buffer from an indeterminate data type and bindings.

Example

use std::borrow::Cow;

let bindings = Cow::Owned(vec![(
        Cow::Borrowed("position"), 0,
        glium::vertex::AttributeType::F32F32,
    ), (
        Cow::Borrowed("color"), 2 * ::std::mem::size_of::<f32>(),
        glium::vertex::AttributeType::F32,
    ),
]);

let data = vec![
    1.0, -0.3, 409.0,
    -0.4, 2.8, 715.0f32
];

let vertex_buffer = unsafe {
    glium::VertexBuffer::new_raw(&display, &data, bindings, 3 * ::std::mem::size_of::<f32>())
};

pub unsafe fn new_raw_dynamic<F: ?Sized>(
facade: &F,
data: &[T],
bindings: VertexFormat,
elements_size: usize
) -> Result<VertexBuffer<T>, CreationError> where
F: Facade,

Dynamic version of new_raw.

pub fn slice<R: RangeArgument<usize>>(
&self,
range: R
) -> Option<VertexBufferSlice<T>>

Accesses a slice of the buffer.

Returns None if the slice is out of range.

pub fn get_bindings(&self) -> &VertexFormat

Returns the associated VertexFormat.

pub fn per_instance(&self) -> Result<PerInstance, InstancingNotSupported>

Creates a marker that instructs glium to use multiple instances.

Instead of calling surface.draw(&vertex_buffer, ...) you can call surface.draw(vertex_buffer.per_instance(), ...). This will draw one instance of the geometry for each element in this buffer. The attributes are still passed to the vertex shader, but each entry is passed for each different instance.

impl<T> VertexBuffer<T> where
T: Copy + Send + 'static,

pub fn into_vertex_buffer_any(self) -> VertexBufferAny

DEPRECATED: use .into() instead. Discard the type information and turn the vertex buffer into a VertexBufferAny.