Trait glium::vertex::Vertex

pub trait Vertex: Copy + Sized {
    fn build_bindings() -> VertexFormat;

    fn is_supported<C: ?Sized>(caps: &C) -> bool
    where
        C: CapabilitiesSource,
    { ... }
}

Trait for structures that represent a vertex.

Instead of implementing this trait yourself, it is recommended to use the implement_vertex! macro instead.

Required Methods

fn build_bindings() -> VertexFormat

Builds the VertexFormat representing the layout of this element.

Provided Methods

fn is_supported<C: ?Sized>(caps: &C) -> bool where
C: CapabilitiesSource,

Returns true if the backend supports this vertex format.