Module glium::vertex_bufferUnstable

Contains everything related to vertex buffers.

The main struct is the VertexBuffer, which represents a buffer in the video memory containing a list of vertices.

In order to create a vertex buffer, you must first create a struct that represents each vertex and implement the glium::vertex_buffer::Vertex trait on it. The #[vertex_format] attribute coming from glium_macros helps you doing that.

#[vertex_format]
#[derive(Copy)]
struct Vertex {
    position: [f32; 3],
    texcoords: [f32; 2],
}

Then you must build a Vec of the vertices that you want to upload, and pass it to VertexBuffer::new.

let data = vec![
    Vertex {
        position: [0.0, 0.0, 0.4],
        texcoords: [0.0, 1.0]
    },
    Vertex {
        position: [12.0, 4.5, -1.8],
        texcoords: [1.0, 0.5]
    },
    Vertex {
        position: [-7.124, 0.1, 0.0],
        texcoords: [0.0, 0.4]
    },
];

let vertex_buffer = glium::vertex_buffer::VertexBuffer::new(&display, data);

Structs

Mapping
A mapping of a buffer.

VertexBuffer
A list of vertices loaded in the graphics card's memory.

VertexBufferAny
A list of vertices loaded in the graphics card's memory.

Enums

AttributeType

VerticesSource
Describes the source to use for the vertices when drawing.

Traits

Attribute
Trait for types that can be used as vertex attributes.

IntoVerticesSource
Objects that can be used as vertex sources.

Vertex
Trait for structures that represent a vertex.

Type Definitions

VertexFormat
Describes the layout of each vertex in a vertex buffer.