Struct glium::vertex_buffer::VertexBufferUnstable

pub struct VertexBuffer<T> {
    // some fields omitted
}

A list of vertices loaded in the graphics card's memory.

Methods

impl<T: Vertex + 'static + Send> VertexBuffer<T>

fn new(display: &Display, data: Vec<T>) -> VertexBuffer<T>

Builds a new vertex buffer.

Example

#[vertex_format]
#[derive(Copy)]
struct Vertex {
    position: [f32; 3],
    texcoords: [f32; 2],
}

let vertex_buffer = glium::VertexBuffer::new(&display, vec![
    Vertex { position: [0.0,  0.0, 0.0], texcoords: [0.0, 1.0] },
    Vertex { position: [5.0, -3.0, 2.0], texcoords: [1.0, 0.0] },
]);

fn new_dynamic(display: &Display, data: Vec<T>) -> VertexBuffer<T>

Builds a new vertex buffer.

This function will create a buffer that has better performances when it is modified frequently.

impl<T: Send + Copy> VertexBuffer<T>

unsafe fn new_raw(display: &Display, data: Vec<T>, bindings: VertexFormat, elements_size: usize) -> VertexBuffer<T>

Builds a new vertex buffer from an undeterminate data type and bindings.

Example

let bindings = vec![(
        format!("position"), 0,
        glium::vertex_buffer::AttributeType::F32F32,
    ), (
        format!("color"), 2 * ::std::mem::size_of::<f32>(),
        glium::vertex_buffer::AttributeType::F32,
    ),
];

let data = vec![
    1.0, -0.3, 409.0,
    -0.4, 2.8, 715.0f32
];

let vertex_buffer = unsafe {
    glium::VertexBuffer::new_raw(&display, data, bindings, 3 * ::std::mem::size_of::<f32>())
};

fn map<'a>(&'a mut self) -> Mapping<'a, T>

Maps the buffer to allow write access to it.

Warning: using this function can slow things down a lot because the function waits for all the previous commands to be executed before returning.

Panic

OpenGL ES doesn't support mapping buffers. Using this function will thus panic. If you want to be compatible with all platforms, it is preferable to disable the gl_extensions feature.

Features

Only available if the gl_extensions feature is enabled.

fn read(&self) -> Vec<T>

Reads the content of the buffer.

This function is usually better if are just doing one punctual read, while map is better if you want to have multiple small reads.

Panic

OpenGL ES doesn't support reading buffers. Using this function will thus panic. If you want to be compatible with all platforms, it is preferable to disable the gl_extensions feature.

Features

Only available if the gl_extensions feature is enabled.

fn read_slice(&self, offset: usize, size: usize) -> Vec<T>

Reads the content of the buffer.

This function is usually better if are just doing one punctual read, while map is better if you want to have multiple small reads.

The offset and size are expressed in number of elements.

Panic

Panics if offset or offset + size are greated than the size of the buffer.

OpenGL ES doesn't support reading buffers. Using this function will thus panic. If you want to be compatible with all platforms, it is preferable to disable the gl_extensions feature.

Features

Only available if the gl_extensions feature is enabled.

impl<T> VertexBuffer<T>

fn get_elements_size(&self) -> usize

Returns the number of bytes between two consecutive elements in the buffer.

fn get_bindings(&self) -> &VertexFormat

Returns the associated VertexFormat.

fn into_vertex_buffer_any(self) -> VertexBufferAny

Lose the type informations and turn the vertex buffer into a VertexBufferAny.