Struct glium::vertex_buffer::VertexBufferAnyUnstable

pub struct VertexBufferAny {
    // some fields omitted
}

A list of vertices loaded in the graphics card's memory.

Contrary to VertexBuffer, this struct doesn't know about the type of data inside the buffer. Therefore you can't map or read it.

This struct is provided for convenience, so that you can have a Vec<VertexBufferAny>, or return a VertexBufferAny instead of a VertexBuffer<MyPrivateVertexType>.

Methods

impl VertexBufferAny

fn get_elements_size(&self) -> usize

Returns the number of bytes between two consecutive elements in the buffer.

fn get_bindings(&self) -> &VertexFormat

Returns the associated VertexFormat.

unsafe fn into_vertex_buffer<T>(self) -> VertexBuffer<T>

Turns the vertex buffer into a VertexBuffer without checking the type.