Crate glium

Easy-to-use, high-level, OpenGL3+ wrapper.

Initialization

This library defines the DisplayBuild trait which is curently implemented only on glutin::WindowBuilder.

Initialization is done by creating a WindowBuilder and calling build_glium.

extern crate glutin;
extern crate glium;

fn main() {
    use glium::DisplayBuild;

    let display = glutin::WindowBuilder::new()
        .with_dimensions(1024, 768)
        .with_title(format!("Hello world"))
        .build_glium().unwrap();
}

The display object is the most important object of this library.

The window you are drawing on will produce events. They can be received by calling display.poll_events().

Complete example

The first step is to create the vertex buffer, which contains the list of all the points that make up our mesh. The elements that we pass to VertexBuffer::new must implement the glium::vertex::VertexFormat trait, which can be easily added for any custom struct thanks to the implement_vertex! macro.

See the vertex module documentation for more informations.

#[derive(Copy, Clone)]
struct Vertex {
    position: [f32; 2],
    color: [f32; 3],
}

implement_vertex!(Vertex, position, color);

let vertex = glium::VertexBuffer::new(&display, vec![
    Vertex { position: [-0.5, -0.5], color: [0.0, 1.0, 0.0] },
    Vertex { position: [ 0.0,  0.5], color: [0.0, 0.0, 1.0] },
    Vertex { position: [ 0.5, -0.5], color: [1.0, 0.0, 0.0] },
]);

We will also need to tell glium how the vertices must be linked together. We could create an index buffer, but since we only have a single triangle the simpler solution here is not to use indices.

use glium::index;
let indices = index::NoIndices(index::PrimitiveType::TrianglesList);

Next, we create the program, which is composed of a vertex shader, a program executed once for each element in our vertex buffer, and a fragment shader, a program executed once for each pixel before it is written on the final image.

The purpose of a program is to instruct the GPU how to process our mesh, in order to obtain pixels.

let program = glium::Program::from_source(&display,
    // vertex shader
    "   #version 110

        uniform mat4 matrix;

        attribute vec2 position;
        attribute vec3 color;

        varying vec3 v_color;

        void main() {
            gl_Position = vec4(position, 0.0, 1.0) * matrix;
            v_color = color;
        }
    ",

    // fragment shader
    "   #version 110
        varying vec3 v_color;

        void main() {
            gl_FragColor = vec4(v_color, 1.0);
        }
    ",

    // optional geometry shader
    None
).unwrap();

Note: teaching you the GLSL language is not covered by this guide.

You may notice that the attribute declarations in the vertex shader match the field names and types of the elements in the vertex buffer. This is required, otherwise drawing will result in an error.

In the example above, one of our shaders contains uniform mat4 matrix;. Uniforms are global variables in our program whose values are chosen by the application.

let uniforms = uniform! {
    matrix: [
        [ 1.0, 0.0, 0.0, 0.0 ],
        [ 0.0, 1.0, 0.0, 0.0 ],
        [ 0.0, 0.0, 1.0, 0.0 ],
        [ 0.0, 0.0, 0.0, 1.0 ]
    ]
};

The value of uniforms can be of any type that implements glium::uniforms::UniformValue. This includes textures and samplers (not covered here). See the uniforms module documentation for more informations.

Now that everything is initialized, we can finally draw something. The display.draw() function will start drawing a new frame and return a Frame object. This Frame object has a draw function, which you can use to draw things.

Its arguments are the source of vertices, source of indices, program, uniforms, and an object of type DrawParameters which contains miscellaneous information specifying how everything should be rendered (depth test, blending, backface culling, etc.).

use glium::Surface;
let mut target = display.draw();
target.clear_color(0.0, 0.0, 0.0, 0.0);  // filling the output with the black color
target.draw(&vertex_buffer, &indices, &program, &uniforms,
            &std::default::Default::default()).unwrap();
target.finish();

Reexports

pub use vertex::{Vertex};

pub use program::ProgramCreationError::{CompilationError, LinkingError, ShaderTypeNotSupported};

pub use texture::{Texture, Texture2d};

Modules

backend
The backend module allows one to link between glium and the OpenGL context..

debug

framebuffer
Framebuffers allow you to customize the color, depth and stencil buffers you will draw on.

index
In order to draw, you need to provide a source of indices which is used to link the vertices together into primitives.

macros
Defines useful macros for glium usage.

pixel_buffer
Pixel buffers are buffers that contain two-dimensional texture data.

program

render_buffer
A render buffer is similar to a texture, but is optimized for usage as a draw target.

texture
A texture is an image loaded in video memory, which can be sampled in your shaders.

uniforms
A uniform is a global variable in your program. In order to draw something, you will need to give glium the values of all your uniforms. Objects that implement the Uniform trait are here to do that.

vertex
Contains everything related to vertex buffers.

Macros

implement_vertex!
Implements the glium::vertex::Vertex trait for the given type.

program!
Builds a program depending on the GLSL version supported by the backend.

uniform!
Returns an implementation-defined type which implements the Uniform trait.

Structs

BlitTarget
Area of a surface in pixels. Similar to a Rect except that dimensions can be negative.

DrawParameters
Represents the parameters to use when drawing.

Frame
Implementation of Surface, targeting the default framebuffer.

IndexBuffer
A list of indices loaded in the graphics card's memory.

LinearSyncFence
Prototype for a SyncFence.

Program
A combination of shaders linked together.

Rect
Area of a surface in pixels.

SyncFence
Provides a way to wait for a server-side operation to be finished.

Version
Describes a version.

VertexBuffer
A list of vertices loaded in the graphics card's memory.

Enums

Api
Describes an OpenGL-related API.

BackfaceCullingMode
Describes how triangles should be filtered before the fragment processing. Backface culling is purely an optimization. If you don't know what this does, just use CullingDisabled.

BlendingFunction
Function that the GPU will use for blending.

DepthTest
The function that the GPU will use to determine whether to write over an existing pixel on the target.

DrawError
Error that can happen while drawing.

GliumCreationError
Error that can happen while creating a glium display.

Handle
Handle to a shader or a program.

LinearBlendingFactor
Indicates which value to multiply each component with.

PolygonMode
Defines how the device should render polygons.

ProgramCreationError
Error that can be triggered when creating a Program.

StencilOperation
Specificies which operation the GPU will do depending on the result of the stencil test.

StencilTest
Specifies which comparaison the GPU will do to determine whether a sample passes the stencil test.

Traits

DisplayBuild
Objects that can build a facade object.

GlObject
Trait for objects that are OpenGL objects.

Surface
Object that can be drawn upon.

Functions

get_supported_glsl_version
Given an API version, this function returns the GLSL version that the implementation is required to support.

Type Definitions

Display
The main object of this library. Controls the whole display.

VertexFormat
Describes the layout of each vertex in a vertex buffer.