Crate gliumUnstable

Easy-to-use high-level OpenGL3+ wrapper.

Initialization

This library defines the DisplayBuild trait which is curently implemented only on glutin::WindowBuilder.

Initialization is done by creating a WindowBuilder and calling build_glium.

extern crate glutin;
extern crate glium;

fn main() {
    use glium::DisplayBuild;

    let display = glutin::WindowBuilder::new()
        .with_dimensions(1024, 768)
        .with_title(format!("Hello world"))
        .build_glium().unwrap();
}

The display object is the most important object of this library.

The window where you are drawing on will produce events. They can be received by calling display.poll_events().

Complete example

We start by creating the vertex buffer, which contains the list of all the points that are part of our mesh. The elements that we pass to VertexBuffer::new must implement the glium::vertex_buffer::VertexFormat trait. We can easily do this by creating a custom struct and adding the #[vertex_format] attribute to it.

You can check the documentation of the vertex_buffer module for more informations.

#[plugin]
extern crate glium_macros;

#[vertex_format]
#[derive(Copy)]
struct Vertex {
    position: [f32; 2],
    color: [f32; 3],
}

let vertex_buffer = glium::VertexBuffer::new(&display, vec![
    Vertex { position: [-0.5, -0.5], color: [0.0, 1.0, 0.0] },
    Vertex { position: [ 0.0,  0.5], color: [0.0, 0.0, 1.0] },
    Vertex { position: [ 0.5, -0.5], color: [1.0, 0.0, 0.0] },
]);

Then we create the index buffer, which contains informations about the primitives (triangles, lines, etc.) that compose our mesh.

The last parameter is a list of indices that represent the positions of our points in the vertex buffer.

let index_buffer = glium::IndexBuffer::new(&display,
    glium::index_buffer::TrianglesList(vec![ 0u16, 1, 2 ]));

Then we create the program, which is composed of a vertex shader, a program executed once for each element in our vertex buffer, and a fragment shader, a program executed once for each pixel before it is written on the final image.

The purpose of a program is to instruct the GPU how to process our mesh in order to obtain pixels.

let program = glium::Program::from_source(&display,
    // vertex shader
    "   #version 110

        uniform mat4 matrix;

        attribute vec2 position;
        attribute vec3 color;

        varying vec3 v_color;

        void main() {
            gl_Position = vec4(position, 0.0, 1.0) * matrix;
            v_color = color;
        }
    ",

    // fragment shader
    "   #version 110
        varying vec3 v_color;

        void main() {
            gl_FragColor = vec4(v_color, 1.0);
        }
    ",

    // optional geometry shader
    None
).unwrap();

Note: Teaching you the GLSL language is not covered by this guide.

You may notice that the attribute declarations in the vertex shader match the field names and types of the elements in the vertex buffer. This is required, or drawing will result in an error.

In the example above, you may notice uniform mat4 matrix;. This is a uniform, in other words a global variable in our program. We will need to tell glium what the value of matrix is by creating an object that implements the glium::uniforms::Uniforms trait.

Similarly to the vertex buffer and vertex format, we can do so by creating a custom struct and adding the #[uniforms] attribute to it.

#[plugin]
extern crate glium_macros;

#[uniforms]
struct Uniforms {
    matrix: [[f32; 4]; 4],
}

let uniforms = Uniforms {
    matrix: [
        [ 1.0, 0.0, 0.0, 0.0 ],
        [ 0.0, 1.0, 0.0, 0.0 ],
        [ 0.0, 0.0, 1.0, 0.0 ],
        [ 0.0, 0.0, 0.0, 1.0 ]
    ],
};

Vertex buffers, index buffers and program should be stored between draws in order to avoid wasting time, but objects that implement the glium::uniforms::Uniforms trait are usually constructed every time you draw.

Fields of our Uniforms object can be any object that implements glium::uniforms::UniformValue. This includes textures and samplers (not covered here). You can check the documentation of the uniforms module for more informations.

Now that everything is initialized, we can finally draw something. To do so, call display.draw() in order to obtain a Frame object. Note that it is also possible to draw on a texture by calling texture.as_surface(), but this is not covered here.

The Frame object has a draw function which you can use to draw things. Its arguments are the vertex buffer, index buffer, program, uniforms, and an object of type DrawParameters which contains miscellaneous informations about how everything should be rendered (depth test, blending, backface culling, etc.).

use glium::Surface;
let mut target = display.draw();
target.clear_color(0.0, 0.0, 0.0, 0.0);  // filling the output with the black color
target.draw(&vertex_buffer, &index_buffer, &program, &uniforms, &std::default::Default::default());
target.finish();

Reexports

pub use index_buffer::IndexBuffer;

pub use vertex_buffer::{VertexBuffer, Vertex, VertexFormat};

pub use program::ProgramCreationError::{CompilationError, LinkingError, ShaderTypeNotSupported};

pub use texture::{Texture, Texture2d};

Modules

debug

framebuffer
Framebuffers allows you to customize the color, depth and stencil buffers you will draw on.

index_buffer
In order to draw, you need to provide a source of indices which is used to link the vertices together into primitives.

pixel_buffer
Pixel buffers are buffer that contain two-dimensional texture data.

render_buffer
A render buffer is similar to a texture, but is optimized for usage as a draw target.

texture
A texture is an image loaded in video memory and that can be sampled in your shaders.

uniforms
A uniform is a global variable in your program. In order to draw something, you will need to tell glium what the values of all your uniforms are. Objects that implement the Uniform trait are here to do that.

vertex_buffer
Contains everything related to vertex buffers.

Structs

Display
The main object of this library. Controls the whole display.

DrawParameters
Represents the parameters to use when drawing.

Frame
Implementation of Surface targetting the default framebuffer.

Program
A combinaison of shaders linked together.

Rect
Area of a surface in pixels.

Enums

BackfaceCullingMode
Describes how triangles should be filtered before the fragment processing. Backface culling is purely an optimization. If you don't know what this does, just use CullingDisabled.

BlendingFunction
Function that the GPU will use for blending.

DepthFunction
The function that the GPU will use to determine whether to write over an existing pixel on the target.

GliumCreationError
Error that can happen while creating a glium display.

PolygonMode
Defines how the device should render polygons.

ProgramCreationError
Error that can be triggered when creating a Program.

Traits

DisplayBuild
Objects that can build a Display object.

Surface
Object which can be drawn upon.