Struct glium::backend::glutin_backend::GlutinFacade

pub struct GlutinFacade {
    // some fields omitted
}

Facade implementation for glutin. Wraps both glium and glutin.

Methods

impl GlutinFacade

fn poll_events(&self) -> PollEventsIter

Reads all events received by the window.

This iterator polls for events and can be exhausted.

fn wait_events(&self) -> WaitEventsIter

Reads all events received by the window.

fn is_closed(&self) -> bool

Returns true if the window has been closed.

fn get_window(&self) -> Option<WinRef>

Returns the underlying window, or None if glium uses a headless context.

fn get_opengl_version(&self) -> Version

Returns the OpenGL version of the current context.

fn draw(&self) -> Frame

Start drawing on the backbuffer.

This function returns a Frame, which can be used to draw on it. When the Frame is destroyed, the buffers are swapped.

Note that destroying a Frame is immediate, even if vsync is enabled.