Struct glium::backend::Context

pub struct Context {
    pub framebuffer_objects: Option<FramebuffersContainer>,
    pub vertex_array_objects: VertexAttributesSystem,
    pub samplers: RefCell<HashMap<SamplerBehavior, SamplerObject>>,
    // some fields omitted
}

Stores the state and information required for glium to execute commands. Most public glium functions require passing a Context.

Fields

framebuffer_objects

vertex_array_objects

samplers

Methods

impl Context

unsafe fn new<B>(backend: B, check_current_context: bool) -> Result<Rc<Context>, GliumCreationError> where B: Backend + 'static

Builds a new context.

The check_current_context parameter tells the context whether it should check if the backend's OpenGL context is the current one before each OpenGL operation.

If you pass false, you must ensure that no other OpenGL context is going to be made current in the same thread as this context. Passing true makes things safe but is slightly slower.

fn get_framebuffer_dimensions(&self) -> (u32, u32)

Calls get_framebuffer_dimensions on the backend object stored by this context.

unsafe fn rebuild<B>(&self, new_backend: B) -> Result<(), GliumCreationError> where B: Backend + 'static

Changes the OpenGL context associated with this context.

The new context must have lists shared with the old one.

fn swap_buffers(&self)

Swaps the buffers in the backend.

fn capabilities(&self) -> &Capabilities

fn get_version(&self) -> &Version

Returns the OpenGL version detected by this context.

fn get_extensions(&self) -> &ExtensionsList

fn get_supported_glsl_version(&self) -> Version

Returns the GLSL version guaranteed to be supported.

fn is_glsl_version_supported(&self, version: &Version) -> bool

Returns true if the given GLSL version is supported.

fn get_max_anisotropy_support(&self) -> Option<u16>

Returns the maximum value that can be used for anisotropic filtering, or None if the hardware doesn't support it.

fn get_max_viewport_dimensions(&self) -> (u32, u32)

Returns the maximum dimensions of the viewport.

Glium will panic if you request a larger viewport than this when drawing.

fn release_shader_compiler(&self)

Releases the shader compiler, indicating that no new programs will be created for a while.

This method is a no-op if it's not available in the implementation.

fn get_free_video_memory(&self) -> Option<usize>

Returns an estimate of the amount of video memory available in bytes.

Returns None if no estimate is available.

fn read_front_buffer<P, T>(&self) -> T where P: PixelValue + Clone + Send, T: Texture2dDataSink<Data=P>

Reads the content of the front buffer.

You will only see the data that has finished being drawn.

This function can return any type that implements Texture2dData.

Example

let pixels: Vec<Vec<(u8, u8, u8)>> = display.read_front_buffer();

unsafe fn exec_in_context<'a, T, F>(&self, action: F) -> T where T: Send + 'static, F: FnOnce() -> T + 'a

Execute an arbitrary closure with the OpenGL context active. Useful if another component needs to directly manipulate OpenGL state.

If action manipulates any OpenGL state, it must be restored before action completes.

fn assert_no_error(&self)

Asserts that there are no OpenGL errors pending.

This function should be used in tests.

fn synchronize(&self)

Waits until all the previous commands have finished being executed.

When you execute OpenGL functions, they are not executed immediately. Instead they are put in a queue. This function waits until all commands have finished being executed, and the queue is empty.

You don't need to call this function manually, except when running benchmarks.