Module glium::buffer

Buffers management in glium

There are three levels of abstraction in glium:

A Buffer corresponds to an OpenGL buffer object. This type is not public.

A SubBuffer corresponds to a part of a Buffer. One buffer can contain one or multiple subbuffers.

The VertexBuffer, IndexBuffer, UniformBuffer, PixelBuffer, ... types are abstractions over a subbuffer indicating their specific purpose. They implement Deref for the subbuffer. These types are in the vertex, index, ... modules.

Structs

Builder
Struct that allows you to build multiple subbuffers stored inside the same buffer.

Mapping
Mapping of a buffer in memory.

SubBuffer
Represents a sub-part of a buffer.

SubBufferAny
Represents a sub-part of a buffer.

SubBufferAnySlice
Slice of a SubBuffer without any type info.

SubBufferMutSlice
Represents a sub-part of a buffer.

SubBufferSlice
Represents a sub-part of a buffer.

Enums

BufferCreationError
Error that can happen when creating a buffer.