Module glium::debug

Structs

TimestampQuery
Allows you to obtain the timestamp inside the OpenGL commands queue.

Enums

DebugCallbackBehavior
Describes the behavior that the debug output should have.

MessageType
Type of a debug message.

Severity
Severity of a debug message.

Source
Source of a debug message.

Type Definitions

DebugCallback
Represents a callback that can be used for the debug output feature of OpenGL.