Enum glium::debug::DebugCallbackBehavior

pub enum DebugCallbackBehavior {
    Ignore,
    DebugMessageOnError,
    PrintAll,
    Custom {
        callback: DebugCallback,
        synchronous: bool,
    },
}

Describes the behavior that the debug output should have.

Variants
Ignore

Don't do anything. This is the default behavior in release.

DebugMessageOnError

Print a message on stdout on error, except in some circumstances like when compiling shaders. This is the default behavior in debug mode.

PrintAll

Print every single output received by the driver.

Custom

Use a custom callback.

Fields of Custom

callback: DebugCallback

The function to be called.

synchronous: bool

Whether or not it should be called immediately (true) or asynchronously (false).