Struct glium::draw_parameters::Stencil

pub struct Stencil {
    pub test_clockwise: StencilTest,
    pub reference_value_clockwise: i32,
    pub write_mask_clockwise: u32,
    pub fail_operation_clockwise: StencilOperation,
    pub pass_depth_fail_operation_clockwise: StencilOperation,
    pub depth_pass_operation_clockwise: StencilOperation,
    pub test_counter_clockwise: StencilTest,
    pub reference_value_counter_clockwise: i32,
    pub write_mask_counter_clockwise: u32,
    pub fail_operation_counter_clockwise: StencilOperation,
    pub pass_depth_fail_operation_counter_clockwise: StencilOperation,
    pub depth_pass_operation_counter_clockwise: StencilOperation,
}

Describes the parameters that must be used for the stencil operations when drawing.

Fields
test_clockwise: StencilTest

A comparison against the existing value in the stencil buffer.

Only relevant for faces that are clockwise on the target surface. Other faces, points and lines use stencil_test_counter_clockwise instead.

The default value is AlwaysPass.

reference_value_clockwise: i32

Reference value that is used by stencil_test_clockwise, stencil_fail_operation_clockwise, stencil_pass_depth_fail_operation_clockwise and stencil_depth_pass_operation_clockwise.

write_mask_clockwise: u32

Allows specifying a mask when writing data on the stencil buffer.

Only relevant for faces that are clockwise on the target surface. Other faces, points and lines use stencil_write_mask_counter_clockwise instead.

The default value is 0xffffffff.

fail_operation_clockwise: StencilOperation

Specifies the operation to do when a fragment fails the stencil test.

The stencil test is the test specified by stencil_test_clockwise.

Only relevant for faces that are clockwise on the target surface. Other faces, points and lines use stencil_fail_operation_counter_clockwise instead.

The default value is Keep.

pass_depth_fail_operation_clockwise: StencilOperation

Specifies the operation to do when a fragment passes the stencil test but fails the depth test.

The stencil test is the test specified by stencil_test_clockwise.

Only relevant for faces that are clockwise on the target surface. Other faces, points and lines use stencil_pass_depth_fail_operation_counter_clockwise instead.

The default value is Keep.

depth_pass_operation_clockwise: StencilOperation

Specifies the operation to do when a fragment passes both the stencil and depth tests.

The stencil test is the test specified by stencil_test_clockwise.

Only relevant for faces that are clockwise on the target surface. Other faces, points and lines use stencil_depth_pass_operation_counter_clockwise instead.

The default value is Keep.

test_counter_clockwise: StencilTest

A comparaison against the existing value in the stencil buffer.

Only relevant for points, lines and faces that are counter-clockwise on the target surface. Other faces use stencil_test_counter_clockwise instead.

The default value is AlwaysPass.

reference_value_counter_clockwise: i32

Reference value that is used by stencil_test_counter_clockwise, stencil_fail_operation_counter_clockwise, stencil_pass_depth_fail_operation_counter_clockwise and stencil_depth_pass_operation_counter_clockwise.

write_mask_counter_clockwise: u32

Allows specifying a mask when writing data on the stencil buffer.

Only relevant for points, lines and faces that are counter-clockwise on the target surface. Other faces use stencil_write_mask_clockwise instead.

The default value is 0xffffffff.

fail_operation_counter_clockwise: StencilOperation

Specifies the operation to do when a fragment fails the stencil test.

The stencil test is the test specified by stencil_test_counter_clockwise.

Only relevant for faces that are counter-clockwise on the target surface. Other faces use stencil_fail_operation_clockwise instead.

The default value is Keep.

pass_depth_fail_operation_counter_clockwise: StencilOperation

Specifies the operation to do when a fragment passes the stencil test but fails the depth test.

The stencil test is the test specified by stencil_test_counter_clockwise.

Only relevant for faces that are counter-clockwise on the target surface. Other faces use stencil_pass_depth_fail_operation_clockwise instead.

The default value is Keep.

depth_pass_operation_counter_clockwise: StencilOperation

Specifies the operation to do when a fragment passes both the stencil and depth tests.

The stencil test is the test specified by stencil_test_counter_clockwise.

Only relevant for faces that are counter-clockwise on the target surface. Other faces use stencil_depth_pass_operation_clockwise instead.

The default value is Keep.