Enum glium::LinearBlendingFactorUnstable

pub enum LinearBlendingFactor {
    Zero,
    One,
    SourceColor,
    OneMinusSourceColor,
    DestinationColor,
    OneMinusDestinationColor,
    SourceAlpha,
    OneMinusSourceAlpha,
    DestinationAlpha,
    OneMinusDestinationAlpha,
}

Indicates which value to multiply each component with.

Variants

Zero

Multiply the source or destination component by zero, which always gives 0.0.

One

Multiply the source or destination component by one, which always gives you the original value.

SourceColor

Multiply the source or destination component by its corresponding value in the source.

If you apply this to the source components, you get the values squared.

OneMinusSourceColor

Equivalent to 1 - SourceColor.

DestinationColor

Multiply the source or destination component by its corresponding value in the destination.

If you apply this to the destination components, you get the values squared.

OneMinusDestinationColor

Equivalent to 1 - DestinationColor.

SourceAlpha

Multiply the source or destination component by the alpha value of the source.

OneMinusSourceAlpha

Multiply the source or destination component by 1.0 minus the alpha value of the source.

DestinationAlpha

Multiply the source or destination component by the alpha value of the destination.

OneMinusDestinationAlpha

Multiply the source or destination component by 1.0 minus the alpha value of the destination.