Enum glium::LinearBlendingFactor

pub enum LinearBlendingFactor {
    Zero,
    One,
    SourceColor,
    OneMinusSourceColor,
    DestinationColor,
    OneMinusDestinationColor,
    SourceAlpha,
    SourceAlphaSaturate,
    OneMinusSourceAlpha,
    DestinationAlpha,
    OneMinusDestinationAlpha,
    ConstantColor,
    OneMinusConstantColor,
    ConstantAlpha,
    OneMinusConstantAlpha,
}

Indicates which value to multiply each component with.

Variants
Zero

Multiply the source or destination component by zero, which always gives 0.0.

One

Multiply the source or destination component by one, which always gives you the original value.

SourceColor

Multiply the source or destination component by its corresponding value in the source.

If you apply this to the source components, you get the values squared.

OneMinusSourceColor

Equivalent to 1 - SourceColor.

DestinationColor

Multiply the source or destination component by its corresponding value in the destination.

If you apply this to the destination components, you get the values squared.

OneMinusDestinationColor

Equivalent to 1 - DestinationColor.

SourceAlpha

Multiply the source or destination component by the alpha value of the source.

SourceAlphaSaturate

Multiply the source or destination component by the smallest value of SourceAlpha and 1 - DestinationAlpha.

OneMinusSourceAlpha

Multiply the source or destination component by 1.0 minus the alpha value of the source.

DestinationAlpha

Multiply the source or destination component by the alpha value of the destination.

OneMinusDestinationAlpha

Multiply the source or destination component by 1.0 minus the alpha value of the destination.

ConstantColor

Multiply the source or destination component by the corresponding value in Blend::const_value.

OneMinusConstantColor

Multiply the source or destination compoent by 1.0 minus the corresponding value in Blend::const_value.

ConstantAlpha

Multiply the source or destination component by the alpha value of Blend::const_value.

OneMinusConstantAlpha

Multiply the source or destination componet by 1.0 minus the alpha value of Blend::const_value.