Enum glium::StencilTest

pub enum StencilTest {
    AlwaysPass,
    AlwaysFail,
    IfLess {
        mask: u32,
    },
    IfLessOrEqual {
        mask: u32,
    },
    IfMore {
        mask: u32,
    },
    IfMoreOrEqual {
        mask: u32,
    },
    IfEqual {
        mask: u32,
    },
    IfNotEqual {
        mask: u32,
    },
}

Specifies which comparaison the GPU will do to determine whether a sample passes the stencil test.

Variants

AlwaysPass

The stencil test always passes.

AlwaysFail

The stencil test always fails.

IfLess

Applies the mask as a bitfield to both the value currently in the stencil buffer and the reference value (stencil_reference_value_clockwise or stencil_reference_value_counter_clockwise), then compares them.

Fields

mask

IfLessOrEqual

Applies the mask as a bitfield to both the value currently in the stencil buffer and the reference value (stencil_reference_value_clockwise or stencil_reference_value_counter_clockwise), then compares them.

Fields

mask

IfMore

Applies the mask as a bitfield to both the value currently in the stencil buffer and the reference value (stencil_reference_value_clockwise or stencil_reference_value_counter_clockwise), then compares them.

Fields

mask

IfMoreOrEqual

Applies the mask as a bitfield to both the value currently in the stencil buffer and the reference value (stencil_reference_value_clockwise or stencil_reference_value_counter_clockwise), then compares them.

Fields

mask

IfEqual

Applies the mask as a bitfield to both the value currently in the stencil buffer and the reference value (stencil_reference_value_clockwise or stencil_reference_value_counter_clockwise), then compares them.

Fields

mask

IfNotEqual

Applies the mask as a bitfield to both the value currently in the stencil buffer and the reference value (stencil_reference_value_clockwise or stencil_reference_value_counter_clockwise), then compares them.

Fields

mask