Enum glium::SwapBuffersError

pub enum SwapBuffersError {
    ContextLost,
    AlreadySwapped,
}

Error that can happen when swapping buffers.

Variants
ContextLost

The OpenGL context has been lost and needs to be recreated. The Display and all the objects associated to it (textures, buffers, programs, etc.) need to be recreated from scratch.

Operations will have no effect. Functions that read textures, buffers, etc. from OpenGL will return uninitialized data instead.

A context loss usually happens on mobile devices when the user puts the application on sleep and wakes it up later. However any OpenGL implementation can theoretically lose the context at any time. Can only happen if calling is_context_loss_possible() returns true.

AlreadySwapped

The buffers have already been swapped.

This error can be returned when set_finish() is called multiple times, or finish() is called after set_finish().