Struct glium::framebuffer::DepthStencilRenderBuffer

pub struct DepthStencilRenderBuffer {
    // some fields omitted
}

A render buffer is similar to a texture, but is optimized for usage as a draw target.

Contrary to a texture, you can't sample or modify the content of the DepthStencilRenderBuffer directly.

Methods

impl DepthStencilRenderBuffer

fn new<F>(facade: &F, format: DepthStencilFormat, width: u32, height: u32) -> DepthStencilRenderBuffer where F: Facade

Builds a new render buffer.