Struct glium::framebuffer::EmptyFrameBuffer

pub struct EmptyFrameBuffer { /* fields omitted */ }

A framebuffer with no attachment at all.

Note that this is only supported on recent hardware.

Methods

impl<'a> EmptyFrameBuffer

pub fn is_supported<C: ?Sized>(context: &C) -> bool where
C: CapabilitiesSource,

Returns true if empty framebuffers are supported by the backend.

pub fn is_layered_supported<C: ?Sized>(context: &C) -> bool where
C: CapabilitiesSource,

Returns true if layered empty framebuffers are supported by the backend.

pub fn get_max_supported_width<C: ?Sized>(context: &C) -> Option<u32> where
C: CapabilitiesSource,

Returns the maximum width of empty framebuffers that the backend supports, or None if empty framebuffers are not supported.

pub fn get_max_supported_height<C: ?Sized>(context: &C) -> Option<u32> where
C: CapabilitiesSource,

Returns the maximum height of empty framebuffers that the backend supports, or None if empty framebuffers are not supported.

pub fn get_max_supported_samples<C: ?Sized>(context: &C) -> Option<u32> where
C: CapabilitiesSource,

Returns the maximum number of samples of empty framebuffers that the backend supports, or None if empty framebuffers are not supported.

pub fn get_max_supported_layers<C: ?Sized>(context: &C) -> Option<u32> where
C: CapabilitiesSource,

Returns the maximum number of layers of empty framebuffers that the backend supports, or None if layered empty framebuffers are not supported.

pub fn new<F: ?Sized>(
facade: &F,
width: u32,
height: u32,
layers: Option<u32>,
samples: Option<u32>,
fixed_samples: bool
) -> Result<EmptyFrameBuffer, ValidationError> where
F: Facade,

Creates a EmptyFrameBuffer.

Panic

Panics if layers or samples is equal to Some(0).