Struct glium::framebuffer::SimpleFrameBufferExperimental

pub struct SimpleFrameBuffer<'a> {
    // some fields omitted
}

A framebuffer which has only one color attachment.

Methods

impl<'a> SimpleFrameBuffer<'a>

fn new<C>(display: &Display, color: &'a C) -> SimpleFrameBuffer<'a> where C: ToColorAttachment

Creates a SimpleFrameBuffer with a single color attachment and no depth nor stencil buffer.

fn with_depth_buffer<C, D>(display: &Display, color: &'a C, depth: &'a D) -> SimpleFrameBuffer<'a> where C: ToColorAttachment, D: ToDepthAttachment

Creates a SimpleFrameBuffer with a single color attachment and a depth buffer, but no stencil buffer.

fn with_depth_and_stencil_buffer<C, D, S>(display: &Display, color: &'a C, depth: &'a D, stencil: &'a S) -> SimpleFrameBuffer<'a> where C: ToColorAttachment, D: ToDepthAttachment, S: ToStencilAttachment

Creates a SimpleFrameBuffer with a single color attachment, a depth buffer, and a stencil buffer.

fn with_stencil_buffer<C, S>(display: &Display, color: &'a C, stencil: &'a S) -> SimpleFrameBuffer<'a> where C: ToColorAttachment, S: ToStencilAttachment

Creates a SimpleFrameBuffer with a single color attachment and a stencil buffer, but no buffer buffer.