Module glium::index

In order to draw, you need to provide a source of indices which is used to link the vertices together into primitives.

There are eleven types of primitives, each one with a corresponding struct:

PointsList

LinesList

LinesListAdjacency

LineStrip

LineStripAdjacency

TrianglesList

TrianglesListAdjacency

TriangleStrip

TriangleStripAdjacency

TriangleFan

Patches

These structs can be turned into an IndexBuffer, which uploads the data in video memory.

There are three ways to specify the indices that must be used:

Passing a reference to one of these structs.

Passing a reference to an IndexBuffer.

NoIndices, which is equivalent to (0, 1, 2, 3, 4, 5, 6, 7, ..).

For performances it is highly recommended to use either an IndexBuffer or NoIndices, and to avoid passing indices in RAM.

When you draw something, a draw command is sent to the GPU and the execution continues immediatly after. But if you pass indices in RAM, the execution has to block until the GPU has finished drawing in order to make sure that the indices are not free'd.

Reexports

pub use self::buffer::IndexBuffer;

Structs

LineStrip
A list of lines connected together stored in RAM.

LineStripAdjacency
A list of lines connected together, with adjacency information, stored in RAM.

LinesList
A list of lines stored in RAM.

LinesListAdjacency
A list of lines, with adjacency information, stored in RAM.

NoIndices
Marker that can be used as an indices source when you don't need indices.

Patches
A list of patches stored in RAM.

PointsList
A list of points stored in RAM.

TriangleFan
A list of triangles stored in RAM.

TriangleStrip
A list of triangles connected together stored in RAM.

TriangleStripAdjacency
A list of triangles connected together, with adjacency information, stored in RAM.

TrianglesList
A list of triangles stored in RAM.

TrianglesListAdjacency
A list of triangles, with adjacency information, stored in RAM.

Enums

IndexType
Type of the indices in an index source.

IndicesSource
Describes a source of indices used for drawing.

PrimitiveType
List of available primitives.

Traits

Index
An index from the index buffer.

IntoIndexBuffer
Object that is convertible to an index buffer.

ToIndicesSource
Can be used as a source of indices when drawing.