Enum glium::index::IndicesSource

pub enum IndicesSource<'a, T: 'a> {
    IndexBuffer {
        buffer: &'a IndexBuffer,
        offset: usize,
        length: usize,
    },
    Buffer {
        pointer: &'a [T],
        primitives: PrimitiveType,
        offset: usize,
        length: usize,
    },
    NoIndices {
        primitives: PrimitiveType,
    },
}

Describes a source of indices used for drawing.

Variants

IndexBuffer

A buffer uploaded in video memory.

Fields

buffer

The buffer.

offset

Offset of the first element of the buffer to use.

length

Number of elements in the buffer to use.

Buffer

A buffer in RAM.

Fields

pointer

Slice of data to use.

primitives

Type of primitives contained in the buffer.

offset

Offset of the first element of the buffer to use.

length

Number of elements in the buffer to use.

NoIndices

Don't use indices. Assemble primitives by using the order in which the vertices are in the vertices source.

Fields

primitives

Type of primitives contained in the vertex source.

Methods

impl<'a, T> IndicesSource<'a, T> where T: Index

fn get_primitives_type(&self) -> PrimitiveType

Returns the type of the primitives.