Struct glium::pixel_buffer::PixelBufferUnstable

pub struct PixelBuffer<T> {
    // some fields omitted
}

Buffer that stores the content of a texture.

The generic type represents the type of pixels that the buffer contains.

Methods

impl<T> PixelBuffer<T> where T: Texture2dData

fn new_empty(display: &Display, capacity: usize) -> PixelBuffer<T>

Builds a new buffer with an uninitialized content.

fn get_size(&self) -> usize

Returns the size in bytes of the buffer.

fn read(&self) -> T

Copies the content of the pixel buffer to the RAM.

This operation is slow and should be done outside of the rendering loop.

Panic

Panics if the pixel buffer is empty.