Module glium::program

Reexports

pub use self::program::{Program, ProgramCreationError};

Structs

Attribute
Information about an attribute of a program (except its name).

Binary
Represents the compiled binary data of a program.

SourceCode
Represents the source code of a program.

TransformFeedbackVarying
Describes a varying that is being output with transform feedback.

Uniform
Information about a uniform (except its name).

UniformBlock
Information about a uniform block (except its name).

UniformBlockMember
Information about a uniform inside a block.

Enums

ProgramCreationInput
Input when creating a program.

TransformFeedbackMode
Describes the mode that is used when transform feedback is enabled.

Traits

IntoProgramCreationInput
Traits for objects that can be turned into ProgramCreationInput.