Enum glium::program::BlockLayout

pub enum BlockLayout {
    Struct {
        members: Vec<(String, BlockLayout)>,
    },
    BasicType {
        ty: UniformType,
        offset_in_buffer: usize,
    },
    Array {
        content: Box<BlockLayout>,
        length: usize,
    },
    DynamicSizedArray {
        content: Box<BlockLayout>,
    },
}

Layout of a shader storage buffer or a uniform buffer.

Variants
Struct

Multiple elements, each having a name.

Fields

members: Vec<(String, BlockLayout)>

The list of elements, with name/layout pairs.

BasicType

A basic element.

Fields

ty: UniformType

Type of data.

offset_in_buffer: usize

Offset of this element in bytes from the start of the buffer.

Array

A fixed-size array.

For example:

uint data[12];

Fields

content: Box<BlockLayout>

Type of data of each element.

length: usize

Number of elements in the array.

DynamicSizedArray

An array whose size isn't known at compile-time. Can only be used as the last element of a buffer.

Its actual size depends on the size of the buffer.

For example:

buffer MyBuffer {
    uint data[];
}

Fields

content: Box<BlockLayout>

Type of data of each element.