Struct glium::program::ProgramUnstable

pub struct Program {
    // some fields omitted
}

A combination of shaders linked together.

Methods

impl Program

fn new<'a, I>(display: &Display, input: I) -> Result<Program, ProgramCreationError> where I: IntoProgramCreationInput<'a>

Builds a new program.

fn from_source<'a>(display: &Display, vertex_shader: &'a str, fragment_shader: &'a str, geometry_shader: Option<&'a str>) -> Result<Program, ProgramCreationError>

Builds a new program from GLSL source code.

A program is a group of shaders linked together.

Parameters

vertex_shader: Source code of the vertex shader.

fragment_shader: Source code of the fragment shader.

geometry_shader: Source code of the geometry shader.

Example

let program = glium::Program::from_source(&display, vertex_source, fragment_source,
    Some(geometry_source));

fn get_binary(&self) -> Binary

Returns the program's compiled binary.

You can store the result in a file, then reload it later. This avoids having to compile the source code every time.

Features

Only available if the gl_program_binary feature is enabled.

fn get_binary_if_supported(&self) -> Option<Binary>

Returns the program's compiled binary.

Same as get_binary but always available. Returns None if the backend doesn't support getting or reloading the program's binary.

fn get_frag_data_location(&self, name: &str) -> Option<u32>

Returns the location of an output fragment, if it exists.

The location is low-level information that is used internally by glium. You probably don't need to call this function.

You can declare output fragments in your shaders by writing:

out vec4 foo;

fn get_uniform(&self, name: &str) -> Option<&Uniform>

Returns informations about a uniform variable, if it exists.

fn get_uniform_blocks(&self) -> &HashMap<String, UniformBlock>

Returns a list of uniform blocks.

fn has_tessellation_shaders(&self) -> bool

Returns true if the program contains a tessellation stage.