Struct glium::DisplayUnstable

pub struct Display {
    // some fields omitted
}

The main object of this library. Controls the whole display.

This object contains a smart pointer to the real implementation. Cloning the display allows you to easily share the Display object throughout your program and between threads.

Methods

impl Display

fn poll_events(&self) -> Vec<Event>

Reads all events received by the window.

fn get_framebuffer_dimensions(&self) -> (usize, usize)

Returns the dimensions of the main framebuffer.

fn draw(&self) -> Frame

Start drawing on the backbuffer.

This function returns a Frame which can be used to draw on it. When the Frame is destroyed, the buffers are swapped.

Note that destroying a Frame is immediate, even if vsync is enabled.

fn get_max_anisotropy_support(&self) -> Option<u16>

Returns the maximum value that can be used for anisotropic filtering, or None if the hardware doesn't support it.

fn get_max_viewport_dimensions(&self) -> (u32, u32)

Returns the maximum dimensions of the viewport that you can pass when drawing.

Glium will panic if you request a larger viewport.

fn release_shader_compiler(&self)

Releases the shader compiler, indicating that no new programs will be created for a while.

Features

This method is always available, but is a no-op if it's not available in the implementation.

fn get_free_video_memory(&self) -> Option<usize>

Returns an estimate of the amount of video memory available in bytes.

Returns None if no estimate is available.

fn set_debug_callback<F>(&self, callback: F) where F: FnMut(String, Source, MessageType, Severity) + Send + Sync

Sets the callback to use when an OpenGL debug message is generated.

Important: some contexts don't support debug output, in which case this function will act as a no-op. Even if the context does support them, you are not guaranteed to get any. Debug messages are just a convenience and are not reliable.

unsafe fn set_debug_callback_sync<F>(&self, callback: F) where F: FnMut(String, Source, MessageType, Severity) + Send + Sync

Sets the callback to use when an OpenGL debug message is generated.

Contrary to set_debug_callback, the callback is called synchronously.

fn read_front_buffer<P, T>(&self) -> T where P: PixelValue + Clone + Send, T: Texture2dData<Data=P>

Reads the content of the front buffer.

You will only see the data that has finished being drawn.

This function can return any type that implements Texture2dData.

Example

# extern crate glium;
# extern crate glutin;
# use glium::DisplayBuild;
# fn main() {
# let display: glium::Display = unsafe { ::std::mem::uninitialized() };
let pixels: Vec<Vec<(u8, u8, u8)>> = display.read_front_buffer();
# }

fn assert_no_error(&self)

Asserts that there are no OpenGL error pending.

This function is supposed to be used in tests.

fn synchronize(&self)

Waits until all the previous commands have finished being executed.

When you execute OpenGL functions, they are not executed immediatly. Instead they are put in a queue. This function waits until all commands have finished being executed and the queue is empty.

You don't need to call this function manually, except when running benchmarks.