Struct glium::DrawParametersUnstable

pub struct DrawParameters {
    pub depth_function: DepthFunction,
    pub depth_range: (f32, f32),
    pub blending_function: Option<BlendingFunction>,
    pub line_width: Option<f32>,
    pub backface_culling: BackfaceCullingMode,
    pub polygon_mode: PolygonMode,
    pub multisampling: bool,
    pub dithering: bool,
    pub viewport: Option<Rect>,
    pub scissor: Option<Rect>,
}

Represents the parameters to use when drawing.

Example:

let params = glium::DrawParameters {
    depth_function: glium::DepthFunction::IfLess,
    .. std::default::Default::default()
};

Fields

depth_function

The function that the GPU will use to determine whether to write over an existing pixel on the target.

See the documentation of DepthFunction for more details.

The default is Overwrite.

depth_range

The range of Z coordinates in surface coordinates.

Just like OpenGL turns X and Y coordinates between -1.0 and 1.0 into surface coordinates, it will also map your Z coordinates to a certain range which you can specify here.

The two values must be between 0.0 and 1.0, anything outside this range will result in a panic. By default the depth range is (0.0, 1.0).

The first value of the tuple must be the "near" value, where -1.0 will be mapped. The second value must be the "far" value, where 1.0 will be mapped. It is possible for the "near" value to be greater than the "far" value.

blending_function

The function that the GPU will use to merge the existing pixel with the pixel that is being written.

None means "don't care" (usually when you know that the alpha is always 1).

line_width

Width in pixels of the lines to draw when drawing lines.

None means "don't care". Use this when you don't draw lines.

backface_culling

Whether or not the GPU should filter out some faces.

After the vertex shader stage, the GPU will try to remove the faces that aren't facing the camera.

See the documentation of BackfaceCullingMode for more infos.

polygon_mode

Sets how to render polygons. The default value is Fill.

See the documentation of PolygonMode for more infos.

multisampling

Whether multisample antialiasing (MSAA) should be used. Default value is true.

Note that you will need to set the appropriate option when creating the window. The recommended way to do is to leave this to true, and adjust the option when creating the window.

dithering

Whether dithering is activated. Default value is true.

Dithering will smoothen the transition between colors in your color buffer.

viewport

Specifies the viewport to use when drawing.

The x and y positions of your vertices are mapped to the viewport so that (-1, -1) corresponds to the lower-left hand corner and (1, 1) corresponds to the top-right hand corner. Any pixel outside of the viewport is discarded.

You can specify a viewport greater than the target if you want to stretch the image.

None means "use the whole surface".

scissor

If specified, only pixels in this rect will be displayed. Default is None.

This is different from a viewport. The image will stretch to fill the viewport, but not the scissor box.