Struct glium::IndexBuffer

pub struct IndexBuffer<T> where T: Index {
    // some fields omitted
}

A list of indices loaded in the graphics card's memory.

Methods

impl<T> IndexBuffer<T> where T: Index

fn new<F, D>(facade: &F, prim: PrimitiveType, data: D) -> IndexBuffer<T> where F: Facade, D: AsRef<[T]>

Builds a new index buffer from a list of indices and a primitive type.

fn dynamic<F, D>(facade: &F, prim: PrimitiveType, data: D) -> IndexBuffer<T> where F: Facade, D: AsRef<[T]>

Builds a new index buffer from a list of indices and a primitive type.

fn empty<F>(facade: &F, prim: PrimitiveType, len: usize) -> IndexBuffer<T> where F: Facade

Builds a new empty index buffer.

fn empty_dynamic<F>(facade: &F, prim: PrimitiveType, len: usize) -> IndexBuffer<T> where F: Facade

Builds a new empty index buffer.

fn get_primitives_type(&self) -> PrimitiveType

Returns the type of primitives associated with this index buffer.

fn get_indices_type(&self) -> IndexType

Returns the data type of the indices inside this index buffer.

fn slice(&self, range: Range<usize>) -> Option<IndexBufferSlice<T>>

Returns None if out of range.