Struct glium::ProgramUnstable

pub struct Program {
    // some fields omitted
}

A combinaison of shaders linked together.

Methods

impl Program

fn new(display: &Display, vertex_shader: &str, fragment_shader: &str, geometry_shader: Option<&str>) -> Result<Program, ProgramCreationError>

Builds a new program.

fn from_source(display: &Display, vertex_shader: &str, fragment_shader: &str, geometry_shader: Option<&str>) -> Result<Program, ProgramCreationError>

Builds a new program from GLSL source code.

A program is a group of shaders linked together.

Parameters

vertex_shader: Source code of the vertex shader.

fragment_shader: Source code of the fragment shader.

geometry_shader: Source code of the geometry shader.

Example

let program = glium::Program::from_source(&display, vertex_source, fragment_source,
    Some(geometry_source));

fn get_frag_data_location(&self, name: &str) -> Option<u32>

Returns the location of an output fragment, if it exists.

The location is a low-level information that is used internally by glium. You probably don't need to call this function.

You can declare output fragments in your shaders by writing:

out vec4 foo;