Struct glium::SyncFence

pub struct SyncFence {
    // some fields omitted
}

Provides a way to wait for a server-side operation to be finished.

Creating a SyncFence injects an element in the commands queue of the backend. When this element is reached, the fence becomes signaled.

Example

let fence = glium::SyncFence::new_if_supported(&display).unwrap();
do_something(&display);
fence.wait();   // blocks until the previous operations have finished

Methods

impl SyncFence

fn new<F>(facade: &F) -> SyncFence where F: Facade

Builds a new SyncFence that is injected in the server.

Features

Only available if the gl_sync feature is enabled.

fn new_if_supported<F>(facade: &F) -> Option<SyncFence> where F: Facade

Builds a new SyncFence that is injected in the server.

Returns None is this is not supported by the backend.

fn wait(self)

Blocks until the operation has finished on the server.