Struct glium::VertexBuffer

pub struct VertexBuffer<T> where T: Copy {
    // some fields omitted
}

A list of vertices loaded in the graphics card's memory.

Methods

impl<T> VertexBuffer<T> where T: Vertex

fn new<F>(facade: &F, data: &[T]) -> Result<VertexBuffer<T>, CreationError> where F: Facade

Builds a new vertex buffer.

Note that operations such as write will be very slow. If you want to modify the buffer from time to time, you should use the dynamic function instead.

Example

#[derive(Copy, Clone)]
struct Vertex {
    position: [f32; 3],
    texcoords: [f32; 2],
}

implement_vertex!(Vertex, position, texcoords);

let vertex_buffer = glium::VertexBuffer::new(&display, &[
    Vertex { position: [0.0,  0.0, 0.0], texcoords: [0.0, 1.0] },
    Vertex { position: [5.0, -3.0, 2.0], texcoords: [1.0, 0.0] },
]);

fn dynamic<F>(facade: &F, data: &[T]) -> Result<VertexBuffer<T>, CreationError> where F: Facade

Builds a new vertex buffer.

This function will create a buffer that is intended to be modified frequently.

fn persistent<F>(facade: &F, data: &[T]) -> Result<VertexBuffer<T>, CreationError> where F: Facade

Builds a new vertex buffer.

fn immutable<F>(facade: &F, data: &[T]) -> Result<VertexBuffer<T>, CreationError> where F: Facade

Builds a new vertex buffer.

fn empty<F>(facade: &F, elements: usize) -> Result<VertexBuffer<T>, CreationError> where F: Facade

Builds an empty vertex buffer.

The parameter indicates the number of elements.

fn empty_dynamic<F>(facade: &F, elements: usize) -> Result<VertexBuffer<T>, CreationError> where F: Facade

Builds an empty vertex buffer.

The parameter indicates the number of elements.

fn empty_persistent<F>(facade: &F, elements: usize) -> Result<VertexBuffer<T>, CreationError> where F: Facade

Builds an empty vertex buffer.

The parameter indicates the number of elements.

fn empty_immutable<F>(facade: &F, elements: usize) -> Result<VertexBuffer<T>, CreationError> where F: Facade

Builds an empty vertex buffer.

The parameter indicates the number of elements.

impl<T> VertexBuffer<T> where T: Copy

unsafe fn new_raw<F>(facade: &F, data: &[T], bindings: VertexFormat, elements_size: usize) -> Result<VertexBuffer<T>, CreationError> where F: Facade

Builds a new vertex buffer from an indeterminate data type and bindings.

Example

use std::borrow::Cow;

let bindings = Cow::Owned(vec![(
        Cow::Borrowed("position"), 0,
        glium::vertex::AttributeType::F32F32,
    ), (
        Cow::Borrowed("color"), 2 * ::std::mem::size_of::<f32>(),
        glium::vertex::AttributeType::F32,
    ),
]);

let data = vec![
    1.0, -0.3, 409.0,
    -0.4, 2.8, 715.0f32
];

let vertex_buffer = unsafe {
    glium::VertexBuffer::new_raw(&display, &data, bindings, 3 * ::std::mem::size_of::<f32>())
};

unsafe fn new_raw_dynamic<F>(facade: &F, data: &[T], bindings: VertexFormat, elements_size: usize) -> Result<VertexBuffer<T>, CreationError> where F: Facade

Dynamic version of new_raw.

fn slice(&self, range: Range<usize>) -> Option<VertexBufferSlice<T>>

Accesses a slice of the buffer.

Returns None if the slice is out of range.

fn get_bindings(&self) -> &VertexFormat

Returns the associated VertexFormat.

fn per_instance(&self) -> Result<PerInstance, InstancingNotSupported>

Creates a marker that instructs glium to use multiple instances.

Instead of calling surface.draw(&vertex_buffer, ...) you can call surface.draw(vertex_buffer.per_instance(), ...). This will draw one instance of the geometry for each element in this buffer. The attributes are still passed to the vertex shader, but each entry is passed for each different instance.

impl<T> VertexBuffer<T> where T: Copy + Send + 'static

fn into_vertex_buffer_any(self) -> VertexBufferAny

DEPRECATED: use .into() instead. Discard the type information and turn the vertex buffer into a VertexBufferAny.