Module glium::textureUnstable

A texture is an image loaded in video memory, which can be sampled in your shaders.

Textures come in ten different dimensions:

Textures with one dimension.

Textures with two dimensions.

Textures with two dimensions and multisampling enabled.

Textures with three dimensions.

Cube textures, which are arrays of six two-dimensional textures corresponding to the six faces of a cube.

Arrays of one-dimensional textures.

Arrays of two-dimensional textures.

Arrays of two-dimensional textures with multisampling enabled.

Arrays of cube textures.

Buffer textures, which are one-dimensional textures that are mapped to a buffer.

In addition to this, there are six kinds of texture formats:

The texture contains floating-point data, with either the Compressed prefix or no prefix at all.

The texture contains signed integers, with the Integral prefix.

The texture contains unsigned integers, with the Unsigned prefix.

The texture contains depth information, with the Depth prefix.

The texture contains stencil information, with the Stencil prefix.

The texture contains depth and stencil information, with the DepthStencil prefix.

Each combination of dimensions and format corresponds to a sampler type in GLSL. For example, an IntegralTexture3d can only be bound to an isampler3D uniform in GLSL. Some combinations don't exist, like DepthBufferTexture.

The difference between compressed textures and uncompressed textures is that you can't do render-to-texture on the former.

The most common types of textures are CompressedTexture2d and Texture2d (the two dimensions being the width and height). These are what you will use most of the time.

Structs

CompressedTexture1d
A compressed one-dimensional texture containing floating-point data.

CompressedTexture1dArray
An array of compressed one-dimensional textures containing floating-point data.

CompressedTexture2d
A compressed two-dimensional texture containing floating-point data.

CompressedTexture2dArray
An array of compressed two-dimensional textures containing floating-point data.

CompressedTexture3d
A compressed three-dimensional texture containing floating-point data.

DepthStencilTexture1d
A one-dimensional texture containing both depth and stencil data.

DepthStencilTexture1dArray
An array of one-dimensional textures containing both depth and stencil data.

DepthStencilTexture2d
A two-dimensional texture containing both depth and stencil data.

DepthStencilTexture2dArray
An array of two-dimensional textures containing both depth and stencil data.

DepthStencilTexture3d
A three-dimensional texture containing both depth and stencil data.

DepthTexture1d
A one-dimensional texture containing depth data.

DepthTexture1dArray
An array of one-dimensional textures containing depth data.

DepthTexture2d
A two-dimensional texture containing depth data.

DepthTexture2dArray
An array of two-dimensional textures containing depth data.

DepthTexture3d
A three-dimensional texture containing depth data.

IntegralTexture1d
A one-dimensional texture containing signed integral data.

IntegralTexture1dArray
An array of one-dimensional textures containing signed integral data.

IntegralTexture2d
A two-dimensional texture containing signed integral data.

IntegralTexture2dArray
An array of two-dimensional textures containing signed integral data.

IntegralTexture3d
A three-dimensional texture containing signed integral data.

StencilTexture1d
A one-dimensional texture containing stencil data.

StencilTexture1dArray
An array of one-dimensional textures containing stencil data.

StencilTexture2d
A two-dimensional texture containing stencil data.

StencilTexture2dArray
An array of two-dimensional textures containing stencil data.

StencilTexture3d
A three-dimensional texture containing stencil data.

Texture1d
A one-dimensional texture containing floating-point data.

Texture1dArray
An array of one-dimensional textures containing floating-point data.

Texture2d
A two-dimensional texture containing floating-point data.

Texture2dArray
An array of two-dimensional textures containing floating-point data.

Texture3d
A three-dimensional texture containing floating-point data.

TextureSurface
Struct that allows you to draw on a texture.

UnsignedTexture1d
A one-dimensional texture containing unsigned integral data.

UnsignedTexture1dArray
An array of one-dimensional textures containing unsigned integral data.

UnsignedTexture2d
A two-dimensional texture containing unsigned integral data.

UnsignedTexture2dArray
An array of two-dimensional textures containing unsigned integral data.

UnsignedTexture3d
A three-dimensional texture containing unsigned integral data.

Enums

ClientFormat
List of client-side pixel formats.

CompressedFormat
List of compressed texture formats.

DepthFormat
List of formats available for depth textures.

DepthStencilFormat
List of formats available for depth-stencil textures.

StencilFormat
List of formats available for stencil textures.

TextureFormat
Format of the internal representation of a texture.

UncompressedFloatFormat
List of uncompressed pixel formats that contain floating-point-like data.

UncompressedIntFormat
List of uncompressed pixel formats that contain signed integral data.

UncompressedUintFormat
List of uncompressed pixel formats that contain unsigned integral data.

Traits

PixelValue
A trait that must be implemented for any type that can represent the value of a pixel.

Texture
Trait that describes a texture.

Texture1dData
Trait that describes data for a one-dimensional texture.

Texture2dData
Trait that describes data for a two-dimensional texture.

Texture3dData
Trait that describes data for a three-dimensional texture.