Trait glium::texture::buffer_texture::TextureBufferContent

pub unsafe trait TextureBufferContent: BufferContent {
    fn get_type() -> TextureBufferContentType;
}

Trait for data types that can be interpreted by a buffer texture.

Required Methods

fn get_type() -> TextureBufferContentType

Returns the enumeration corresponding to elements of this data type.