Enum glium::texture::CompressedFormat

pub enum CompressedFormat {
    RgtcFormatU,
    RgtcFormatI,
    RgtcFormatUU,
    RgtcFormatII,
    BptcUnorm4,
    BptcSignedFloat3,
    BptcUnsignedFloat3,
    S3tcDxt1NoAlpha,
    S3tcDxt1Alpha,
    S3tcDxt3Alpha,
    S3tcDxt5Alpha,
}

List of compressed texture formats.

Variants
RgtcFormatU

Red/green compressed texture with one unsigned component.

RgtcFormatI

Red/green compressed texture with one signed component.

RgtcFormatUU

Red/green compressed texture with two unsigned components.

RgtcFormatII

Red/green compressed texture with two signed components.

BptcUnorm4

BPTC format with four components represented as integers.

BptcSignedFloat3

BPTC format with three components (no alpha) represented as signed floats.

BptcUnsignedFloat3

BPTC format with three components (no alpha) represented as unsigned floats.

S3tcDxt1NoAlpha

S3TC DXT1 without alpha, see https://www.opengl.org/wiki/S3_Texture_Compression.

S3tcDxt1Alpha

S3TC DXT1 with 1-bit alpha, see https://www.opengl.org/wiki/S3_Texture_Compression.

S3tcDxt3Alpha

S3TC DXT3, see https://www.opengl.org/wiki/S3_Texture_Compression.

S3tcDxt5Alpha

S3TC DXT5, see https://www.opengl.org/wiki/S3_Texture_Compression.

Methods

impl CompressedFormat

pub fn get_formats_list() -> Vec<CompressedFormat>

Returns a list of all the possible values of this enumeration.

pub fn to_texture_format(self) -> TextureFormat

Turns this format into a more generic TextureFormat.

pub fn is_supported<C: ?Sized>(&self, context: &C) -> bool where
C: CapabilitiesSource,

Returns true if this format is supported by the backend.