Enum glium::texture::UncompressedFloatFormatUnstable

pub enum UncompressedFloatFormat {
    U8,
    I8,
    U16,
    I16,
    U8U8,
    I8I8,
    U16U16,
    I16I16,
    U3U32U,
    U4U4U4,
    U5U5U5,
    U8U8U8,
    I8I8I8,
    U10U10U10,
    U12U12U12,
    I16I16I16,
    U2U2U2U2,
    U4U4U4U4,
    U5U5U5U1,
    U8U8U8U8,
    I8I8I8I8,
    U10U10U10U2,
    U12U12U12U12,
    U16U16U16U16,
    F16,
    F16F16,
    F16F16F16,
    F16F16F16F16,
    F32,
    F32F32,
    F32F32F32,
    F32F32F32F32,
    F11F11F10,
    F9F9F9,
}

List of uncompressed pixel formats that contain floating points-like data.

Some formats are marked as "guaranteed to be supported". What this means is that you are certain that the backend will use exactly these formats. If you try to use a format that is not supported by the backend, it will automatically fall back to a larger format.

Variants

U8

Guaranteed to be supported for both textures and renderbuffers.

I8

Guaranteed to be supported for textures.

U16

Guaranteed to be supported for both textures and renderbuffers.

I16

Guaranteed to be supported for textures.

U8U8

Guaranteed to be supported for both textures and renderbuffers.

I8I8

Guaranteed to be supported for textures.

U16U16

Guaranteed to be supported for both textures and renderbuffers.

I16I16

Guaranteed to be supported for textures.

U3U32U

U4U4U4

U5U5U5

U8U8U8

Guaranteed to be supported for textures.

I8I8I8

Guaranteed to be supported for textures.

U10U10U10

U12U12U12

I16I16I16

Guaranteed to be supported for textures.

U2U2U2U2

U4U4U4U4

U5U5U5U1

U8U8U8U8

Guaranteed to be supported for both textures and renderbuffers.

I8I8I8I8

Guaranteed to be supported for textures.

U10U10U10U2

Guaranteed to be supported for both textures and renderbuffers.

U12U12U12U12

U16U16U16U16

Guaranteed to be supported for both textures and renderbuffers.

F16

Guaranteed to be supported for both textures and renderbuffers.

F16F16

Guaranteed to be supported for both textures and renderbuffers.

F16F16F16

Guaranteed to be supported for textures.

F16F16F16F16

Guaranteed to be supported for both textures and renderbuffers.

F32

Guaranteed to be supported for both textures and renderbuffers.

F32F32

Guaranteed to be supported for both textures and renderbuffers.

F32F32F32

Guaranteed to be supported for textures.

F32F32F32F32

Guaranteed to be supported for both textures and renderbuffers.

F11F11F10

Guaranteed to be supported for both textures and renderbuffers.

F9F9F9

Uses three components of 9 bits of precision that all share the same exponent.

Use this format only if all the components are approximately equal.

Guaranteed to be supported for textures.