Enum glium::texture::UncompressedIntFormat

pub enum UncompressedIntFormat {
    I8,
    I16,
    I32,
    I8I8,
    I16I16,
    I32I32,
    I8I8I8,
    I16I16I16,
    I32I32I32,
    I8I8I8I8,
    I16I16I16I16,
    I32I32I32I32,
}

List of uncompressed pixel formats that contain signed integral data.

Variants

I8

I16

I32

I8I8

I16I16

I32I32

I8I8I8

I16I16I16

May not be supported by renderbuffers.

I32I32I32

May not be supported by renderbuffers.

I8I8I8I8

May not be supported by renderbuffers.

I16I16I16I16

I32I32I32I32

Methods

impl UncompressedIntFormat

fn get_formats_list() -> Vec<UncompressedIntFormat>

Returns a list of all the possible values of this enumeration.

fn to_texture_format(self) -> TextureFormat

Turns this format into a more generic TextureFormat.

fn is_supported<C>(&self, context: &C) -> bool where C: CapabilitiesSource

Returns true if this format is supported by the backend.

fn is_color_renderable<C>(&self, context: &C) -> bool where C: CapabilitiesSource

Returns true if a texture or renderbuffer with this format can be used as a framebuffer attachment.