Module glium::texture::pixel_buffer

Pixel buffers are buffers that contain two-dimensional texture data.

Contrary to textures, pixel buffers are stored in a client-defined format. They are used to transfer data to or from the video memory, before or after being turned into a texture.

Structs

PixelBuffer
Buffer that stores the content of a texture.