Struct glium::texture::TextureAny

pub struct TextureAny {
    // some fields omitted
}

A texture whose type isn't fixed at compile-time.

Methods

impl TextureAny

fn get_width(&self) -> u32

Returns the width of the texture.

fn get_height(&self) -> Option<u32>

Returns the height of the texture.

fn get_depth(&self) -> Option<u32>

Returns the depth of the texture.

fn get_array_size(&self) -> Option<u32>

Returns the array size of the texture.

fn get_mipmap_levels(&self) -> u32

Returns the number of mipmap levels of the texture.

fn get_texture_type(&self) -> Dimensions

Returns the type of the texture (1D, 2D, 3D, etc.).

fn get_internal_format(&self) -> Result<InternalFormat, GetFormatError>

Determines the internal format of this texture.

fn mipmap(&self, layer: u32, level: u32) -> Option<TextureAnyMipmap>

Returns a structure that represents a specific mipmap of the texture.

Returns None if out of range.