Struct glium::texture::TextureAnyImage

pub struct TextureAnyImage<'a> { /* fields omitted */ }

Represents a specific 2D image of a texture. 1D textures are considered as having a height of 1.

Methods

impl<'a> TextureAnyImage<'a>

pub fn get_texture(&self) -> &'a TextureAny

Returns the texture.

pub fn get_level(&self) -> u32

Returns the level of the texture.

pub fn get_layer(&self) -> u32

Returns the layer of the texture.

pub fn get_cubemap_layer(&self) -> Option<CubeLayer>

Returns the cubemap layer of this image, or None if the texture is not a cubemap.

pub fn get_width(&self) -> u32

Returns the width of this texture slice.

pub fn get_height(&self) -> Option<u32>

Returns the height of this texture slice.

pub fn get_samples(&self) -> Option<u32>

Returns the number of samples of the texture.

pub fn raw_read<T, P>(&self, rect: &Rect) -> T where
T: Texture2dDataSink<P>,
P: PixelValue,

Reads the content of the image.

Panic

Panics if the rect is out of range.

Panics if it fails to read the texture.

pub fn raw_read_to_pixel_buffer<P>(&self, rect: &Rect, dest: &PixelBuffer<P>) where
P: PixelValue,

Reads the content of the image to a pixel buffer.

Panic

Panics if the rect is out of range.

Panics if the buffer is not large enough.

Panics if it fails to read the texture.

pub fn raw_clear_buffer<D>(&self, data: D) where
D: Into<ClearBufferData>,

Clears the content of the texture to a specific value.

Panic

Panics if data does not match the kind of texture. For example passing a [i32; 4] when using a regular (float) texture.