Trait glium::texture::PixelValue

pub unsafe trait PixelValue: Copy + Clone + Send + 'static {
    fn get_format() -> ClientFormat;
}

A trait that must be implemented for any type that can represent the value of a pixel.

Required methods

fn get_format() -> ClientFormat

Returns corresponding client format.

Loading content...