Trait glium::texture::PixelValueExperimental

pub trait PixelValue: Copy + Send {
    fn get_format(_: Option<Self>) -> ClientFormat;
}

A trait that must be implemented for any type that can represent the value of a pixel.

Required Methods

fn get_format(_: Option<Self>) -> ClientFormat

Returns corresponding client format.