Trait glium::texture::TextureUnstable

pub trait Texture {
    fn get_width(&self) -> u32;
    fn get_height(&self) -> Option<u32>;
    fn get_depth(&self) -> Option<u32>;
    fn get_array_size(&self) -> Option<u32>;
}

Trait that describes a texture.

Required Methods

fn get_width(&self) -> u32

Returns the width in pixels of the texture.

fn get_height(&self) -> Option<u32>

Returns the height in pixels of the texture, or None for one dimension textures.

fn get_depth(&self) -> Option<u32>

Returns the depth in pixels of the texture, or None for one or two dimension textures.

fn get_array_size(&self) -> Option<u32>

Returns the number of textures in the array, or None for non-arrays.