Trait glium::texture::Texture2dDataSink

pub trait Texture2dDataSink<T> {
    fn from_raw(data: Cow<[T]>, width: u32, height: u32) -> Self
    where
        [T]: ToOwned;
}

Trait that describes types that can be built from two-dimensional texture data.

The parameter indicates the type of pixels accepted by this sink.

You are especially encouraged to implement this trait with the parameter (u8, u8, u8, u8), as this is the only format that is guaranteed to be supported by OpenGL when reading pixels.

Required methods

fn from_raw(data: Cow<[T]>, width: u32, height: u32) -> Self where
[T]: ToOwned,

Builds a new object from raw data.

Loading content...