Trait glium::DisplayBuildUnstable

pub trait DisplayBuild {
    fn build_glium(self) -> Result<Display, GliumCreationError>;
}

Objects that can build a Display object.

Required Methods

fn build_glium(self) -> Result<Display, GliumCreationError>

Build a context and a Display to draw on it.

Performances a compatibility check to make sure that all core elements of glium are supported by the implementation.