Module glium::uniformsUnstable

A uniform is a global variable in your program. In order to draw something, you will need to tell glium what the values of all your uniforms are. Objects that implement the Uniform trait are here to do that.

The recommended way to is to create your own structure and put the #[uniforms] attribute to it.

For example:

#[plugin]
extern crate glium_macros;

#[uniforms]
struct Uniforms<'a> {
    texture: &'a glium::Texture2d,
    matrix: [[f32; 4]; 4],
}

let uniforms = Uniforms {
    texture: tex,
    matrix: matrix,
};

Each field must implement the UniformValue trait.

Sampler

In order to customize the way a texture is being sampled, you must use a Sampler.

use std::default::Default;
let uniforms = glium::uniforms::UniformsStorage::new("texture",
    glium::uniforms::Sampler(&texture, glium::uniforms::SamplerBehavior {
        magnify_filter: glium::uniforms::MagnifySamplerFilter::Nearest,
        .. Default::default()
    }));

Structs

EmptyUniforms
Object that can be used when you don't have any uniform.

Sampler
A sampler.

SamplerBehavior
Behavior of a sampler.

UniformsStorage
Stores uniforms.

Enums

MagnifySamplerFilter
The function that the GPU will use when loading the value of a texel.

MinifySamplerFilter
The function that the GPU will use when loading the value of a texel.

SamplerWrapFunction
Function to use for out-of-bounds samples.

UniformType
Type of a uniform in a program.

UniformValue
Represents a value to bind to a uniform.

Traits

IntoUniformValue
Represents a value that can be used as the value of a uniform.

Uniforms
Object that contains the values of all the uniforms to bind to a program.