Enum glium::uniforms::LayoutMismatchError

pub enum LayoutMismatchError {
    TypeMismatch {
        expected: UniformType,
        obtained: UniformType,
    },
    LayoutMismatch {
        expected: BlockLayout,
        obtained: BlockLayout,
    },
    OffsetMismatch {
        expected: usize,
        obtained: usize,
    },
    MemberMismatch {
        member: String,
        err: Box<LayoutMismatchError>,
    },
    MissingField {
        name: String,
    },
}

Error about a block layout mismatch.

Variants

TypeMismatch

There is a mismatch in the type of one element.

Fields

expected

Type expected by the shader.

obtained

Type that you gave.

LayoutMismatch

The expected layout is totally different from what we have.

Fields

expected

Layout expected by the shader.

obtained

Layout of the input.

OffsetMismatch

The type of data is good, but there is a misalignment.

Fields

expected

Expected offset of a member.

obtained

Offset of the same member in the input.

MemberMismatch

There is a mismatch in a submember of this layout.

This is kind of a hierarchy inside the LayoutMismatchErrors.

Fields

member

Name of the field.

err

The sub-error.

MissingField

A field is missing in either the expected of the input data layout.

Fields

name

Name of the field.