Struct glium::uniforms::SamplerBehaviorUnstable

pub struct SamplerBehavior {
    pub wrap_function: (SamplerWrapFunction, SamplerWrapFunction, SamplerWrapFunction),
    pub minify_filter: MinifySamplerFilter,
    pub magnify_filter: MagnifySamplerFilter,
    pub max_anisotropy: u16,
}

Behavior of a sampler.

Fields

wrap_function

Functions to use for the X, Y, and Z coordinates.

minify_filter

Filter to use when mignifying the texture.

magnify_filter

Filter to use when magnifying the texture.

max_anisotropy

1 means no anisotropic filtering, any value superior to 1 does.

Compatibility

This parameter is always available. However it is ignored on hardware that does not support anisotropic filtering.

If you set the value to a value higher than what the hardware supports, it will be clamped.