Struct glium::uniforms::UniformBuffer

pub struct UniformBuffer<T: ?Sized> where
    T: Content,  { /* fields omitted */ }

Buffer that contains a uniform block.

For example, to use a UniformBuffer<[u32; 8]>, you must declare it as

uniform MyBlock {
    uint array[8];
};

and pass it to uniform! like this:

uniform! {
    MyBlock: &buffer,
}

Methods

impl<T> UniformBuffer<T> where
T: Copy,

pub fn new<F: ?Sized>(
facade: &F,
data: T
) -> Result<UniformBuffer<T>, BufferCreationError> where
F: Facade,

Uploads data in the uniforms buffer.

pub fn dynamic<F: ?Sized>(
facade: &F,
data: T
) -> Result<UniformBuffer<T>, BufferCreationError> where
F: Facade,

Uploads data in the uniforms buffer.

pub fn persistent<F: ?Sized>(
facade: &F,
data: T
) -> Result<UniformBuffer<T>, BufferCreationError> where
F: Facade,

Uploads data in the uniforms buffer.

pub fn immutable<F: ?Sized>(
facade: &F,
data: T
) -> Result<UniformBuffer<T>, BufferCreationError> where
F: Facade,

Uploads data in the uniforms buffer.

pub fn empty<F: ?Sized>(
facade: &F
) -> Result<UniformBuffer<T>, BufferCreationError> where
F: Facade,

Creates an empty buffer.

pub fn empty_dynamic<F: ?Sized>(
facade: &F
) -> Result<UniformBuffer<T>, BufferCreationError> where
F: Facade,

Creates an empty buffer.

pub fn empty_persistent<F: ?Sized>(
facade: &F
) -> Result<UniformBuffer<T>, BufferCreationError> where
F: Facade,

Creates an empty buffer.

pub fn empty_immutable<F: ?Sized>(
facade: &F
) -> Result<UniformBuffer<T>, BufferCreationError> where
F: Facade,

Creates an empty buffer.

impl<T: ?Sized> UniformBuffer<T> where
T: Content,

pub fn empty_unsized<F: ?Sized>(
facade: &F,
size: usize
) -> Result<UniformBuffer<T>, BufferCreationError> where
F: Facade,

Creates an empty buffer.

Panic

Panics if the size passed as parameter is not suitable for the type of data.

pub fn empty_unsized_dynamic<F: ?Sized>(
facade: &F,
size: usize
) -> Result<UniformBuffer<T>, BufferCreationError> where
F: Facade,

Creates an empty buffer.

Panic

Panics if the size passed as parameter is not suitable for the type of data.

pub fn empty_unsized_persistent<F: ?Sized>(
facade: &F,
size: usize
) -> Result<UniformBuffer<T>, BufferCreationError> where
F: Facade,

Creates an empty buffer.

Panic

Panics if the size passed as parameter is not suitable for the type of data.

pub fn empty_unsized_immutable<F: ?Sized>(
facade: &F,
size: usize
) -> Result<UniformBuffer<T>, BufferCreationError> where
F: Facade,

Creates an empty buffer.

Panic

Panics if the size passed as parameter is not suitable for the type of data.