Trait glium::uniforms::UniformBlock

pub trait UniformBlock {
    fn matches(&BlockLayout, base_offset: usize) -> Result<(), LayoutMismatchError>;
    fn build_layout(base_offset: usize) -> BlockLayout;
}

Objects that are suitable for being inside a uniform block or a SSBO.

Required Methods

fn matches(&BlockLayout, base_offset: usize) -> Result<(), LayoutMismatchError>

Checks whether the uniforms' layout matches the given block if Self starts at the given offset.

fn build_layout(base_offset: usize) -> BlockLayout

Builds the BlockLayout corresponding to the current object.