Module glium::vertex

Contains everything related to vertex buffers.

The main struct is the VertexBuffer, which represents a buffer in the video memory, containing a list of vertices.

In order to create a vertex buffer, you must first create a struct that represents each vertex, and implement the glium::vertex::Vertex trait on it. The implement_vertex! macro helps you with that.

#[derive(Copy, Clone)]
struct Vertex {
    position: [f32; 3],
    texcoords: [f32; 2],
}

implement_vertex!(Vertex, position, texcoords);

Next, build a Vec of the vertices that you want to upload, and pass it to VertexBuffer::new.

let data = vec![
    Vertex {
        position: [0.0, 0.0, 0.4],
        texcoords: [0.0, 1.0]
    },
    Vertex {
        position: [12.0, 4.5, -1.8],
        texcoords: [1.0, 0.5]
    },
    Vertex {
        position: [-7.124, 0.1, 0.0],
        texcoords: [0.0, 0.4]
    },
];

let vertex_buffer = glium::vertex::VertexBuffer::new(&display, data);

Reexports

pub use self::buffer::{VertexBuffer};

pub use self::format::{VertexFormat};

Structs

Mapping
A mapping of a buffer.

PerInstance
Marker that instructs glium that the buffer is to be used per instance.

VertexBufferAny
A list of vertices loaded in the graphics card's memory.

VertexBufferAnySlice
Represents a slice of a VertexBufferAny.

VertexBufferSlice
Represents a slice of a VertexBuffer.

Enums

AttributeType

VerticesSource
Describes the source to use for the vertices when drawing.

Traits

Attribute
Trait for types that can be used as vertex attributes.

IntoVerticesSource
Objects that can be used as vertex sources.

MultiVerticesSource
Objects that describe multiple vertex sources.

Vertex
Trait for structures that represent a vertex.