Enum glium::vertex::AttributeType

pub enum AttributeType {
    I8,
    I8I8,
    I8I8I8,
    I8I8I8I8,
    U8,
    U8U8,
    U8U8U8,
    U8U8U8U8,
    I16,
    I16I16,
    I16I16I16,
    I16I16I16I16,
    U16,
    U16U16,
    U16U16U16,
    U16U16U16U16,
    I32,
    I32I32,
    I32I32I32,
    I32I32I32I32,
    U32,
    U32U32,
    U32U32U32,
    U32U32U32U32,
    I64,
    I64I64,
    I64I64I64,
    I64I64I64I64,
    U64,
    U64U64,
    U64U64U64,
    U64U64U64U64,
    F16,
    F16F16,
    F16F16F16,
    F16F16F16F16,
    F16x2x2,
    F16x2x3,
    F16x2x4,
    F16x3x2,
    F16x3x3,
    F16x3x4,
    F16x4x2,
    F16x4x3,
    F16x4x4,
    F32,
    F32F32,
    F32F32F32,
    F32F32F32F32,
    F32x2x2,
    F32x2x3,
    F32x2x4,
    F32x3x2,
    F32x3x3,
    F32x3x4,
    F32x4x2,
    F32x4x3,
    F32x4x4,
    F64,
    F64F64,
    F64F64F64,
    F64F64F64F64,
    F64x2x2,
    F64x2x3,
    F64x2x4,
    F64x3x2,
    F64x3x3,
    F64x3x4,
    F64x4x2,
    F64x4x3,
    F64x4x4,
    I2I10I10I10Reversed,
    U2U10U10U10Reversed,
    I10I10I10I2,
    U10U10U10U2,
    F10F11F11UnsignedIntReversed,
    FixedFloatI16U16,
}

Variants

I8

I8I8

I8I8I8

I8I8I8I8

U8

U8U8

U8U8U8

U8U8U8U8

I16

I16I16

I16I16I16

I16I16I16I16

U16

U16U16

U16U16U16

U16U16U16U16

I32

I32I32

I32I32I32

I32I32I32I32

U32

U32U32

U32U32U32

U32U32U32U32

I64

I64I64

I64I64I64

I64I64I64I64

U64

U64U64

U64U64U64

U64U64U64U64

F16

F16F16

F16F16F16

F16F16F16F16

F16x2x2

2x2 matrix of f16s

F16x2x3

2x3 matrix of f16s

F16x2x4

2x3 matrix of f16s

F16x3x2

3x2 matrix of f16s

F16x3x3

3x3 matrix of f16s

F16x3x4

3x4 matrix of f16s

F16x4x2

4x2 matrix of f16s

F16x4x3

4x3 matrix of f16s

F16x4x4

4x4 matrix of f16s

F32

F32F32

F32F32F32

F32F32F32F32

F32x2x2

2x2 matrix of f32s

F32x2x3

2x3 matrix of f32s

F32x2x4

2x3 matrix of f32s

F32x3x2

3x2 matrix of f32s

F32x3x3

3x3 matrix of f32s

F32x3x4

3x4 matrix of f32s

F32x4x2

4x2 matrix of f32s

F32x4x3

4x3 matrix of f32s

F32x4x4

4x4 matrix of f32s

F64

Warning: using f64s can be very slow.

F64F64

Warning: using f64s can be very slow.

F64F64F64

Warning: using f64s can be very slow.

F64F64F64F64

Warning: using f64s can be very slow.

F64x2x2

2x2 matrix of f64s Warning: using f64s can be very slow.

F64x2x3

2x3 matrix of f64s Warning: using f64s can be very slow.

F64x2x4

2x3 matrix of f64s Warning: using f64s can be very slow.

F64x3x2

3x2 matrix of f64s Warning: using f64s can be very slow.

F64x3x3

3x3 matrix of f64s Warning: using f64s can be very slow.

F64x3x4

3x4 matrix of f64s Warning: using f64s can be very slow.

F64x4x2

4x2 matrix of f64s Warning: using f64s can be very slow.

F64x4x3

4x3 matrix of f64s Warning: using f64s can be very slow.

F64x4x4

4x4 matrix of f64s Warning: using f64s can be very slow.

I2I10I10I10Reversed

From MSB to LSB: two bits for the alpha, ten bits for the blue, ten bits for the green, ten bits for the red.

Corresponds to GL_INT_2_10_10_10_REV.

U2U10U10U10Reversed

From MSB to LSB: two bits for the alpha, ten bits for the blue, ten bits for the green, ten bits for the red.

Corresponds to GL_UNSIGNED_INT_2_10_10_10_REV.

I10I10I10I2

Corresponds to GL_INT_10_10_10_2.

U10U10U10U2

Corresponds to GL_UNSIGNED_INT_10_10_10_2.

F10F11F11UnsignedIntReversed

Three floating points values turned into unsigned integers./

Corresponds to GL_UNSIGNED_INT_10F_11F_11F_REV.

FixedFloatI16U16

Fixed floating points. A 16bits signed value followed by the 16bits unsigned exponent.

Corresponds to GL_FIXED.

Implementations

impl AttributeType

pub fn is_supported<C: ?Sized>(&self, caps: &C) -> bool where
C: CapabilitiesSource,

Returns true if the backend supports this type of attribute.

pub fn get_size_bytes(&self) -> usize

Returns the size in bytes of a value of this type.

pub fn get_num_components(&self) -> usize

Returns the number of values for this type.