Enum glium::vertex::VerticesSource

pub enum VerticesSource<'a> {
    VertexBuffer(&'a VertexBufferAny, Option<Sender<LinearSyncFence>>, usize, usize),
    PerInstanceBuffer(&'a PerInstanceAttributesBufferAny, Option<Sender<LinearSyncFence>>),
}

Describes the source to use for the vertices when drawing.

Variants

VertexBuffer

A buffer uploaded in the video memory.

If the second parameter is Some, then a fence must be sent with this sender for when the buffer stops being used.

The third and fourth parameters are the offset and length of the buffer.

PerInstanceBuffer

A buffer uploaded in the video memory.

If the second parameter is Some, then a fence must be sent with this sender for when the buffer stops being used.