Enum glium::vertex::VerticesSource

pub enum VerticesSource<'a> {
    VertexBuffer(BufferAnySlice<'a>, &'a VertexFormat, bool),
    Marker {
        len: usize,
        per_instance: bool,
    },
}

Describes the source to use for the vertices when drawing.

Variants
VertexBuffer(BufferAnySlice<'a>, &'a VertexFormat, bool)

A buffer uploaded in the video memory.

The second parameter is the number of vertices in the buffer.

The third parameter tells whether or not this buffer is "per instance" (true) or "per vertex" (false).

Marker

A marker indicating a "phantom list of attributes".

Fields

len: usize

Number of attributes.

per_instance: bool

Whether or not this buffer is "per instance" (true) or "per vertex" (false).