Struct glium::vertex::VertexBufferAny

pub struct VertexBufferAny { /* fields omitted */ }

A list of vertices loaded in the graphics card's memory.

Contrary to VertexBuffer, this struct doesn't know about the type of data inside the buffer. Therefore you can't map or read it.

This struct is provided for convenience, so that you can have a Vec<VertexBufferAny>, or return a VertexBufferAny instead of a VertexBuffer<MyPrivateVertexType>.

Methods

impl VertexBufferAny

pub fn get_elements_size(&self) -> usize

Returns the number of bytes between two consecutive elements in the buffer.

pub fn len(&self) -> usize

Returns the number of elements in the buffer.

pub fn get_bindings(&self) -> &VertexFormat

Returns the associated VertexFormat.

pub unsafe fn into_vertex_buffer<T: Copy>(self) -> VertexBuffer<T>

Turns the vertex buffer into a VertexBuffer without checking the type.

pub fn per_instance(&self) -> Result<PerInstance, InstancingNotSupported>

Creates a marker that instructs glium to use multiple instances.

Instead of calling surface.draw(&vertex_buffer, ...) you can call surface.draw(vertex_buffer.per_instance(), ...). This will draw one instance of the geometry for each element in this buffer. The attributes are still passed to the vertex shader, but each entry is passed for each different instance.