Struct glium::vertex::VertexBufferSlice

pub struct VertexBufferSlice<'b, T: 'b> where
    T: Copy,  { /* fields omitted */ }

Represents a slice of a VertexBuffer.

Methods

impl<'b, T: 'b> VertexBufferSlice<'b, T> where
T: Copy + Content,

pub fn per_instance(&'b self) -> Result<PerInstance, InstancingNotSupported>

Creates a marker that instructs glium to use multiple instances.

Instead of calling surface.draw(&vertex_buffer.slice(...).unwrap(), ...) you can call surface.draw(vertex_buffer.slice(...).unwrap().per_instance(), ...). This will draw one instance of the geometry for each element in this buffer slice. The attributes are still passed to the vertex shader, but each entry is passed for each different instance.