Struct glium::vertex::VertexBufferSlice

pub struct VertexBufferSlice<'b, T: 'b> {
    // some fields omitted
}

Represents a slice of a VertexBuffer.

Methods

impl<'b, T> VertexBufferSlice<'b, T> where T: Send + Copy + 'static

fn read(&self) -> Vec<T>

Reads the content of the slice.

This function is usually better if are just doing one punctual read, while map is better if you want to have multiple small reads.

Features

Only available if the gl_read_buffer feature is enabled.

fn read_if_supported(&self) -> Option<Vec<T>>

Reads the content of the buffer.

This function is usually better if are just doing one punctual read, while map is better if you want to have multiple small reads.

fn write(&self, data: Vec<T>)

Writes some vertices to the buffer.

Panic

Panics if the length of data is different from the length of this slice.