Trait glium::vertex::Attribute

pub unsafe trait Attribute: Sized {
    fn get_type() -> AttributeType;

    fn is_supported<C>(caps: &C) -> bool where C: CapabilitiesSource { ... }
}

Trait for types that can be used as vertex attributes.

Required Methods

fn get_type() -> AttributeType

Get the type of data.

Provided Methods

fn is_supported<C>(caps: &C) -> bool where C: CapabilitiesSource

Returns true if the backend supports this type of attribute.